% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_service.R
\name{alexaforbusiness}
\alias{alexaforbusiness}
\title{Alexa For Business}
\usage{
alexaforbusiness(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\value{
A client for the service. You can call the service's operations using
syntax like \code{svc$operation(...)}, where \code{svc} is the name you've assigned
to the client. The available operations are listed in the
Operations section.
}
\description{
Alexa for Business helps you use Alexa in your organization. Alexa for
Business provides you with the tools to manage Alexa devices, enroll
your users, and assign skills, at scale. You can build your own
context-aware voice skills using the Alexa Skills Kit and the Alexa for
Business API operations. You can also make these available as private
skills for your organization. Alexa for Business makes it efficient to
voice-enable your products and services, thus providing context-aware
voice experiences for your customers. Device makers building with the
Alexa Voice Service (AVS) can create fully integrated solutions,
register their products with Alexa for Business, and manage them as
shared devices in their organization.
}
\section{Service syntax}{
\preformatted{svc <- alexaforbusiness(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=alexaforbusiness_approve_skill]{approve_skill} \tab Associates a skill with the organization under the customer's AWS account\cr
\link[=alexaforbusiness_associate_contact_with_address_book]{associate_contact_with_address_book} \tab Associates a contact with a given address book\cr
\link[=alexaforbusiness_associate_device_with_network_profile]{associate_device_with_network_profile} \tab Associates a device with the specified network profile\cr
\link[=alexaforbusiness_associate_device_with_room]{associate_device_with_room} \tab Associates a device with a given room\cr
\link[=alexaforbusiness_associate_skill_group_with_room]{associate_skill_group_with_room} \tab Associates a skill group with a given room\cr
\link[=alexaforbusiness_associate_skill_with_skill_group]{associate_skill_with_skill_group} \tab Associates a skill with a skill group\cr
\link[=alexaforbusiness_associate_skill_with_users]{associate_skill_with_users} \tab Makes a private skill available for enrolled users to enable on their devices\cr
\link[=alexaforbusiness_create_address_book]{create_address_book} \tab Creates an address book with the specified details\cr
\link[=alexaforbusiness_create_business_report_schedule]{create_business_report_schedule} \tab Creates a recurring schedule for usage reports to deliver to the specified S3 location with a specified daily or weekly interval\cr
\link[=alexaforbusiness_create_conference_provider]{create_conference_provider} \tab Adds a new conference provider under the user's AWS account\cr
\link[=alexaforbusiness_create_contact]{create_contact} \tab Creates a contact with the specified details\cr
\link[=alexaforbusiness_create_gateway_group]{create_gateway_group} \tab Creates a gateway group with the specified details\cr
\link[=alexaforbusiness_create_network_profile]{create_network_profile} \tab Creates a network profile with the specified details\cr
\link[=alexaforbusiness_create_profile]{create_profile} \tab Creates a new room profile with the specified details\cr
\link[=alexaforbusiness_create_room]{create_room} \tab Creates a room with the specified details\cr
\link[=alexaforbusiness_create_skill_group]{create_skill_group} \tab Creates a skill group with a specified name and description\cr
\link[=alexaforbusiness_create_user]{create_user} \tab Creates a user\cr
\link[=alexaforbusiness_delete_address_book]{delete_address_book} \tab Deletes an address book by the address book ARN\cr
\link[=alexaforbusiness_delete_business_report_schedule]{delete_business_report_schedule} \tab Deletes the recurring report delivery schedule with the specified schedule ARN\cr
\link[=alexaforbusiness_delete_conference_provider]{delete_conference_provider} \tab Deletes a conference provider\cr
\link[=alexaforbusiness_delete_contact]{delete_contact} \tab Deletes a contact by the contact ARN\cr
\link[=alexaforbusiness_delete_device]{delete_device} \tab Removes a device from Alexa For Business\cr
\link[=alexaforbusiness_delete_device_usage_data]{delete_device_usage_data} \tab When this action is called for a specified shared device, it allows authorized users to delete the device's entire previous history of voice input data and associated response data\cr
\link[=alexaforbusiness_delete_gateway_group]{delete_gateway_group} \tab Deletes a gateway group\cr
\link[=alexaforbusiness_delete_network_profile]{delete_network_profile} \tab Deletes a network profile by the network profile ARN\cr
\link[=alexaforbusiness_delete_profile]{delete_profile} \tab Deletes a room profile by the profile ARN\cr
\link[=alexaforbusiness_delete_room]{delete_room} \tab Deletes a room by the room ARN\cr
\link[=alexaforbusiness_delete_room_skill_parameter]{delete_room_skill_parameter} \tab Deletes room skill parameter details by room, skill, and parameter key ID\cr
\link[=alexaforbusiness_delete_skill_authorization]{delete_skill_authorization} \tab Unlinks a third-party account from a skill\cr
\link[=alexaforbusiness_delete_skill_group]{delete_skill_group} \tab Deletes a skill group by skill group ARN\cr
\link[=alexaforbusiness_delete_user]{delete_user} \tab Deletes a specified user by user ARN and enrollment ARN\cr
\link[=alexaforbusiness_disassociate_contact_from_address_book]{disassociate_contact_from_address_book} \tab Disassociates a contact from a given address book\cr
\link[=alexaforbusiness_disassociate_device_from_room]{disassociate_device_from_room} \tab Disassociates a device from its current room\cr
\link[=alexaforbusiness_disassociate_skill_from_skill_group]{disassociate_skill_from_skill_group} \tab Disassociates a skill from a skill group\cr
\link[=alexaforbusiness_disassociate_skill_from_users]{disassociate_skill_from_users} \tab Makes a private skill unavailable for enrolled users and prevents them from enabling it on their devices\cr
\link[=alexaforbusiness_disassociate_skill_group_from_room]{disassociate_skill_group_from_room} \tab Disassociates a skill group from a specified room\cr
\link[=alexaforbusiness_forget_smart_home_appliances]{forget_smart_home_appliances} \tab Forgets smart home appliances associated to a room\cr
\link[=alexaforbusiness_get_address_book]{get_address_book} \tab Gets address the book details by the address book ARN\cr
\link[=alexaforbusiness_get_conference_preference]{get_conference_preference} \tab Retrieves the existing conference preferences\cr
\link[=alexaforbusiness_get_conference_provider]{get_conference_provider} \tab Gets details about a specific conference provider\cr
\link[=alexaforbusiness_get_contact]{get_contact} \tab Gets the contact details by the contact ARN\cr
\link[=alexaforbusiness_get_device]{get_device} \tab Gets the details of a device by device ARN\cr
\link[=alexaforbusiness_get_gateway]{get_gateway} \tab Retrieves the details of a gateway\cr
\link[=alexaforbusiness_get_gateway_group]{get_gateway_group} \tab Retrieves the details of a gateway group\cr
\link[=alexaforbusiness_get_invitation_configuration]{get_invitation_configuration} \tab Retrieves the configured values for the user enrollment invitation email template\cr
\link[=alexaforbusiness_get_network_profile]{get_network_profile} \tab Gets the network profile details by the network profile ARN\cr
\link[=alexaforbusiness_get_profile]{get_profile} \tab Gets the details of a room profile by profile ARN\cr
\link[=alexaforbusiness_get_room]{get_room} \tab Gets room details by room ARN\cr
\link[=alexaforbusiness_get_room_skill_parameter]{get_room_skill_parameter} \tab Gets room skill parameter details by room, skill, and parameter key ARN\cr
\link[=alexaforbusiness_get_skill_group]{get_skill_group} \tab Gets skill group details by skill group ARN\cr
\link[=alexaforbusiness_list_business_report_schedules]{list_business_report_schedules} \tab Lists the details of the schedules that a user configured\cr
\link[=alexaforbusiness_list_conference_providers]{list_conference_providers} \tab Lists conference providers under a specific AWS account\cr
\link[=alexaforbusiness_list_device_events]{list_device_events} \tab Lists the device event history, including device connection status, for up to 30 days\cr
\link[=alexaforbusiness_list_gateway_groups]{list_gateway_groups} \tab Retrieves a list of gateway group summaries\cr
\link[=alexaforbusiness_list_gateways]{list_gateways} \tab Retrieves a list of gateway summaries\cr
\link[=alexaforbusiness_list_skills]{list_skills} \tab Lists all enabled skills in a specific skill group\cr
\link[=alexaforbusiness_list_skills_store_categories]{list_skills_store_categories} \tab Lists all categories in the Alexa skill store\cr
\link[=alexaforbusiness_list_skills_store_skills_by_category]{list_skills_store_skills_by_category} \tab Lists all skills in the Alexa skill store by category\cr
\link[=alexaforbusiness_list_smart_home_appliances]{list_smart_home_appliances} \tab Lists all of the smart home appliances associated with a room\cr
\link[=alexaforbusiness_list_tags]{list_tags} \tab Lists all tags for the specified resource\cr
\link[=alexaforbusiness_put_conference_preference]{put_conference_preference} \tab Sets the conference preferences on a specific conference provider at the account level\cr
\link[=alexaforbusiness_put_invitation_configuration]{put_invitation_configuration} \tab Configures the email template for the user enrollment invitation with the specified attributes\cr
\link[=alexaforbusiness_put_room_skill_parameter]{put_room_skill_parameter} \tab Updates room skill parameter details by room, skill, and parameter key ID\cr
\link[=alexaforbusiness_put_skill_authorization]{put_skill_authorization} \tab Links a user's account to a third-party skill provider\cr
\link[=alexaforbusiness_register_avs_device]{register_avs_device} \tab Registers an Alexa-enabled device built by an Original Equipment Manufacturer (OEM) using Alexa Voice Service (AVS)\cr
\link[=alexaforbusiness_reject_skill]{reject_skill} \tab Disassociates a skill from the organization under a user's AWS account\cr
\link[=alexaforbusiness_resolve_room]{resolve_room} \tab Determines the details for the room from which a skill request was invoked\cr
\link[=alexaforbusiness_revoke_invitation]{revoke_invitation} \tab Revokes an invitation and invalidates the enrollment URL\cr
\link[=alexaforbusiness_search_address_books]{search_address_books} \tab Searches address books and lists the ones that meet a set of filter and sort criteria\cr
\link[=alexaforbusiness_search_contacts]{search_contacts} \tab Searches contacts and lists the ones that meet a set of filter and sort criteria\cr
\link[=alexaforbusiness_search_devices]{search_devices} \tab Searches devices and lists the ones that meet a set of filter criteria\cr
\link[=alexaforbusiness_search_network_profiles]{search_network_profiles} \tab Searches network profiles and lists the ones that meet a set of filter and sort criteria\cr
\link[=alexaforbusiness_search_profiles]{search_profiles} \tab Searches room profiles and lists the ones that meet a set of filter criteria\cr
\link[=alexaforbusiness_search_rooms]{search_rooms} \tab Searches rooms and lists the ones that meet a set of filter and sort criteria\cr
\link[=alexaforbusiness_search_skill_groups]{search_skill_groups} \tab Searches skill groups and lists the ones that meet a set of filter and sort criteria\cr
\link[=alexaforbusiness_search_users]{search_users} \tab Searches users and lists the ones that meet a set of filter and sort criteria\cr
\link[=alexaforbusiness_send_announcement]{send_announcement} \tab Triggers an asynchronous flow to send text, SSML, or audio announcements to rooms that are identified by a search or filter\cr
\link[=alexaforbusiness_send_invitation]{send_invitation} \tab Sends an enrollment invitation email with a URL to a user\cr
\link[=alexaforbusiness_start_device_sync]{start_device_sync} \tab Resets a device and its account to the known default settings\cr
\link[=alexaforbusiness_start_smart_home_appliance_discovery]{start_smart_home_appliance_discovery} \tab Initiates the discovery of any smart home appliances associated with the room\cr
\link[=alexaforbusiness_tag_resource]{tag_resource} \tab Adds metadata tags to a specified resource\cr
\link[=alexaforbusiness_untag_resource]{untag_resource} \tab Removes metadata tags from a specified resource\cr
\link[=alexaforbusiness_update_address_book]{update_address_book} \tab Updates address book details by the address book ARN\cr
\link[=alexaforbusiness_update_business_report_schedule]{update_business_report_schedule} \tab Updates the configuration of the report delivery schedule with the specified schedule ARN\cr
\link[=alexaforbusiness_update_conference_provider]{update_conference_provider} \tab Updates an existing conference provider's settings\cr
\link[=alexaforbusiness_update_contact]{update_contact} \tab Updates the contact details by the contact ARN\cr
\link[=alexaforbusiness_update_device]{update_device} \tab Updates the device name by device ARN\cr
\link[=alexaforbusiness_update_gateway]{update_gateway} \tab Updates the details of a gateway\cr
\link[=alexaforbusiness_update_gateway_group]{update_gateway_group} \tab Updates the details of a gateway group\cr
\link[=alexaforbusiness_update_network_profile]{update_network_profile} \tab Updates a network profile by the network profile ARN\cr
\link[=alexaforbusiness_update_profile]{update_profile} \tab Updates an existing room profile by room profile ARN\cr
\link[=alexaforbusiness_update_room]{update_room} \tab Updates room details by room ARN\cr
\link[=alexaforbusiness_update_skill_group]{update_skill_group} \tab Updates skill group details by skill group ARN
}
}

\examples{
\dontrun{
svc <- alexaforbusiness()
svc$approve_skill(
  Foo = 123
)
}

}
