% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_approve_skill}
\alias{alexaforbusiness_approve_skill}
\title{Associates a skill with the organization under the customer's AWS
account}
\usage{
alexaforbusiness_approve_skill(SkillId)
}
\arguments{
\item{SkillId}{[required] The unique identifier of the skill.}
}
\value{
An empty list.
}
\description{
Associates a skill with the organization under the customer's AWS
account. If a skill is private, the user implicitly accepts access to
this skill during enablement.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$approve_skill(
  SkillId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
