% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_associate_contact_with_address_book}
\alias{alexaforbusiness_associate_contact_with_address_book}
\title{Associates a contact with a given address book}
\usage{
alexaforbusiness_associate_contact_with_address_book(ContactArn,
  AddressBookArn)
}
\arguments{
\item{ContactArn}{[required] The ARN of the contact to associate with an address book.}

\item{AddressBookArn}{[required] The ARN of the address book with which to associate the contact.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Associates a contact with a given address book.
}
\section{Request syntax}{
\preformatted{svc$associate_contact_with_address_book(
  ContactArn = "string",
  AddressBookArn = "string"
)
}
}

\keyword{internal}
