% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_associate_device_with_network_profile}
\alias{alexaforbusiness_associate_device_with_network_profile}
\title{Associates a device with the specified network profile}
\usage{
alexaforbusiness_associate_device_with_network_profile(DeviceArn,
  NetworkProfileArn)
}
\arguments{
\item{DeviceArn}{[required] The device ARN.}

\item{NetworkProfileArn}{[required] The ARN of the network profile to associate with a device.}
}
\value{
An empty list.
}
\description{
Associates a device with the specified network profile.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$associate_device_with_network_profile(
  DeviceArn = "string",
  NetworkProfileArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
