% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_associate_skill_group_with_room}
\alias{alexaforbusiness_associate_skill_group_with_room}
\title{Associates a skill group with a given room}
\usage{
alexaforbusiness_associate_skill_group_with_room(SkillGroupArn, RoomArn)
}
\arguments{
\item{SkillGroupArn}{The ARN of the skill group to associate with a room. Required.}

\item{RoomArn}{The ARN of the room with which to associate the skill group. Required.}
}
\value{
An empty list.
}
\description{
Associates a skill group with a given room. This enables all skills in
the associated skill group on all devices in the room.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$associate_skill_group_with_room(
  SkillGroupArn = "string",
  RoomArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
