% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_create_address_book}
\alias{alexaforbusiness_create_address_book}
\title{Creates an address book with the specified details}
\usage{
alexaforbusiness_create_address_book(Name, Description,
  ClientRequestToken, Tags)
}
\arguments{
\item{Name}{[required] The name of the address book.}

\item{Description}{The description of the address book.}

\item{ClientRequestToken}{A unique, user-specified identifier for the request that ensures
idempotency.}

\item{Tags}{The tags to be added to the specified resource. Do not provide system
tags.}
}
\value{
A list with the following syntax:\preformatted{list(
  AddressBookArn = "string"
)
}
}
\description{
Creates an address book with the specified details.
}
\section{Request syntax}{
\preformatted{svc$create_address_book(
  Name = "string",
  Description = "string",
  ClientRequestToken = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
