% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_create_conference_provider}
\alias{alexaforbusiness_create_conference_provider}
\title{Adds a new conference provider under the user's AWS account}
\usage{
alexaforbusiness_create_conference_provider(ConferenceProviderName,
  ConferenceProviderType, IPDialIn, PSTNDialIn, MeetingSetting,
  ClientRequestToken)
}
\arguments{
\item{ConferenceProviderName}{[required] The name of the conference provider.}

\item{ConferenceProviderType}{[required] Represents a type within a list of predefined types.}

\item{IPDialIn}{The IP endpoint and protocol for calling.}

\item{PSTNDialIn}{The information for PSTN conferencing.}

\item{MeetingSetting}{[required] The meeting settings for the conference provider.}

\item{ClientRequestToken}{The request token of the client.}
}
\description{
Adds a new conference provider under the user\'s AWS account.
}
\section{Request syntax}{
\preformatted{svc$create_conference_provider(
  ConferenceProviderName = "string",
  ConferenceProviderType = "CHIME"|"BLUEJEANS"|"FUZE"|"GOOGLE_HANGOUTS"|"POLYCOM"|"RINGCENTRAL"|"SKYPE_FOR_BUSINESS"|"WEBEX"|"ZOOM"|"CUSTOM",
  IPDialIn = list(
    Endpoint = "string",
    CommsProtocol = "SIP"|"SIPS"|"H323"
  ),
  PSTNDialIn = list(
    CountryCode = "string",
    PhoneNumber = "string",
    OneClickIdDelay = "string",
    OneClickPinDelay = "string"
  ),
  MeetingSetting = list(
    RequirePin = "YES"|"NO"|"OPTIONAL"
  ),
  ClientRequestToken = "string"
)
}
}

\keyword{internal}
