% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_create_contact}
\alias{alexaforbusiness_create_contact}
\title{Creates a contact with the specified details}
\usage{
alexaforbusiness_create_contact(DisplayName, FirstName, LastName,
  PhoneNumber, PhoneNumbers, SipAddresses, ClientRequestToken)
}
\arguments{
\item{DisplayName}{The name of the contact to display on the console.}

\item{FirstName}{[required] The first name of the contact that is used to call the contact on the
device.}

\item{LastName}{The last name of the contact that is used to call the contact on the
device.}

\item{PhoneNumber}{The phone number of the contact in E.164 format. The phone number type
defaults to WORK. You can specify PhoneNumber or PhoneNumbers. We
recommend that you use PhoneNumbers, which lets you specify the phone
number type and multiple numbers.}

\item{PhoneNumbers}{The list of phone numbers for the contact.}

\item{SipAddresses}{The list of SIP addresses for the contact.}

\item{ClientRequestToken}{A unique, user-specified identifier for this request that ensures
idempotency.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ContactArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Creates a contact with the specified details.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_contact(
  DisplayName = "string",
  FirstName = "string",
  LastName = "string",
  PhoneNumber = "string",
  PhoneNumbers = list(
    list(
      Number = "string",
      Type = "MOBILE"|"WORK"|"HOME"
    )
  ),
  SipAddresses = list(
    list(
      Uri = "string",
      Type = "WORK"
    )
  ),
  ClientRequestToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
