% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_create_gateway_group}
\alias{alexaforbusiness_create_gateway_group}
\title{Creates a gateway group with the specified details}
\usage{
alexaforbusiness_create_gateway_group(Name, Description,
  ClientRequestToken)
}
\arguments{
\item{Name}{[required] The name of the gateway group.}

\item{Description}{The description of the gateway group.}

\item{ClientRequestToken}{[required] A unique, user-specified identifier for the request that ensures
idempotency.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  GatewayGroupArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Creates a gateway group with the specified details.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_gateway_group(
  Name = "string",
  Description = "string",
  ClientRequestToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
