% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_create_network_profile}
\alias{alexaforbusiness_create_network_profile}
\title{Creates a network profile with the specified details}
\usage{
alexaforbusiness_create_network_profile(NetworkProfileName, Description,
  Ssid, SecurityType, EapMethod, CurrentPassword, NextPassword,
  CertificateAuthorityArn, TrustAnchors, ClientRequestToken, Tags)
}
\arguments{
\item{NetworkProfileName}{[required] The name of the network profile associated with a device.}

\item{Description}{Detailed information about a device's network profile.}

\item{Ssid}{[required] The SSID of the Wi-Fi network.}

\item{SecurityType}{[required] The security type of the Wi-Fi network. This can be WPA2_ENTERPRISE,
WPA2_PSK, WPA_PSK, WEP, or OPEN.}

\item{EapMethod}{The authentication standard that is used in the EAP framework.
Currently, EAP_TLS is supported.}

\item{CurrentPassword}{The current password of the Wi-Fi network.}

\item{NextPassword}{The next, or subsequent, password of the Wi-Fi network. This password is
asynchronously transmitted to the device and is used when the password
of the network changes to NextPassword.}

\item{CertificateAuthorityArn}{The ARN of the Private Certificate Authority (PCA) created in AWS
Certificate Manager (ACM). This is used to issue certificates to the
devices.}

\item{TrustAnchors}{The root certificates of your authentication server that is installed on
your devices and used to trust your authentication server during EAP
negotiation.}

\item{ClientRequestToken}{[required]}

\item{Tags}{The tags to be added to the specified resource. Do not provide system
tags.}
}
\value{
A list with the following syntax:\preformatted{list(
  NetworkProfileArn = "string"
)
}
}
\description{
Creates a network profile with the specified details.
}
\section{Request syntax}{
\preformatted{svc$create_network_profile(
  NetworkProfileName = "string",
  Description = "string",
  Ssid = "string",
  SecurityType = "OPEN"|"WEP"|"WPA_PSK"|"WPA2_PSK"|"WPA2_ENTERPRISE",
  EapMethod = "EAP_TLS",
  CurrentPassword = "string",
  NextPassword = "string",
  CertificateAuthorityArn = "string",
  TrustAnchors = list(
    "string"
  ),
  ClientRequestToken = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
