% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_create_skill_group}
\alias{alexaforbusiness_create_skill_group}
\title{Creates a skill group with a specified name and description}
\usage{
alexaforbusiness_create_skill_group(SkillGroupName, Description,
  ClientRequestToken, Tags)
}
\arguments{
\item{SkillGroupName}{[required] The name for the skill group.}

\item{Description}{The description for the skill group.}

\item{ClientRequestToken}{A unique, user-specified identifier for this request that ensures
idempotency.}

\item{Tags}{The tags for the skill group.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  SkillGroupArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Creates a skill group with a specified name and description.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_skill_group(
  SkillGroupName = "string",
  Description = "string",
  ClientRequestToken = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
