% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_create_user}
\alias{alexaforbusiness_create_user}
\title{Creates a user}
\usage{
alexaforbusiness_create_user(UserId, FirstName, LastName, Email,
  ClientRequestToken, Tags)
}
\arguments{
\item{UserId}{[required] The ARN for the user.}

\item{FirstName}{The first name for the user.}

\item{LastName}{The last name for the user.}

\item{Email}{The email address for the user.}

\item{ClientRequestToken}{A unique, user-specified identifier for this request that ensures
idempotency.}

\item{Tags}{The tags for the user.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  UserArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Creates a user.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_user(
  UserId = "string",
  FirstName = "string",
  LastName = "string",
  Email = "string",
  ClientRequestToken = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
