% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_delete_address_book}
\alias{alexaforbusiness_delete_address_book}
\title{Deletes an address book by the address book ARN}
\usage{
alexaforbusiness_delete_address_book(AddressBookArn)
}
\arguments{
\item{AddressBookArn}{[required] The ARN of the address book to delete.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes an address book by the address book ARN.
}
\section{Request syntax}{
\preformatted{svc$delete_address_book(
  AddressBookArn = "string"
)
}
}

\keyword{internal}
