% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_delete_device_usage_data}
\alias{alexaforbusiness_delete_device_usage_data}
\title{When this action is called for a specified shared device, it allows
authorized users to delete the device's entire previous history of voice
input data and associated response data}
\usage{
alexaforbusiness_delete_device_usage_data(DeviceArn, DeviceUsageType)
}
\arguments{
\item{DeviceArn}{[required] The ARN of the device.}

\item{DeviceUsageType}{[required] The type of usage data to delete.}
}
\value{
An empty list.
}
\description{
When this action is called for a specified shared device, it allows
authorized users to delete the device's entire previous history of voice
input data and associated response data. This action can be called once
every 24 hours for a specific shared device.
}
\section{Request syntax}{
\preformatted{svc$delete_device_usage_data(
  DeviceArn = "string",
  DeviceUsageType = "VOICE"
)
}
}

\keyword{internal}
