% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_delete_skill_group}
\alias{alexaforbusiness_delete_skill_group}
\title{Deletes a skill group by skill group ARN}
\usage{
alexaforbusiness_delete_skill_group(SkillGroupArn)
}
\arguments{
\item{SkillGroupArn}{The ARN of the skill group to delete. Required.}
}
\value{
An empty list.
}
\description{
Deletes a skill group by skill group ARN.
}
\section{Request syntax}{
\preformatted{svc$delete_skill_group(
  SkillGroupArn = "string"
)
}
}

\keyword{internal}
