% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_disassociate_contact_from_address_book}
\alias{alexaforbusiness_disassociate_contact_from_address_book}
\title{Disassociates a contact from a given address book}
\usage{
alexaforbusiness_disassociate_contact_from_address_book(ContactArn,
  AddressBookArn)
}
\arguments{
\item{ContactArn}{[required] The ARN of the contact to disassociate from an address book.}

\item{AddressBookArn}{[required] The ARN of the address from which to disassociate the contact.}
}
\value{
An empty list.
}
\description{
Disassociates a contact from a given address book.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$disassociate_contact_from_address_book(
  ContactArn = "string",
  AddressBookArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
