% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_disassociate_device_from_room}
\alias{alexaforbusiness_disassociate_device_from_room}
\title{Disassociates a device from its current room}
\usage{
alexaforbusiness_disassociate_device_from_room(DeviceArn)
}
\arguments{
\item{DeviceArn}{The ARN of the device to disassociate from a room. Required.}
}
\value{
An empty list.
}
\description{
Disassociates a device from its current room. The device continues to be
connected to the Wi-Fi network and is still registered to the account.
The device settings and skills are removed from the room.
}
\section{Request syntax}{
\preformatted{svc$disassociate_device_from_room(
  DeviceArn = "string"
)
}
}

\keyword{internal}
