% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_disassociate_skill_from_skill_group}
\alias{alexaforbusiness_disassociate_skill_from_skill_group}
\title{Disassociates a skill from a skill group}
\usage{
alexaforbusiness_disassociate_skill_from_skill_group(SkillGroupArn,
  SkillId)
}
\arguments{
\item{SkillGroupArn}{The unique identifier of a skill. Required.}

\item{SkillId}{[required] The ARN of a skill group to associate to a skill.}
}
\value{
An empty list.
}
\description{
Disassociates a skill from a skill group.
}
\section{Request syntax}{
\preformatted{svc$disassociate_skill_from_skill_group(
  SkillGroupArn = "string",
  SkillId = "string"
)
}
}

\keyword{internal}
