% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_disassociate_skill_from_users}
\alias{alexaforbusiness_disassociate_skill_from_users}
\title{Makes a private skill unavailable for enrolled users and prevents them
from enabling it on their devices}
\usage{
alexaforbusiness_disassociate_skill_from_users(SkillId)
}
\arguments{
\item{SkillId}{[required] The private skill ID you want to make unavailable for enrolled users.}
}
\value{
An empty list.
}
\description{
Makes a private skill unavailable for enrolled users and prevents them
from enabling it on their devices.
}
\section{Request syntax}{
\preformatted{svc$disassociate_skill_from_users(
  SkillId = "string"
)
}
}

\keyword{internal}
