% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_disassociate_skill_group_from_room}
\alias{alexaforbusiness_disassociate_skill_group_from_room}
\title{Disassociates a skill group from a specified room}
\usage{
alexaforbusiness_disassociate_skill_group_from_room(SkillGroupArn,
  RoomArn)
}
\arguments{
\item{SkillGroupArn}{The ARN of the skill group to disassociate from a room. Required.}

\item{RoomArn}{The ARN of the room from which the skill group is to be disassociated.
Required.}
}
\value{
An empty list.
}
\description{
Disassociates a skill group from a specified room. This disables all
skills in the skill group on all devices in the room.
}
\section{Request syntax}{
\preformatted{svc$disassociate_skill_group_from_room(
  SkillGroupArn = "string",
  RoomArn = "string"
)
}
}

\keyword{internal}
