% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_get_contact}
\alias{alexaforbusiness_get_contact}
\title{Gets the contact details by the contact ARN}
\usage{
alexaforbusiness_get_contact(ContactArn)
}
\arguments{
\item{ContactArn}{[required] The ARN of the contact for which to request details.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Contact = list(
    ContactArn = "string",
    DisplayName = "string",
    FirstName = "string",
    LastName = "string",
    PhoneNumber = "string",
    PhoneNumbers = list(
      list(
        Number = "string",
        Type = "MOBILE"|"WORK"|"HOME"
      )
    ),
    SipAddresses = list(
      list(
        Uri = "string",
        Type = "WORK"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets the contact details by the contact ARN.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_contact(
  ContactArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
