% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_get_gateway}
\alias{alexaforbusiness_get_gateway}
\title{Retrieves the details of a gateway}
\usage{
alexaforbusiness_get_gateway(GatewayArn)
}
\arguments{
\item{GatewayArn}{[required] The ARN of the gateway to get.}
}
\value{
A list with the following syntax:\preformatted{list(
  Gateway = list(
    Arn = "string",
    Name = "string",
    Description = "string",
    GatewayGroupArn = "string",
    SoftwareVersion = "string"
  )
)
}
}
\description{
Retrieves the details of a gateway.
}
\section{Request syntax}{
\preformatted{svc$get_gateway(
  GatewayArn = "string"
)
}
}

\keyword{internal}
