% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_get_gateway_group}
\alias{alexaforbusiness_get_gateway_group}
\title{Retrieves the details of a gateway group}
\usage{
alexaforbusiness_get_gateway_group(GatewayGroupArn)
}
\arguments{
\item{GatewayGroupArn}{[required] The ARN of the gateway group to get.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  GatewayGroup = list(
    Arn = "string",
    Name = "string",
    Description = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the details of a gateway group.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_gateway_group(
  GatewayGroupArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
