% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_get_invitation_configuration}
\alias{alexaforbusiness_get_invitation_configuration}
\title{Retrieves the configured values for the user enrollment invitation email
template}
\usage{
alexaforbusiness_get_invitation_configuration()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  OrganizationName = "string",
  ContactEmail = "string",
  PrivateSkillIds = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the configured values for the user enrollment invitation email
template.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_invitation_configuration()
}\if{html}{\out{</div>}}
}

\keyword{internal}
