% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_get_network_profile}
\alias{alexaforbusiness_get_network_profile}
\title{Gets the network profile details by the network profile ARN}
\usage{
alexaforbusiness_get_network_profile(NetworkProfileArn)
}
\arguments{
\item{NetworkProfileArn}{[required] The ARN of the network profile associated with a device.}
}
\value{
A list with the following syntax:\preformatted{list(
  NetworkProfile = list(
    NetworkProfileArn = "string",
    NetworkProfileName = "string",
    Description = "string",
    Ssid = "string",
    SecurityType = "OPEN"|"WEP"|"WPA_PSK"|"WPA2_PSK"|"WPA2_ENTERPRISE",
    EapMethod = "EAP_TLS",
    CurrentPassword = "string",
    NextPassword = "string",
    CertificateAuthorityArn = "string",
    TrustAnchors = list(
      "string"
    )
  )
)
}
}
\description{
Gets the network profile details by the network profile ARN.
}
\section{Request syntax}{
\preformatted{svc$get_network_profile(
  NetworkProfileArn = "string"
)
}
}

\keyword{internal}
