% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_get_room}
\alias{alexaforbusiness_get_room}
\title{Gets room details by room ARN}
\usage{
alexaforbusiness_get_room(RoomArn)
}
\arguments{
\item{RoomArn}{The ARN of the room for which to request details. Required.}
}
\description{
Gets room details by room ARN.
}
\section{Request syntax}{
\preformatted{svc$get_room(
  RoomArn = "string"
)
}
}

\keyword{internal}
