% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_list_conference_providers}
\alias{alexaforbusiness_list_conference_providers}
\title{Lists conference providers under a specific AWS account}
\usage{
alexaforbusiness_list_conference_providers(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{The tokens used for pagination.}

\item{MaxResults}{The maximum number of conference providers to be returned, per paginated
calls.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ConferenceProviders = list(
    list(
      Arn = "string",
      Name = "string",
      Type = "CHIME"|"BLUEJEANS"|"FUZE"|"GOOGLE_HANGOUTS"|"POLYCOM"|"RINGCENTRAL"|"SKYPE_FOR_BUSINESS"|"WEBEX"|"ZOOM"|"CUSTOM",
      IPDialIn = list(
        Endpoint = "string",
        CommsProtocol = "SIP"|"SIPS"|"H323"
      ),
      PSTNDialIn = list(
        CountryCode = "string",
        PhoneNumber = "string",
        OneClickIdDelay = "string",
        OneClickPinDelay = "string"
      ),
      MeetingSetting = list(
        RequirePin = "YES"|"NO"|"OPTIONAL"
      )
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists conference providers under a specific AWS account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_conference_providers(
  NextToken = "string",
  MaxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
