% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_list_smart_home_appliances}
\alias{alexaforbusiness_list_smart_home_appliances}
\title{Lists all of the smart home appliances associated with a room}
\usage{
alexaforbusiness_list_smart_home_appliances(RoomArn, MaxResults,
  NextToken)
}
\arguments{
\item{RoomArn}{[required] The room that the appliances are associated with.}

\item{MaxResults}{The maximum number of appliances to be returned, per paginated calls.}

\item{NextToken}{The tokens used for pagination.}
}
\value{
A list with the following syntax:\preformatted{list(
  SmartHomeAppliances = list(
    list(
      FriendlyName = "string",
      Description = "string",
      ManufacturerName = "string"
    )
  ),
  NextToken = "string"
)
}
}
\description{
Lists all of the smart home appliances associated with a room.
}
\section{Request syntax}{
\preformatted{svc$list_smart_home_appliances(
  RoomArn = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
