% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_put_invitation_configuration}
\alias{alexaforbusiness_put_invitation_configuration}
\title{Configures the email template for the user enrollment invitation with
the specified attributes}
\usage{
alexaforbusiness_put_invitation_configuration(OrganizationName,
  ContactEmail, PrivateSkillIds)
}
\arguments{
\item{OrganizationName}{[required] The name of the organization sending the enrollment invite to a user.}

\item{ContactEmail}{The email ID of the organization or individual contact that the enrolled
user can use.}

\item{PrivateSkillIds}{The list of private skill IDs that you want to recommend to the user to
enable in the invitation.}
}
\description{
Configures the email template for the user enrollment invitation with
the specified attributes.
}
\section{Request syntax}{
\preformatted{svc$put_invitation_configuration(
  OrganizationName = "string",
  ContactEmail = "string",
  PrivateSkillIds = list(
    "string"
  )
)
}
}

\keyword{internal}
