% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_reject_skill}
\alias{alexaforbusiness_reject_skill}
\title{Disassociates a skill from the organization under a user's AWS account}
\usage{
alexaforbusiness_reject_skill(SkillId)
}
\arguments{
\item{SkillId}{[required] The unique identifier of the skill.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Disassociates a skill from the organization under a user's AWS account.
If the skill is a private skill, it moves to an AcceptStatus of PENDING.
Any private or public skill that is rejected can be added later by
calling the ApproveSkill API.
}
\section{Request syntax}{
\preformatted{svc$reject_skill(
  SkillId = "string"
)
}
}

\keyword{internal}
