% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_resolve_room}
\alias{alexaforbusiness_resolve_room}
\title{Determines the details for the room from which a skill request was
invoked}
\usage{
alexaforbusiness_resolve_room(UserId, SkillId)
}
\arguments{
\item{UserId}{[required] The ARN of the user. Required.}

\item{SkillId}{[required] The ARN of the skill that was requested. Required.}
}
\value{
A list with the following syntax:\preformatted{list(
  RoomArn = "string",
  RoomName = "string",
  RoomSkillParameters = list(
    list(
      ParameterKey = "string",
      ParameterValue = "string"
    )
  )
)
}
}
\description{
Determines the details for the room from which a skill request was
invoked. This operation is used by skill developers.

To query ResolveRoom from an Alexa skill, the skill ID needs to be
authorized. When the skill is using an AWS Lambda function, the skill is
automatically authorized when you publish your skill as a private skill
to your AWS account. Skills that are hosted using a custom web service
must be manually authorized. To get your skill authorized, contact AWS
Support with your AWS account ID that queries the ResolveRoom API and
skill ID.
}
\section{Request syntax}{
\preformatted{svc$resolve_room(
  UserId = "string",
  SkillId = "string"
)
}
}

\keyword{internal}
