% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_revoke_invitation}
\alias{alexaforbusiness_revoke_invitation}
\title{Revokes an invitation and invalidates the enrollment URL}
\usage{
alexaforbusiness_revoke_invitation(UserArn, EnrollmentId)
}
\arguments{
\item{UserArn}{The ARN of the user for whom to revoke an enrollment invitation.
Required.}

\item{EnrollmentId}{The ARN of the enrollment invitation to revoke. Required.}
}
\value{
An empty list.
}
\description{
Revokes an invitation and invalidates the enrollment URL.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$revoke_invitation(
  UserArn = "string",
  EnrollmentId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
