% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_search_address_books}
\alias{alexaforbusiness_search_address_books}
\title{Searches address books and lists the ones that meet a set of filter and
sort criteria}
\usage{
alexaforbusiness_search_address_books(Filters, SortCriteria, NextToken,
  MaxResults)
}
\arguments{
\item{Filters}{The filters to use to list a specified set of address books. The
supported filter key is AddressBookName.}

\item{SortCriteria}{The sort order to use in listing the specified set of address books. The
supported sort key is AddressBookName.}

\item{NextToken}{An optional token returned from a prior request. Use this token for
pagination of results from this action. If this parameter is specified,
the response only includes results beyond the token, up to the value
specified by MaxResults.}

\item{MaxResults}{The maximum number of results to include in the response. If more
results exist than the specified MaxResults value, a token is included
in the response so that the remaining results can be retrieved.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AddressBooks = list(
    list(
      AddressBookArn = "string",
      Name = "string",
      Description = "string"
    )
  ),
  NextToken = "string",
  TotalCount = 123
)
}\if{html}{\out{</div>}}
}
\description{
Searches address books and lists the ones that meet a set of filter and
sort criteria.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$search_address_books(
  Filters = list(
    list(
      Key = "string",
      Values = list(
        "string"
      )
    )
  ),
  SortCriteria = list(
    list(
      Key = "string",
      Value = "ASC"|"DESC"
    )
  ),
  NextToken = "string",
  MaxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
