% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_search_devices}
\alias{alexaforbusiness_search_devices}
\title{Searches devices and lists the ones that meet a set of filter criteria}
\usage{
alexaforbusiness_search_devices(NextToken, MaxResults, Filters,
  SortCriteria)
}
\arguments{
\item{NextToken}{An optional token returned from a prior request. Use this token for
pagination of results from this action. If this parameter is specified,
the response includes only results beyond the token, up to the value
specified by \code{MaxResults}.}

\item{MaxResults}{The maximum number of results to include in the response. If more
results exist than the specified \code{MaxResults} value, a token is included
in the response so that the remaining results can be retrieved.}

\item{Filters}{The filters to use to list a specified set of devices. Supported filter
keys are DeviceName, DeviceStatus, DeviceStatusDetailCode, RoomName,
DeviceType, DeviceSerialNumber, UnassociatedOnly, ConnectionStatus
(ONLINE and OFFLINE), NetworkProfileName, NetworkProfileArn, Feature,
and FailureCode.}

\item{SortCriteria}{The sort order to use in listing the specified set of devices. Supported
sort keys are DeviceName, DeviceStatus, RoomName, DeviceType,
DeviceSerialNumber, ConnectionStatus, NetworkProfileName,
NetworkProfileArn, Feature, and FailureCode.}
}
\value{
A list with the following syntax:\preformatted{list(
  Devices = list(
    list(
      DeviceArn = "string",
      DeviceSerialNumber = "string",
      DeviceType = "string",
      DeviceName = "string",
      SoftwareVersion = "string",
      MacAddress = "string",
      DeviceStatus = "READY"|"PENDING"|"WAS_OFFLINE"|"DEREGISTERED"|"FAILED",
      NetworkProfileArn = "string",
      NetworkProfileName = "string",
      RoomArn = "string",
      RoomName = "string",
      DeviceStatusInfo = list(
        DeviceStatusDetails = list(
          list(
            Feature = "BLUETOOTH"|"VOLUME"|"NOTIFICATIONS"|"LISTS"|"SKILLS"|"NETWORK_PROFILE"|"SETTINGS"|"ALL",
            Code = "DEVICE_SOFTWARE_UPDATE_NEEDED"|"DEVICE_WAS_OFFLINE"|"CREDENTIALS_ACCESS_FAILURE"|"TLS_VERSION_MISMATCH"|"ASSOCIATION_REJECTION"|"AUTHENTICATION_FAILURE"|"DHCP_FAILURE"|"INTERNET_UNAVAILABLE"|"DNS_FAILURE"|"UNKNOWN_FAILURE"|"CERTIFICATE_ISSUING_LIMIT_EXCEEDED"|"INVALID_CERTIFICATE_AUTHORITY"|"NETWORK_PROFILE_NOT_FOUND"|"INVALID_PASSWORD_STATE"|"PASSWORD_NOT_FOUND"|"PASSWORD_MANAGER_ACCESS_DENIED"|"CERTIFICATE_AUTHORITY_ACCESS_DENIED"
          )
        ),
        ConnectionStatus = "ONLINE"|"OFFLINE",
        ConnectionStatusUpdatedTime = as.POSIXct(
          "2015-01-01"
        )
      ),
      CreatedTime = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  NextToken = "string",
  TotalCount = 123
)
}
}
\description{
Searches devices and lists the ones that meet a set of filter criteria.
}
\section{Request syntax}{
\preformatted{svc$search_devices(
  NextToken = "string",
  MaxResults = 123,
  Filters = list(
    list(
      Key = "string",
      Values = list(
        "string"
      )
    )
  ),
  SortCriteria = list(
    list(
      Key = "string",
      Value = "ASC"|"DESC"
    )
  )
)
}
}

\keyword{internal}
