% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_search_network_profiles}
\alias{alexaforbusiness_search_network_profiles}
\title{Searches network profiles and lists the ones that meet a set of filter
and sort criteria}
\usage{
alexaforbusiness_search_network_profiles(NextToken, MaxResults, Filters,
  SortCriteria)
}
\arguments{
\item{NextToken}{An optional token returned from a prior request. Use this token for
pagination of results from this action. If this parameter is specified,
the response includes only results beyond the token, up to the value
specified by MaxResults.}

\item{MaxResults}{The maximum number of results to include in the response. If more
results exist than the specified MaxResults value, a token is included
in the response so that the remaining results can be retrieved.}

\item{Filters}{The filters to use to list a specified set of network profiles. Valid
filters are NetworkProfileName, Ssid, and SecurityType.}

\item{SortCriteria}{The sort order to use to list the specified set of network profiles.
Valid sort criteria includes NetworkProfileName, Ssid, and SecurityType.}
}
\value{
A list with the following syntax:\preformatted{list(
  NetworkProfiles = list(
    list(
      NetworkProfileArn = "string",
      NetworkProfileName = "string",
      Description = "string",
      Ssid = "string",
      SecurityType = "OPEN"|"WEP"|"WPA_PSK"|"WPA2_PSK"|"WPA2_ENTERPRISE",
      EapMethod = "EAP_TLS",
      CertificateAuthorityArn = "string"
    )
  ),
  NextToken = "string",
  TotalCount = 123
)
}
}
\description{
Searches network profiles and lists the ones that meet a set of filter
and sort criteria.
}
\section{Request syntax}{
\preformatted{svc$search_network_profiles(
  NextToken = "string",
  MaxResults = 123,
  Filters = list(
    list(
      Key = "string",
      Values = list(
        "string"
      )
    )
  ),
  SortCriteria = list(
    list(
      Key = "string",
      Value = "ASC"|"DESC"
    )
  )
)
}
}

\keyword{internal}
