% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_search_rooms}
\alias{alexaforbusiness_search_rooms}
\title{Searches rooms and lists the ones that meet a set of filter and sort
criteria}
\usage{
alexaforbusiness_search_rooms(NextToken, MaxResults, Filters,
  SortCriteria)
}
\arguments{
\item{NextToken}{An optional token returned from a prior request. Use this token for
pagination of results from this action. If this parameter is specified,
the response includes only results beyond the token, up to the value
specified by \code{MaxResults}.}

\item{MaxResults}{The maximum number of results to include in the response. If more
results exist than the specified \code{MaxResults} value, a token is included
in the response so that the remaining results can be retrieved.}

\item{Filters}{The filters to use to list a specified set of rooms. The supported
filter keys are RoomName and ProfileName.}

\item{SortCriteria}{The sort order to use in listing the specified set of rooms. The
supported sort keys are RoomName and ProfileName.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Rooms = list(
    list(
      RoomArn = "string",
      RoomName = "string",
      Description = "string",
      ProviderCalendarId = "string",
      ProfileArn = "string",
      ProfileName = "string"
    )
  ),
  NextToken = "string",
  TotalCount = 123
)
}\if{html}{\out{</div>}}
}
\description{
Searches rooms and lists the ones that meet a set of filter and sort
criteria.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$search_rooms(
  NextToken = "string",
  MaxResults = 123,
  Filters = list(
    list(
      Key = "string",
      Values = list(
        "string"
      )
    )
  ),
  SortCriteria = list(
    list(
      Key = "string",
      Value = "ASC"|"DESC"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
