% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_search_users}
\alias{alexaforbusiness_search_users}
\title{Searches users and lists the ones that meet a set of filter and sort
criteria}
\usage{
alexaforbusiness_search_users(NextToken, MaxResults, Filters,
  SortCriteria)
}
\arguments{
\item{NextToken}{An optional token returned from a prior request. Use this token for
pagination of results from this action. If this parameter is specified,
the response includes only results beyond the token, up to the value
specified by \code{MaxResults}. Required.}

\item{MaxResults}{The maximum number of results to include in the response. If more
results exist than the specified \code{MaxResults} value, a token is included
in the response so that the remaining results can be retrieved.
Required.}

\item{Filters}{The filters to use for listing a specific set of users. Required.
Supported filter keys are UserId, FirstName, LastName, Email, and
EnrollmentStatus.}

\item{SortCriteria}{The sort order to use in listing the filtered set of users. Required.
Supported sort keys are UserId, FirstName, LastName, Email, and
EnrollmentStatus.}
}
\description{
Searches users and lists the ones that meet a set of filter and sort
criteria.
}
\section{Request syntax}{
\preformatted{svc$search_users(
  NextToken = "string",
  MaxResults = 123,
  Filters = list(
    list(
      Key = "string",
      Values = list(
        "string"
      )
    )
  ),
  SortCriteria = list(
    list(
      Key = "string",
      Value = "ASC"|"DESC"
    )
  )
)
}
}

\keyword{internal}
