% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_send_invitation}
\alias{alexaforbusiness_send_invitation}
\title{Sends an enrollment invitation email with a URL to a user}
\usage{
alexaforbusiness_send_invitation(UserArn)
}
\arguments{
\item{UserArn}{The ARN of the user to whom to send an invitation. Required.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Sends an enrollment invitation email with a URL to a user. The URL is
valid for 30 days or until you call this operation again, whichever
comes first.
}
\section{Request syntax}{
\preformatted{svc$send_invitation(
  UserArn = "string"
)
}
}

\keyword{internal}
