% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_start_device_sync}
\alias{alexaforbusiness_start_device_sync}
\title{Resets a device and its account to the known default settings}
\usage{
alexaforbusiness_start_device_sync(RoomArn, DeviceArn, Features)
}
\arguments{
\item{RoomArn}{The ARN of the room with which the device to sync is associated.
Required.}

\item{DeviceArn}{The ARN of the device to sync. Required.}

\item{Features}{[required] Request structure to start the device sync. Required.}
}
\value{
An empty list.
}
\description{
Resets a device and its account to the known default settings. This
clears all information and settings set by previous users in the
following ways:
\itemize{
\item Bluetooth - This unpairs all bluetooth devices paired with your echo
device.
\item Volume - This resets the echo device's volume to the default value.
\item Notifications - This clears all notifications from your echo device.
\item Lists - This clears all to-do items from your echo device.
\item Settings - This internally syncs the room's profile (if the device
is assigned to a room), contacts, address books, delegation access
for account linking, and communications (if enabled on the room
profile).
}
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$start_device_sync(
  RoomArn = "string",
  DeviceArn = "string",
  Features = list(
    "BLUETOOTH"|"VOLUME"|"NOTIFICATIONS"|"LISTS"|"SKILLS"|"NETWORK_PROFILE"|"SETTINGS"|"ALL"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
