% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_start_smart_home_appliance_discovery}
\alias{alexaforbusiness_start_smart_home_appliance_discovery}
\title{Initiates the discovery of any smart home appliances associated with the
room}
\usage{
alexaforbusiness_start_smart_home_appliance_discovery(RoomArn)
}
\arguments{
\item{RoomArn}{[required] The room where smart home appliance discovery was initiated.}
}
\value{
An empty list.
}
\description{
Initiates the discovery of any smart home appliances associated with the
room.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$start_smart_home_appliance_discovery(
  RoomArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
