% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_untag_resource}
\alias{alexaforbusiness_untag_resource}
\title{Removes metadata tags from a specified resource}
\usage{
alexaforbusiness_untag_resource(Arn, TagKeys)
}
\arguments{
\item{Arn}{[required] The ARN of the resource from which to remove metadata tags. Required.}

\item{TagKeys}{[required] The tags to be removed from the specified resource. Do not provide
system tags. Required.}
}
\value{
An empty list.
}
\description{
Removes metadata tags from a specified resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  Arn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
