% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_update_contact}
\alias{alexaforbusiness_update_contact}
\title{Updates the contact details by the contact ARN}
\usage{
alexaforbusiness_update_contact(ContactArn, DisplayName, FirstName,
  LastName, PhoneNumber, PhoneNumbers, SipAddresses)
}
\arguments{
\item{ContactArn}{[required] The ARN of the contact to update.}

\item{DisplayName}{The updated display name of the contact.}

\item{FirstName}{The updated first name of the contact.}

\item{LastName}{The updated last name of the contact.}

\item{PhoneNumber}{The updated phone number of the contact. The phone number type defaults
to WORK. You can either specify PhoneNumber or PhoneNumbers. We
recommend that you use PhoneNumbers, which lets you specify the phone
number type and multiple numbers.}

\item{PhoneNumbers}{The list of phone numbers for the contact.}

\item{SipAddresses}{The list of SIP addresses for the contact.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Updates the contact details by the contact ARN.
}
\section{Request syntax}{
\preformatted{svc$update_contact(
  ContactArn = "string",
  DisplayName = "string",
  FirstName = "string",
  LastName = "string",
  PhoneNumber = "string",
  PhoneNumbers = list(
    list(
      Number = "string",
      Type = "MOBILE"|"WORK"|"HOME"
    )
  ),
  SipAddresses = list(
    list(
      Uri = "string",
      Type = "WORK"
    )
  )
)
}
}

\keyword{internal}
