% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_update_device}
\alias{alexaforbusiness_update_device}
\title{Updates the device name by device ARN}
\usage{
alexaforbusiness_update_device(DeviceArn, DeviceName)
}
\arguments{
\item{DeviceArn}{The ARN of the device to update. Required.}

\item{DeviceName}{The updated device name. Required.}
}
\value{
An empty list.
}
\description{
Updates the device name by device ARN.
}
\section{Request syntax}{
\preformatted{svc$update_device(
  DeviceArn = "string",
  DeviceName = "string"
)
}
}

\keyword{internal}
