% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_update_gateway}
\alias{alexaforbusiness_update_gateway}
\title{Updates the details of a gateway}
\usage{
alexaforbusiness_update_gateway(GatewayArn, Name, Description,
  SoftwareVersion)
}
\arguments{
\item{GatewayArn}{[required] The ARN of the gateway to update.}

\item{Name}{The updated name of the gateway.}

\item{Description}{The updated description of the gateway.}

\item{SoftwareVersion}{The updated software version of the gateway. The gateway automatically
updates its software version during normal operation.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Updates the details of a gateway. If any optional field is not provided,
the existing corresponding value is left unmodified.
}
\section{Request syntax}{
\preformatted{svc$update_gateway(
  GatewayArn = "string",
  Name = "string",
  Description = "string",
  SoftwareVersion = "string"
)
}
}

\keyword{internal}
