% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_update_gateway_group}
\alias{alexaforbusiness_update_gateway_group}
\title{Updates the details of a gateway group}
\usage{
alexaforbusiness_update_gateway_group(GatewayGroupArn, Name,
  Description)
}
\arguments{
\item{GatewayGroupArn}{[required] The ARN of the gateway group to update.}

\item{Name}{The updated name of the gateway group.}

\item{Description}{The updated description of the gateway group.}
}
\value{
An empty list.
}
\description{
Updates the details of a gateway group. If any optional field is not
provided, the existing corresponding value is left unmodified.
}
\section{Request syntax}{
\preformatted{svc$update_gateway_group(
  GatewayGroupArn = "string",
  Name = "string",
  Description = "string"
)
}
}

\keyword{internal}
