% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_update_network_profile}
\alias{alexaforbusiness_update_network_profile}
\title{Updates a network profile by the network profile ARN}
\usage{
alexaforbusiness_update_network_profile(NetworkProfileArn,
  NetworkProfileName, Description, CurrentPassword, NextPassword,
  CertificateAuthorityArn, TrustAnchors)
}
\arguments{
\item{NetworkProfileArn}{[required] The ARN of the network profile associated with a device.}

\item{NetworkProfileName}{The name of the network profile associated with a device.}

\item{Description}{Detailed information about a device's network profile.}

\item{CurrentPassword}{The current password of the Wi-Fi network.}

\item{NextPassword}{The next, or subsequent, password of the Wi-Fi network. This password is
asynchronously transmitted to the device and is used when the password
of the network changes to NextPassword.}

\item{CertificateAuthorityArn}{The ARN of the Private Certificate Authority (PCA) created in AWS
Certificate Manager (ACM). This is used to issue certificates to the
devices.}

\item{TrustAnchors}{The root certificate(s) of your authentication server that will be
installed on your devices and used to trust your authentication server
during EAP negotiation.}
}
\value{
An empty list.
}
\description{
Updates a network profile by the network profile ARN.
}
\section{Request syntax}{
\preformatted{svc$update_network_profile(
  NetworkProfileArn = "string",
  NetworkProfileName = "string",
  Description = "string",
  CurrentPassword = "string",
  NextPassword = "string",
  CertificateAuthorityArn = "string",
  TrustAnchors = list(
    "string"
  )
)
}
}

\keyword{internal}
