% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_update_room}
\alias{alexaforbusiness_update_room}
\title{Updates room details by room ARN}
\usage{
alexaforbusiness_update_room(RoomArn, RoomName, Description,
  ProviderCalendarId, ProfileArn)
}
\arguments{
\item{RoomArn}{The ARN of the room to update.}

\item{RoomName}{The updated name for the room.}

\item{Description}{The updated description for the room.}

\item{ProviderCalendarId}{The updated provider calendar ARN for the room.}

\item{ProfileArn}{The updated profile ARN for the room.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Updates room details by room ARN.
}
\section{Request syntax}{
\preformatted{svc$update_room(
  RoomArn = "string",
  RoomName = "string",
  Description = "string",
  ProviderCalendarId = "string",
  ProfileArn = "string"
)
}
}

\keyword{internal}
