% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_associate_phone_number_with_user}
\alias{chime_associate_phone_number_with_user}
\title{Associates a phone number with the specified Amazon Chime user}
\usage{
chime_associate_phone_number_with_user(AccountId, UserId,
  E164PhoneNumber)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{UserId}{[required] The user ID.}

\item{E164PhoneNumber}{[required] The phone number, in E.164 format.}
}
\value{
An empty list.
}
\description{
Associates a phone number with the specified Amazon Chime user.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$associate_phone_number_with_user(
  AccountId = "string",
  UserId = "string",
  E164PhoneNumber = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
