% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_associate_phone_numbers_with_voice_connector}
\alias{chime_associate_phone_numbers_with_voice_connector}
\title{Associates phone numbers with the specified Amazon Chime Voice Connector}
\usage{
chime_associate_phone_numbers_with_voice_connector(VoiceConnectorId,
  E164PhoneNumbers, ForceAssociate)
}
\arguments{
\item{VoiceConnectorId}{[required] The Amazon Chime Voice Connector ID.}

\item{E164PhoneNumbers}{[required] List of phone numbers, in E.164 format.}

\item{ForceAssociate}{If true, associates the provided phone numbers with the provided Amazon
Chime Voice Connector and removes any previously existing associations.
If false, does not associate any phone numbers that have previously
existing associations.}
}
\value{
A list with the following syntax:\preformatted{list(
  PhoneNumberErrors = list(
    list(
      PhoneNumberId = "string",
      ErrorCode = "BadRequest"|"Conflict"|"Forbidden"|"NotFound"|"PreconditionFailed"|"ResourceLimitExceeded"|"ServiceFailure"|"AccessDenied"|"ServiceUnavailable"|"Throttled"|"Throttling"|"Unauthorized"|"Unprocessable"|"VoiceConnectorGroupAssociationsExist"|"PhoneNumberAssociationsExist",
      ErrorMessage = "string"
    )
  )
)
}
}
\description{
Associates phone numbers with the specified Amazon Chime Voice
Connector.
}
\section{Request syntax}{
\preformatted{svc$associate_phone_numbers_with_voice_connector(
  VoiceConnectorId = "string",
  E164PhoneNumbers = list(
    "string"
  ),
  ForceAssociate = TRUE|FALSE
)
}
}

\keyword{internal}
