% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_associate_signin_delegate_groups_with_account}
\alias{chime_associate_signin_delegate_groups_with_account}
\title{Associates the specified sign-in delegate groups with the specified
Amazon Chime account}
\usage{
chime_associate_signin_delegate_groups_with_account(AccountId,
  SigninDelegateGroups)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{SigninDelegateGroups}{[required] The sign-in delegate groups.}
}
\value{
An empty list.
}
\description{
Associates the specified sign-in delegate groups with the specified
Amazon Chime account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$associate_signin_delegate_groups_with_account(
  AccountId = "string",
  SigninDelegateGroups = list(
    list(
      GroupName = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
