% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_batch_create_channel_membership}
\alias{chime_batch_create_channel_membership}
\title{Adds a specified number of users to a channel}
\usage{
chime_batch_create_channel_membership(ChannelArn, Type, MemberArns,
  ChimeBearer)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel to which you're adding users.}

\item{Type}{The membership type of a user, \code{DEFAULT} or \code{HIDDEN}. Default members
are always returned as part of
\code{\link[=chime_list_channel_memberships]{list_channel_memberships}}. Hidden
members are only returned if the type filter in
\code{\link[=chime_list_channel_memberships]{list_channel_memberships}} equals
\code{HIDDEN}. Otherwise hidden members are not returned. This is only
supported by moderators.}

\item{MemberArns}{[required] The ARNs of the members you want to add to the channel.}

\item{ChimeBearer}{The \code{AppInstanceUserArn} of the user that makes the API call.}
}
\value{
A list with the following syntax:\preformatted{list(
  BatchChannelMemberships = list(
    InvitedBy = list(
      Arn = "string",
      Name = "string"
    ),
    Type = "DEFAULT"|"HIDDEN",
    Members = list(
      list(
        Arn = "string",
        Name = "string"
      )
    ),
    ChannelArn = "string"
  ),
  Errors = list(
    list(
      MemberArn = "string",
      ErrorCode = "BadRequest"|"Conflict"|"Forbidden"|"NotFound"|"PreconditionFailed"|"ResourceLimitExceeded"|"ServiceFailure"|"AccessDenied"|"ServiceUnavailable"|"Throttled"|"Throttling"|"Unauthorized"|"Unprocessable"|"VoiceConnectorGroupAssociationsExist"|"PhoneNumberAssociationsExist",
      ErrorMessage = "string"
    )
  )
)
}
}
\description{
Adds a specified number of users to a channel.
}
\section{Request syntax}{
\preformatted{svc$batch_create_channel_membership(
  ChannelArn = "string",
  Type = "DEFAULT"|"HIDDEN",
  MemberArns = list(
    "string"
  ),
  ChimeBearer = "string"
)
}
}

\keyword{internal}
