% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_batch_suspend_user}
\alias{chime_batch_suspend_user}
\title{Suspends up to 50 users from a Team or EnterpriseLWA Amazon Chime
account}
\usage{
chime_batch_suspend_user(AccountId, UserIdList)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{UserIdList}{[required] The request containing the user IDs to suspend.}
}
\value{
A list with the following syntax:\preformatted{list(
  UserErrors = list(
    list(
      UserId = "string",
      ErrorCode = "BadRequest"|"Conflict"|"Forbidden"|"NotFound"|"PreconditionFailed"|"ResourceLimitExceeded"|"ServiceFailure"|"AccessDenied"|"ServiceUnavailable"|"Throttled"|"Throttling"|"Unauthorized"|"Unprocessable"|"VoiceConnectorGroupAssociationsExist"|"PhoneNumberAssociationsExist",
      ErrorMessage = "string"
    )
  )
)
}
}
\description{
Suspends up to 50 users from a \code{Team} or \code{EnterpriseLWA} Amazon Chime
account. For more information about different account types, see
\href{https://docs.aws.amazon.com/chime/latest/ag/manage-chime-account.html}{Managing Your Amazon Chime Accounts}
in the \emph{Amazon Chime Administration Guide}.

Users suspended from a \code{Team} account are disassociated from the
account,but they can continue to use Amazon Chime as free users. To
remove the suspension from suspended \code{Team} account users, invite them
to the \code{Team} account again. You can use the
\code{\link[=chime_invite_users]{invite_users}} action to do so.

Users suspended from an \code{EnterpriseLWA} account are immediately signed
out of Amazon Chime and can no longer sign in. To remove the suspension
from suspended \code{EnterpriseLWA} account users, use the
\code{\link[=chime_batch_unsuspend_user]{batch_unsuspend_user}} action.

To sign out users without suspending them, use the
\code{\link[=chime_logout_user]{logout_user}} action.
}
\section{Request syntax}{
\preformatted{svc$batch_suspend_user(
  AccountId = "string",
  UserIdList = list(
    "string"
  )
)
}
}

\keyword{internal}
