% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_batch_unsuspend_user}
\alias{chime_batch_unsuspend_user}
\title{Removes the suspension from up to 50 previously suspended users for the
specified Amazon Chime EnterpriseLWA account}
\usage{
chime_batch_unsuspend_user(AccountId, UserIdList)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{UserIdList}{[required] The request containing the user IDs to unsuspend.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  UserErrors = list(
    list(
      UserId = "string",
      ErrorCode = "BadRequest"|"Conflict"|"Forbidden"|"NotFound"|"PreconditionFailed"|"ResourceLimitExceeded"|"ServiceFailure"|"AccessDenied"|"ServiceUnavailable"|"Throttled"|"Throttling"|"Unauthorized"|"Unprocessable"|"VoiceConnectorGroupAssociationsExist"|"PhoneNumberAssociationsExist",
      ErrorMessage = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Removes the suspension from up to 50 previously suspended users for the
specified Amazon Chime \code{EnterpriseLWA} account. Only users on
\code{EnterpriseLWA} accounts can be unsuspended using this action. For more
information about different account types, see \href{https://docs.aws.amazon.com/chime/latest/ag/manage-chime-account.html}{Managing Your Amazon Chime Accounts}
in the \emph{Amazon Chime Administration Guide}.

Previously suspended users who are unsuspended using this action are
returned to \code{Registered} status. Users who are not previously suspended
are ignored.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$batch_unsuspend_user(
  AccountId = "string",
  UserIdList = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
