% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_batch_update_phone_number}
\alias{chime_batch_update_phone_number}
\title{Updates phone number product types or calling names}
\usage{
chime_batch_update_phone_number(UpdatePhoneNumberRequestItems)
}
\arguments{
\item{UpdatePhoneNumberRequestItems}{[required] The request containing the phone number IDs and product types or calling
names to update.}
}
\value{
A list with the following syntax:\preformatted{list(
  PhoneNumberErrors = list(
    list(
      PhoneNumberId = "string",
      ErrorCode = "BadRequest"|"Conflict"|"Forbidden"|"NotFound"|"PreconditionFailed"|"ResourceLimitExceeded"|"ServiceFailure"|"AccessDenied"|"ServiceUnavailable"|"Throttled"|"Throttling"|"Unauthorized"|"Unprocessable"|"VoiceConnectorGroupAssociationsExist"|"PhoneNumberAssociationsExist",
      ErrorMessage = "string"
    )
  )
)
}
}
\description{
Updates phone number product types or calling names. You can update one
attribute at a time for each \code{UpdatePhoneNumberRequestItem}. For
example, you can update the product type or the calling name.

For toll-free numbers, you cannot use the Amazon Chime Business Calling
product type. For numbers outside the US, you must use the Amazon Chime
SIP Media Application Dial-In product type.

Updates to outbound calling names can take 72 hours to complete. Pending
updates to outbound calling names must be complete before you can
request another update.
}
\section{Request syntax}{
\preformatted{svc$batch_update_phone_number(
  UpdatePhoneNumberRequestItems = list(
    list(
      PhoneNumberId = "string",
      ProductType = "BusinessCalling"|"VoiceConnector"|"SipMediaApplicationDialIn",
      CallingName = "string"
    )
  )
)
}
}

\keyword{internal}
