% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_create_account}
\alias{chime_create_account}
\title{Creates an Amazon Chime account under the administrator's AWS account}
\usage{
chime_create_account(Name)
}
\arguments{
\item{Name}{[required] The name of the Amazon Chime account.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Account = list(
    AwsAccountId = "string",
    AccountId = "string",
    Name = "string",
    AccountType = "Team"|"EnterpriseDirectory"|"EnterpriseLWA"|"EnterpriseOIDC",
    CreatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    DefaultLicense = "Basic"|"Plus"|"Pro"|"ProTrial",
    SupportedLicenses = list(
      "Basic"|"Plus"|"Pro"|"ProTrial"
    ),
    SigninDelegateGroups = list(
      list(
        GroupName = "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Creates an Amazon Chime account under the administrator's AWS account.
Only \code{Team} account types are currently supported for this action. For
more information about different account types, see \href{https://docs.aws.amazon.com/chime/latest/ag/manage-chime-account.html}{Managing Your Amazon Chime Accounts}
in the \emph{Amazon Chime Administration Guide}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_account(
  Name = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
