% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_create_app_instance_admin}
\alias{chime_create_app_instance_admin}
\title{Promotes an AppInstanceUser to an AppInstanceAdmin}
\usage{
chime_create_app_instance_admin(AppInstanceAdminArn, AppInstanceArn)
}
\arguments{
\item{AppInstanceAdminArn}{[required] The ARN of the administrator of the current \code{AppInstance}.}

\item{AppInstanceArn}{[required] The ARN of the \code{AppInstance}.}
}
\value{
A list with the following syntax:\preformatted{list(
  AppInstanceAdmin = list(
    Arn = "string",
    Name = "string"
  ),
  AppInstanceArn = "string"
)
}
}
\description{
Promotes an \code{AppInstanceUser} to an \code{AppInstanceAdmin}. The promoted
user can perform the following actions.
\itemize{
\item \code{ChannelModerator} actions across all channels in the \code{AppInstance}.
\item \code{\link[=chime_delete_channel_message]{delete_channel_message}} actions.
}

Only an \code{AppInstanceUser} can be promoted to an \code{AppInstanceAdmin} role.
}
\section{Request syntax}{
\preformatted{svc$create_app_instance_admin(
  AppInstanceAdminArn = "string",
  AppInstanceArn = "string"
)
}
}

\keyword{internal}
