% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_create_channel_ban}
\alias{chime_create_channel_ban}
\title{Permanently bans a member from a channel}
\usage{
chime_create_channel_ban(ChannelArn, MemberArn)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the ban request.}

\item{MemberArn}{[required] The ARN of the member being banned.}
}
\description{
Permanently bans a member from a channel. Moderators can't add banned
members to a channel. To undo a ban, you first have to
\code{DeleteChannelBan}, and then \code{CreateChannelMembership}. Bans are cleaned
up when you delete users or channels.

If you ban a user who is already part of a channel, that user is
automatically kicked from the channel.
}
\section{Request syntax}{
\preformatted{svc$create_channel_ban(
  ChannelArn = "string",
  MemberArn = "string"
)
}
}

\keyword{internal}
