% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_create_channel_membership}
\alias{chime_create_channel_membership}
\title{Adds a user to a channel}
\usage{
chime_create_channel_membership(ChannelArn, MemberArn, Type)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel to which you're adding users.}

\item{MemberArn}{[required] The ARN of the member you want to add to the channel.}

\item{Type}{[required] The membership type of a user, \code{DEFAULT} or \code{HIDDEN}. Default members
are always returned as part of
\code{\link[=chime_list_channel_memberships]{list_channel_memberships}}. Hidden
members are only returned if the type filter in
\code{\link[=chime_list_channel_memberships]{list_channel_memberships}} equals
\code{HIDDEN}. Otherwise hidden members are not returned. This is only
supported by moderators.}
}
\value{
A list with the following syntax:\preformatted{list(
  ChannelArn = "string",
  Member = list(
    Arn = "string",
    Name = "string"
  )
)
}
}
\description{
Adds a user to a channel. The \code{InvitedBy} response field is derived from
the request header. A channel member can:
\itemize{
\item List messages
\item Send messages
\item Receive messages
\item Edit their own messages
\item Leave the channel
}

Privacy settings impact this action as follows:
\itemize{
\item Public Channels: You do not need to be a member to list messages,
but you must be a member to send messages.
\item Private Channels: You must be a member to list or send messages.
}
}
\section{Request syntax}{
\preformatted{svc$create_channel_membership(
  ChannelArn = "string",
  MemberArn = "string",
  Type = "DEFAULT"|"HIDDEN"
)
}
}

\keyword{internal}
