% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_create_channel_moderator}
\alias{chime_create_channel_moderator}
\title{Creates a new ChannelModerator}
\usage{
chime_create_channel_moderator(ChannelArn, ChannelModeratorArn,
  ChimeBearer)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel.}

\item{ChannelModeratorArn}{[required] The ARN of the moderator.}

\item{ChimeBearer}{The \code{AppInstanceUserArn} of the user that makes the API call.}
}
\value{
A list with the following syntax:\preformatted{list(
  ChannelArn = "string",
  ChannelModerator = list(
    Arn = "string",
    Name = "string"
  )
)
}
}
\description{
Creates a new \code{ChannelModerator}. A channel moderator can:
\itemize{
\item Add and remove other members of the channel.
\item Add and remove other moderators of the channel.
\item Add and remove user bans for the channel.
\item Redact messages in the channel.
\item List messages in the channel.
}

The \code{x-amz-chime-bearer} request header is mandatory. Use the
\code{AppInstanceUserArn} of the user that makes the API call as the value in
the header.
}
\section{Request syntax}{
\preformatted{svc$create_channel_moderator(
  ChannelArn = "string",
  ChannelModeratorArn = "string",
  ChimeBearer = "string"
)
}
}

\keyword{internal}
