% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_create_meeting_with_attendees}
\alias{chime_create_meeting_with_attendees}
\title{Creates a new Amazon Chime SDK meeting in the specified media Region,
with attendees}
\usage{
chime_create_meeting_with_attendees(ClientRequestToken,
  ExternalMeetingId, MeetingHostId, MediaRegion, Tags,
  NotificationsConfiguration, Attendees)
}
\arguments{
\item{ClientRequestToken}{[required] The unique identifier for the client request. Use a different token for
different meetings.}

\item{ExternalMeetingId}{The external meeting ID.}

\item{MeetingHostId}{Reserved.}

\item{MediaRegion}{The Region in which to create the meeting. Default: \code{us-east-1}.

Available values: \code{af-south-1}, \code{ap-northeast-1}, \code{ap-northeast-2},
\code{ap-south-1}, \code{ap-southeast-1}, \code{ap-southeast-2}, \code{ca-central-1},
\code{eu-central-1}, \code{eu-north-1}, \code{eu-south-1}, \code{eu-west-1}, \code{eu-west-2},
\code{eu-west-3}, \code{sa-east-1}, \code{us-east-1}, \code{us-east-2}, \code{us-west-1},
\code{us-west-2}.}

\item{Tags}{The tag key-value pairs.}

\item{NotificationsConfiguration}{}

\item{Attendees}{The request containing the attendees to create.}
}
\value{
A list with the following syntax:\preformatted{list(
  Meeting = list(
    MeetingId = "string",
    ExternalMeetingId = "string",
    MediaPlacement = list(
      AudioHostUrl = "string",
      AudioFallbackUrl = "string",
      ScreenDataUrl = "string",
      ScreenSharingUrl = "string",
      ScreenViewingUrl = "string",
      SignalingUrl = "string",
      TurnControlUrl = "string"
    ),
    MediaRegion = "string"
  ),
  Attendees = list(
    list(
      ExternalUserId = "string",
      AttendeeId = "string",
      JoinToken = "string"
    )
  ),
  Errors = list(
    list(
      ExternalUserId = "string",
      ErrorCode = "string",
      ErrorMessage = "string"
    )
  )
)
}
}
\description{
Creates a new Amazon Chime SDK meeting in the specified media Region,
with attendees. For more information about specifying media Regions, see
\href{https://docs.aws.amazon.com/chime/latest/dg/chime-sdk-meetings-regions.html}{Amazon Chime SDK Media Regions}
in the \emph{Amazon Chime Developer Guide}. For more information about the
Amazon Chime SDK, see \href{https://docs.aws.amazon.com/chime/latest/dg/meetings-sdk.html}{Using the Amazon Chime SDK} in
the \emph{Amazon Chime Developer Guide}.
}
\section{Request syntax}{
\preformatted{svc$create_meeting_with_attendees(
  ClientRequestToken = "string",
  ExternalMeetingId = "string",
  MeetingHostId = "string",
  MediaRegion = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  NotificationsConfiguration = list(
    SnsTopicArn = "string",
    SqsQueueArn = "string"
  ),
  Attendees = list(
    list(
      ExternalUserId = "string",
      Tags = list(
        list(
          Key = "string",
          Value = "string"
        )
      )
    )
  )
)
}
}

\keyword{internal}
