% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_create_phone_number_order}
\alias{chime_create_phone_number_order}
\title{Creates an order for phone numbers to be provisioned}
\usage{
chime_create_phone_number_order(ProductType, E164PhoneNumbers)
}
\arguments{
\item{ProductType}{[required] The phone number product type.}

\item{E164PhoneNumbers}{[required] List of phone numbers, in E.164 format.}
}
\value{
A list with the following syntax:\preformatted{list(
  PhoneNumberOrder = list(
    PhoneNumberOrderId = "string",
    ProductType = "BusinessCalling"|"VoiceConnector",
    Status = "Processing"|"Successful"|"Failed"|"Partial",
    OrderedPhoneNumbers = list(
      list(
        E164PhoneNumber = "string",
        Status = "Processing"|"Acquired"|"Failed"
      )
    ),
    CreatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    UpdatedTimestamp = as.POSIXct(
      "2015-01-01"
    )
  )
)
}
}
\description{
Creates an order for phone numbers to be provisioned. Choose from Amazon
Chime Business Calling and Amazon Chime Voice Connector product types.
For toll-free numbers, you must use the Amazon Chime Voice Connector
product type.
}
\section{Request syntax}{
\preformatted{svc$create_phone_number_order(
  ProductType = "BusinessCalling"|"VoiceConnector",
  E164PhoneNumbers = list(
    "string"
  )
)
}
}

\keyword{internal}
