% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_create_room}
\alias{chime_create_room}
\title{Creates a chat room for the specified Amazon Chime Enterprise account}
\usage{
chime_create_room(AccountId, Name, ClientRequestToken)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{Name}{[required] The room name.}

\item{ClientRequestToken}{The idempotency token for the request.}
}
\value{
A list with the following syntax:\preformatted{list(
  Room = list(
    RoomId = "string",
    Name = "string",
    AccountId = "string",
    CreatedBy = "string",
    CreatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    UpdatedTimestamp = as.POSIXct(
      "2015-01-01"
    )
  )
)
}
}
\description{
Creates a chat room for the specified Amazon Chime Enterprise account.
}
\section{Request syntax}{
\preformatted{svc$create_room(
  AccountId = "string",
  Name = "string",
  ClientRequestToken = "string"
)
}
}

\keyword{internal}
