% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_create_room_membership}
\alias{chime_create_room_membership}
\title{Adds a member to a chat room in an Amazon Chime Enterprise account}
\usage{
chime_create_room_membership(AccountId, RoomId, MemberId, Role)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{RoomId}{[required] The room ID.}

\item{MemberId}{[required] The Amazon Chime member ID (user ID or bot ID).}

\item{Role}{The role of the member.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  RoomMembership = list(
    RoomId = "string",
    Member = list(
      MemberId = "string",
      MemberType = "User"|"Bot"|"Webhook",
      Email = "string",
      FullName = "string",
      AccountId = "string"
    ),
    Role = "Administrator"|"Member",
    InvitedBy = "string",
    UpdatedTimestamp = as.POSIXct(
      "2015-01-01"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Adds a member to a chat room in an Amazon Chime Enterprise account. A
member can be either a user or a bot. The member role designates whether
the member is a chat room administrator or a general chat room member.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_room_membership(
  AccountId = "string",
  RoomId = "string",
  MemberId = "string",
  Role = "Administrator"|"Member"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
