% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_create_sip_media_application_call}
\alias{chime_create_sip_media_application_call}
\title{Creates an outbound call to a phone number from the phone number
specified in the request, and it invokes the endpoint of the specified
sipMediaApplicationId}
\usage{
chime_create_sip_media_application_call(FromPhoneNumber, ToPhoneNumber,
  SipMediaApplicationId)
}
\arguments{
\item{FromPhoneNumber}{The phone number that a user calls from.}

\item{ToPhoneNumber}{The phone number that the user dials in order to connect to a meeting}

\item{SipMediaApplicationId}{[required] The ID of the SIP media application.}
}
\description{
Creates an outbound call to a phone number from the phone number
specified in the request, and it invokes the endpoint of the specified
\code{sipMediaApplicationId}.
}
\section{Request syntax}{
\preformatted{svc$create_sip_media_application_call(
  FromPhoneNumber = "string",
  ToPhoneNumber = "string",
  SipMediaApplicationId = "string"
)
}
}

\keyword{internal}
