% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_create_voice_connector}
\alias{chime_create_voice_connector}
\title{Creates an Amazon Chime Voice Connector under the administrator's AWS
account}
\usage{
chime_create_voice_connector(Name, AwsRegion, RequireEncryption)
}
\arguments{
\item{Name}{[required] The name of the Amazon Chime Voice Connector.}

\item{AwsRegion}{The AWS Region in which the Amazon Chime Voice Connector is created.
Default value: \code{us-east-1}.}

\item{RequireEncryption}{[required] When enabled, requires encryption for the Amazon Chime Voice Connector.}
}
\description{
Creates an Amazon Chime Voice Connector under the administrator's AWS
account. You can choose to create an Amazon Chime Voice Connector in a
specific AWS Region.

Enabling CreateVoiceConnectorRequest$RequireEncryption configures your
Amazon Chime Voice Connector to use TLS transport for SIP signaling and
Secure RTP (SRTP) for media. Inbound calls use TLS transport, and
unencrypted outbound calls are blocked.
}
\section{Request syntax}{
\preformatted{svc$create_voice_connector(
  Name = "string",
  AwsRegion = "us-east-1"|"us-west-2",
  RequireEncryption = TRUE|FALSE
)
}
}

\keyword{internal}
