% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_delete_app_instance_admin}
\alias{chime_delete_app_instance_admin}
\title{Demotes an AppInstanceAdmin to an AppInstanceUser}
\usage{
chime_delete_app_instance_admin(AppInstanceAdminArn, AppInstanceArn)
}
\arguments{
\item{AppInstanceAdminArn}{[required] The ARN of the app instance's administrator.}

\item{AppInstanceArn}{[required] The ARN of the app instance.}
}
\value{
An empty list.
}
\description{
Demotes an \code{AppInstanceAdmin} to an \code{AppInstanceUser}. This action does
not delete the user.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_app_instance_admin(
  AppInstanceAdminArn = "string",
  AppInstanceArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
