% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_delete_app_instance_streaming_configurations}
\alias{chime_delete_app_instance_streaming_configurations}
\title{Deletes the streaming configurations of an app instance}
\usage{
chime_delete_app_instance_streaming_configurations(AppInstanceArn)
}
\arguments{
\item{AppInstanceArn}{[required] The ARN of the streaming configurations being deleted.}
}
\value{
An empty list.
}
\description{
Deletes the streaming configurations of an app instance.
}
\section{Request syntax}{
\preformatted{svc$delete_app_instance_streaming_configurations(
  AppInstanceArn = "string"
)
}
}

\keyword{internal}
