% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_delete_app_instance_user}
\alias{chime_delete_app_instance_user}
\title{Deletes an AppInstanceUser}
\usage{
chime_delete_app_instance_user(AppInstanceUserArn)
}
\arguments{
\item{AppInstanceUserArn}{[required] The ARN of the user request being deleted.}
}
\value{
An empty list.
}
\description{
Deletes an \code{AppInstanceUser}.
}
\section{Request syntax}{
\preformatted{svc$delete_app_instance_user(
  AppInstanceUserArn = "string"
)
}
}

\keyword{internal}
