% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_delete_events_configuration}
\alias{chime_delete_events_configuration}
\title{Deletes the events configuration that allows a bot to receive outgoing
events}
\usage{
chime_delete_events_configuration(AccountId, BotId)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{BotId}{[required] The bot ID.}
}
\value{
An empty list.
}
\description{
Deletes the events configuration that allows a bot to receive outgoing
events.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_events_configuration(
  AccountId = "string",
  BotId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
