% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_delete_meeting}
\alias{chime_delete_meeting}
\title{Deletes the specified Amazon Chime SDK meeting}
\usage{
chime_delete_meeting(MeetingId)
}
\arguments{
\item{MeetingId}{[required] The Amazon Chime SDK meeting ID.}
}
\description{
Deletes the specified Amazon Chime SDK meeting. When a meeting is
deleted, its attendees are also deleted and clients can no longer join
it. For more information about the Amazon Chime SDK, see \href{https://docs.aws.amazon.com/chime/latest/dg/meetings-sdk.html}{Using the Amazon Chime SDK} in
the \emph{Amazon Chime Developer Guide}.
}
\section{Request syntax}{
\preformatted{svc$delete_meeting(
  MeetingId = "string"
)
}
}

\keyword{internal}
