% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_delete_phone_number}
\alias{chime_delete_phone_number}
\title{Moves the specified phone number into the Deletion queue}
\usage{
chime_delete_phone_number(PhoneNumberId)
}
\arguments{
\item{PhoneNumberId}{[required] The phone number ID.}
}
\value{
An empty list.
}
\description{
Moves the specified phone number into the \strong{Deletion queue}. A phone
number must be disassociated from any users or Amazon Chime Voice
Connectors before it can be deleted.

Deleted phone numbers remain in the \strong{Deletion queue} for 7 days before
they are deleted permanently.
}
\section{Request syntax}{
\preformatted{svc$delete_phone_number(
  PhoneNumberId = "string"
)
}
}

\keyword{internal}
