% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_delete_room}
\alias{chime_delete_room}
\title{Deletes a chat room in an Amazon Chime Enterprise account}
\usage{
chime_delete_room(AccountId, RoomId)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{RoomId}{[required] The chat room ID.}
}
\value{
An empty list.
}
\description{
Deletes a chat room in an Amazon Chime Enterprise account.
}
\section{Request syntax}{
\preformatted{svc$delete_room(
  AccountId = "string",
  RoomId = "string"
)
}
}

\keyword{internal}
