% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_delete_sip_media_application}
\alias{chime_delete_sip_media_application}
\title{Deletes a SIP media application}
\usage{
chime_delete_sip_media_application(SipMediaApplicationId)
}
\arguments{
\item{SipMediaApplicationId}{[required] The SIP media application ID.}
}
\value{
An empty list.
}
\description{
Deletes a SIP media application.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_sip_media_application(
  SipMediaApplicationId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
