% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_delete_sip_rule}
\alias{chime_delete_sip_rule}
\title{Deletes a SIP rule}
\usage{
chime_delete_sip_rule(SipRuleId)
}
\arguments{
\item{SipRuleId}{[required] The SIP rule ID.}
}
\value{
An empty list.
}
\description{
Deletes a SIP rule. You must disable a SIP rule before you can delete
it.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_sip_rule(
  SipRuleId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
