% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_delete_voice_connector_emergency_calling_configuration}
\alias{chime_delete_voice_connector_emergency_calling_configuration}
\title{Deletes the emergency calling configuration details from the specified
Amazon Chime Voice Connector}
\usage{
chime_delete_voice_connector_emergency_calling_configuration(
  VoiceConnectorId)
}
\arguments{
\item{VoiceConnectorId}{[required] The Amazon Chime Voice Connector ID.}
}
\value{
An empty list.
}
\description{
Deletes the emergency calling configuration details from the specified
Amazon Chime Voice Connector.
}
\section{Request syntax}{
\preformatted{svc$delete_voice_connector_emergency_calling_configuration(
  VoiceConnectorId = "string"
)
}
}

\keyword{internal}
