% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_delete_voice_connector_group}
\alias{chime_delete_voice_connector_group}
\title{Deletes the specified Amazon Chime Voice Connector group}
\usage{
chime_delete_voice_connector_group(VoiceConnectorGroupId)
}
\arguments{
\item{VoiceConnectorGroupId}{[required] The Amazon Chime Voice Connector group ID.}
}
\value{
An empty list.
}
\description{
Deletes the specified Amazon Chime Voice Connector group. Any
\code{VoiceConnectorItems} and phone numbers associated with the group must
be removed before it can be deleted.
}
\section{Request syntax}{
\preformatted{svc$delete_voice_connector_group(
  VoiceConnectorGroupId = "string"
)
}
}

\keyword{internal}
