% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_delete_voice_connector_origination}
\alias{chime_delete_voice_connector_origination}
\title{Deletes the origination settings for the specified Amazon Chime Voice
Connector}
\usage{
chime_delete_voice_connector_origination(VoiceConnectorId)
}
\arguments{
\item{VoiceConnectorId}{[required] The Amazon Chime Voice Connector ID.}
}
\value{
An empty list.
}
\description{
Deletes the origination settings for the specified Amazon Chime Voice
Connector.

If emergency calling is configured for the Amazon Chime Voice Connector,
it must be deleted prior to deleting the origination settings.
}
\section{Request syntax}{
\preformatted{svc$delete_voice_connector_origination(
  VoiceConnectorId = "string"
)
}
}

\keyword{internal}
