% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_delete_voice_connector_streaming_configuration}
\alias{chime_delete_voice_connector_streaming_configuration}
\title{Deletes the streaming configuration for the specified Amazon Chime Voice
Connector}
\usage{
chime_delete_voice_connector_streaming_configuration(VoiceConnectorId)
}
\arguments{
\item{VoiceConnectorId}{[required] The Amazon Chime Voice Connector ID.}
}
\value{
An empty list.
}
\description{
Deletes the streaming configuration for the specified Amazon Chime Voice
Connector.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_voice_connector_streaming_configuration(
  VoiceConnectorId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
