% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_delete_voice_connector_termination_credentials}
\alias{chime_delete_voice_connector_termination_credentials}
\title{Deletes the specified SIP credentials used by your equipment to
authenticate during call termination}
\usage{
chime_delete_voice_connector_termination_credentials(VoiceConnectorId,
  Usernames)
}
\arguments{
\item{VoiceConnectorId}{[required] The Amazon Chime Voice Connector ID.}

\item{Usernames}{[required] The RFC2617 compliant username associated with the SIP credentials, in
US-ASCII format.}
}
\value{
An empty list.
}
\description{
Deletes the specified SIP credentials used by your equipment to
authenticate during call termination.
}
\section{Request syntax}{
\preformatted{svc$delete_voice_connector_termination_credentials(
  VoiceConnectorId = "string",
  Usernames = list(
    "string"
  )
)
}
}

\keyword{internal}
