% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_describe_app_instance_user}
\alias{chime_describe_app_instance_user}
\title{Returns the full details of an AppInstanceUser}
\usage{
chime_describe_app_instance_user(AppInstanceUserArn)
}
\arguments{
\item{AppInstanceUserArn}{[required] The ARN of the \code{AppInstanceUser}.}
}
\value{
A list with the following syntax:\preformatted{list(
  AppInstanceUser = list(
    AppInstanceUserArn = "string",
    Name = "string",
    CreatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    Metadata = "string",
    LastUpdatedTimestamp = as.POSIXct(
      "2015-01-01"
    )
  )
)
}
}
\description{
Returns the full details of an \code{AppInstanceUser} .
}
\section{Request syntax}{
\preformatted{svc$describe_app_instance_user(
  AppInstanceUserArn = "string"
)
}
}

\keyword{internal}
