% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_describe_channel}
\alias{chime_describe_channel}
\title{Returns the full details of a channel in an Amazon Chime AppInstance}
\usage{
chime_describe_channel(ChannelArn, ChimeBearer)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel.}

\item{ChimeBearer}{The \code{AppInstanceUserArn} of the user that makes the API call.}
}
\value{
A list with the following syntax:\preformatted{list(
  Channel = list(
    Name = "string",
    ChannelArn = "string",
    Mode = "UNRESTRICTED"|"RESTRICTED",
    Privacy = "PUBLIC"|"PRIVATE",
    Metadata = "string",
    CreatedBy = list(
      Arn = "string",
      Name = "string"
    ),
    CreatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    LastMessageTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    LastUpdatedTimestamp = as.POSIXct(
      "2015-01-01"
    )
  )
)
}
}
\description{
Returns the full details of a channel in an Amazon Chime \code{AppInstance}.

The \code{x-amz-chime-bearer} request header is mandatory. Use the
\code{AppInstanceUserArn} of the user that makes the API call as the value in
the header.
}
\section{Request syntax}{
\preformatted{svc$describe_channel(
  ChannelArn = "string",
  ChimeBearer = "string"
)
}
}

\keyword{internal}
