% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_describe_channel_ban}
\alias{chime_describe_channel_ban}
\title{Returns the full details of a channel ban}
\usage{
chime_describe_channel_ban(ChannelArn, MemberArn)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel from which the user is banned.}

\item{MemberArn}{[required] The ARN of the member being banned.}
}
\value{
A list with the following syntax:\preformatted{list(
  ChannelBan = list(
    Member = list(
      Arn = "string",
      Name = "string"
    ),
    ChannelArn = "string",
    CreatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    CreatedBy = list(
      Arn = "string",
      Name = "string"
    )
  )
)
}
}
\description{
Returns the full details of a channel ban.
}
\section{Request syntax}{
\preformatted{svc$describe_channel_ban(
  ChannelArn = "string",
  MemberArn = "string"
)
}
}

\keyword{internal}
