% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_describe_channel_membership}
\alias{chime_describe_channel_membership}
\title{Returns the full details of a user's channel membership}
\usage{
chime_describe_channel_membership(ChannelArn, MemberArn)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel.}

\item{MemberArn}{[required] The ARN of the member.}
}
\description{
Returns the full details of a user's channel membership.
}
\section{Request syntax}{
\preformatted{svc$describe_channel_membership(
  ChannelArn = "string",
  MemberArn = "string"
)
}
}

\keyword{internal}
