% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_describe_channel_membership_for_app_instance_user}
\alias{chime_describe_channel_membership_for_app_instance_user}
\title{Returns the details of a channel based on the membership of the
AppInstanceUser specified}
\usage{
chime_describe_channel_membership_for_app_instance_user(ChannelArn,
  AppInstanceUserArn)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel to which the user belongs.}

\item{AppInstanceUserArn}{[required] The ARN of the user in a channel.}
}
\description{
Returns the details of a channel based on the membership of the
\code{AppInstanceUser} specified.
}
\section{Request syntax}{
\preformatted{svc$describe_channel_membership_for_app_instance_user(
  ChannelArn = "string",
  AppInstanceUserArn = "string"
)
}
}

\keyword{internal}
