% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_describe_channel_moderated_by_app_instance_user}
\alias{chime_describe_channel_moderated_by_app_instance_user}
\title{Returns the full details of a channel moderated by the specified
AppInstanceUser}
\usage{
chime_describe_channel_moderated_by_app_instance_user(ChannelArn,
  AppInstanceUserArn, ChimeBearer)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the moderated channel.}

\item{AppInstanceUserArn}{[required] The ARN of the \code{AppInstanceUser} in the moderated channel.}

\item{ChimeBearer}{The \code{AppInstanceUserArn} of the user that makes the API call.}
}
\value{
A list with the following syntax:\preformatted{list(
  Channel = list(
    ChannelSummary = list(
      Name = "string",
      ChannelArn = "string",
      Mode = "UNRESTRICTED"|"RESTRICTED",
      Privacy = "PUBLIC"|"PRIVATE",
      Metadata = "string",
      LastMessageTimestamp = as.POSIXct(
        "2015-01-01"
      )
    )
  )
)
}
}
\description{
Returns the full details of a channel moderated by the specified
\code{AppInstanceUser}.

The \code{x-amz-chime-bearer} request header is mandatory. Use the
\code{AppInstanceUserArn} of the user that makes the API call as the value in
the header.
}
\section{Request syntax}{
\preformatted{svc$describe_channel_moderated_by_app_instance_user(
  ChannelArn = "string",
  AppInstanceUserArn = "string",
  ChimeBearer = "string"
)
}
}

\keyword{internal}
