% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_disassociate_phone_number_from_user}
\alias{chime_disassociate_phone_number_from_user}
\title{Disassociates the primary provisioned phone number from the specified
Amazon Chime user}
\usage{
chime_disassociate_phone_number_from_user(AccountId, UserId)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{UserId}{[required] The user ID.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Disassociates the primary provisioned phone number from the specified
Amazon Chime user.
}
\section{Request syntax}{
\preformatted{svc$disassociate_phone_number_from_user(
  AccountId = "string",
  UserId = "string"
)
}
}

\keyword{internal}
