% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_disassociate_phone_numbers_from_voice_connector}
\alias{chime_disassociate_phone_numbers_from_voice_connector}
\title{Disassociates the specified phone numbers from the specified Amazon
Chime Voice Connector}
\usage{
chime_disassociate_phone_numbers_from_voice_connector(VoiceConnectorId,
  E164PhoneNumbers)
}
\arguments{
\item{VoiceConnectorId}{[required] The Amazon Chime Voice Connector ID.}

\item{E164PhoneNumbers}{[required] List of phone numbers, in E.164 format.}
}
\value{
A list with the following syntax:\preformatted{list(
  PhoneNumberErrors = list(
    list(
      PhoneNumberId = "string",
      ErrorCode = "BadRequest"|"Conflict"|"Forbidden"|"NotFound"|"PreconditionFailed"|"ResourceLimitExceeded"|"ServiceFailure"|"AccessDenied"|"ServiceUnavailable"|"Throttled"|"Throttling"|"Unauthorized"|"Unprocessable"|"VoiceConnectorGroupAssociationsExist"|"PhoneNumberAssociationsExist",
      ErrorMessage = "string"
    )
  )
)
}
}
\description{
Disassociates the specified phone numbers from the specified Amazon
Chime Voice Connector.
}
\section{Request syntax}{
\preformatted{svc$disassociate_phone_numbers_from_voice_connector(
  VoiceConnectorId = "string",
  E164PhoneNumbers = list(
    "string"
  )
)
}
}

\keyword{internal}
