% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_disassociate_signin_delegate_groups_from_account}
\alias{chime_disassociate_signin_delegate_groups_from_account}
\title{Disassociates the specified sign-in delegate groups from the specified
Amazon Chime account}
\usage{
chime_disassociate_signin_delegate_groups_from_account(AccountId,
  GroupNames)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{GroupNames}{[required] The sign-in delegate group names.}
}
\description{
Disassociates the specified sign-in delegate groups from the specified
Amazon Chime account.
}
\section{Request syntax}{
\preformatted{svc$disassociate_signin_delegate_groups_from_account(
  AccountId = "string",
  GroupNames = list(
    "string"
  )
)
}
}

\keyword{internal}
