% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_account_settings}
\alias{chime_get_account_settings}
\title{Retrieves account settings for the specified Amazon Chime account ID,
such as remote control and dial out settings}
\usage{
chime_get_account_settings(AccountId)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}
}
\value{
A list with the following syntax:\preformatted{list(
  AccountSettings = list(
    DisableRemoteControl = TRUE|FALSE,
    EnableDialOut = TRUE|FALSE
  )
)
}
}
\description{
Retrieves account settings for the specified Amazon Chime account ID,
such as remote control and dial out settings. For more information about
these settings, see \href{https://docs.aws.amazon.com/chime/latest/ag/mtg-settings.html}{Use the Policies Page} in
the \emph{Amazon Chime Administration Guide}.
}
\section{Request syntax}{
\preformatted{svc$get_account_settings(
  AccountId = "string"
)
}
}

\keyword{internal}
