% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_app_instance_retention_settings}
\alias{chime_get_app_instance_retention_settings}
\title{Gets the retention settings for an app instance}
\usage{
chime_get_app_instance_retention_settings(AppInstanceArn)
}
\arguments{
\item{AppInstanceArn}{[required] The ARN of the app instance.}
}
\description{
Gets the retention settings for an app instance.
}
\section{Request syntax}{
\preformatted{svc$get_app_instance_retention_settings(
  AppInstanceArn = "string"
)
}
}

\keyword{internal}
