% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_attendee}
\alias{chime_get_attendee}
\title{Gets the Amazon Chime SDK attendee details for a specified meeting ID
and attendee ID}
\usage{
chime_get_attendee(MeetingId, AttendeeId)
}
\arguments{
\item{MeetingId}{[required] The Amazon Chime SDK meeting ID.}

\item{AttendeeId}{[required] The Amazon Chime SDK attendee ID.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Attendee = list(
    ExternalUserId = "string",
    AttendeeId = "string",
    JoinToken = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets the Amazon Chime SDK attendee details for a specified meeting ID
and attendee ID. For more information about the Amazon Chime SDK, see
\href{https://docs.aws.amazon.com/chime-sdk/latest/dg/meetings-sdk.html}{Using the Amazon Chime SDK}
in the \emph{Amazon Chime Developer Guide}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_attendee(
  MeetingId = "string",
  AttendeeId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
