% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_bot}
\alias{chime_get_bot}
\title{Retrieves details for the specified bot, such as bot email address, bot
type, status, and display name}
\usage{
chime_get_bot(AccountId, BotId)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{BotId}{[required] The bot ID.}
}
\description{
Retrieves details for the specified bot, such as bot email address, bot
type, status, and display name.
}
\section{Request syntax}{
\preformatted{svc$get_bot(
  AccountId = "string",
  BotId = "string"
)
}
}

\keyword{internal}
