% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_channel_message}
\alias{chime_get_channel_message}
\title{Gets the full details of a channel message}
\usage{
chime_get_channel_message(ChannelArn, MessageId, ChimeBearer)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel.}

\item{MessageId}{[required] The ID of the message.}

\item{ChimeBearer}{The \code{AppInstanceUserArn} of the user that makes the API call.}
}
\value{
A list with the following syntax:\preformatted{list(
  ChannelMessage = list(
    ChannelArn = "string",
    MessageId = "string",
    Content = "string",
    Metadata = "string",
    Type = "STANDARD"|"CONTROL",
    CreatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    LastEditedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    LastUpdatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    Sender = list(
      Arn = "string",
      Name = "string"
    ),
    Redacted = TRUE|FALSE,
    Persistence = "PERSISTENT"|"NON_PERSISTENT"
  )
)
}
}
\description{
Gets the full details of a channel message.

The x-amz-chime-bearer request header is mandatory. Use the
\code{AppInstanceUserArn} of the user that makes the API call as the value in
the header.
}
\section{Request syntax}{
\preformatted{svc$get_channel_message(
  ChannelArn = "string",
  MessageId = "string",
  ChimeBearer = "string"
)
}
}

\keyword{internal}
