% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_events_configuration}
\alias{chime_get_events_configuration}
\title{Gets details for an events configuration that allows a bot to receive
outgoing events, such as an HTTPS endpoint or Lambda function ARN}
\usage{
chime_get_events_configuration(AccountId, BotId)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{BotId}{[required] The bot ID.}
}
\value{
A list with the following syntax:\preformatted{list(
  EventsConfiguration = list(
    BotId = "string",
    OutboundEventsHTTPSEndpoint = "string",
    LambdaFunctionArn = "string"
  )
)
}
}
\description{
Gets details for an events configuration that allows a bot to receive
outgoing events, such as an HTTPS endpoint or Lambda function ARN.
}
\section{Request syntax}{
\preformatted{svc$get_events_configuration(
  AccountId = "string",
  BotId = "string"
)
}
}

\keyword{internal}
