% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_global_settings}
\alias{chime_get_global_settings}
\title{Retrieves global settings for the administrator's AWS account, such as
Amazon Chime Business Calling and Amazon Chime Voice Connector settings}
\usage{
chime_get_global_settings()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  BusinessCalling = list(
    CdrBucket = "string"
  ),
  VoiceConnector = list(
    CdrBucket = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves global settings for the administrator's AWS account, such as
Amazon Chime Business Calling and Amazon Chime Voice Connector settings.
}
\keyword{internal}
