% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_meeting}
\alias{chime_get_meeting}
\title{Gets the Amazon Chime SDK meeting details for the specified meeting ID}
\usage{
chime_get_meeting(MeetingId)
}
\arguments{
\item{MeetingId}{[required] The Amazon Chime SDK meeting ID.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Meeting = list(
    MeetingId = "string",
    ExternalMeetingId = "string",
    MediaPlacement = list(
      AudioHostUrl = "string",
      AudioFallbackUrl = "string",
      ScreenDataUrl = "string",
      ScreenSharingUrl = "string",
      ScreenViewingUrl = "string",
      SignalingUrl = "string",
      TurnControlUrl = "string"
    ),
    MediaRegion = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets the Amazon Chime SDK meeting details for the specified meeting ID.
For more information about the Amazon Chime SDK, see \href{https://docs.aws.amazon.com/chime-sdk/latest/dg/meetings-sdk.html}{Using the Amazon Chime SDK}
in the \emph{Amazon Chime Developer Guide}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_meeting(
  MeetingId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
