% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_messaging_session_endpoint}
\alias{chime_get_messaging_session_endpoint}
\title{The endpoint for the messaging session}
\usage{
chime_get_messaging_session_endpoint()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Endpoint = list(
    Url = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
The endpoint for the messaging session.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_messaging_session_endpoint()
}\if{html}{\out{</div>}}
}

\keyword{internal}
