% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_phone_number}
\alias{chime_get_phone_number}
\title{Retrieves details for the specified phone number ID, such as
associations, capabilities, and product type}
\usage{
chime_get_phone_number(PhoneNumberId)
}
\arguments{
\item{PhoneNumberId}{[required] The phone number ID.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  PhoneNumber = list(
    PhoneNumberId = "string",
    E164PhoneNumber = "string",
    Type = "Local"|"TollFree",
    ProductType = "BusinessCalling"|"VoiceConnector",
    Status = "AcquireInProgress"|"AcquireFailed"|"Unassigned"|"Assigned"|"ReleaseInProgress"|"DeleteInProgress"|"ReleaseFailed"|"DeleteFailed",
    Capabilities = list(
      InboundCall = TRUE|FALSE,
      OutboundCall = TRUE|FALSE,
      InboundSMS = TRUE|FALSE,
      OutboundSMS = TRUE|FALSE,
      InboundMMS = TRUE|FALSE,
      OutboundMMS = TRUE|FALSE
    ),
    Associations = list(
      list(
        Value = "string",
        Name = "AccountId"|"UserId"|"VoiceConnectorId"|"VoiceConnectorGroupId"|"SipRuleId",
        AssociatedTimestamp = as.POSIXct(
          "2015-01-01"
        )
      )
    ),
    CallingName = "string",
    CallingNameStatus = "Unassigned"|"UpdateInProgress"|"UpdateSucceeded"|"UpdateFailed",
    CreatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    UpdatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    DeletionTimestamp = as.POSIXct(
      "2015-01-01"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves details for the specified phone number ID, such as
associations, capabilities, and product type.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_phone_number(
  PhoneNumberId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
