% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_phone_number_order}
\alias{chime_get_phone_number_order}
\title{Retrieves details for the specified phone number order, such as order
creation timestamp, phone numbers in E}
\usage{
chime_get_phone_number_order(PhoneNumberOrderId)
}
\arguments{
\item{PhoneNumberOrderId}{[required] The ID for the phone number order.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  PhoneNumberOrder = list(
    PhoneNumberOrderId = "string",
    ProductType = "BusinessCalling"|"VoiceConnector",
    Status = "Processing"|"Successful"|"Failed"|"Partial",
    OrderedPhoneNumbers = list(
      list(
        E164PhoneNumber = "string",
        Status = "Processing"|"Acquired"|"Failed"
      )
    ),
    CreatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    UpdatedTimestamp = as.POSIXct(
      "2015-01-01"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves details for the specified phone number order, such as order
creation timestamp, phone numbers in E.164 format, product type, and
order status.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_phone_number_order(
  PhoneNumberOrderId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
