% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_proxy_session}
\alias{chime_get_proxy_session}
\title{Gets the specified proxy session details for the specified Amazon Chime
Voice Connector}
\usage{
chime_get_proxy_session(VoiceConnectorId, ProxySessionId)
}
\arguments{
\item{VoiceConnectorId}{[required] The Amazon Chime voice connector ID.}

\item{ProxySessionId}{[required] The proxy session ID.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ProxySession = list(
    VoiceConnectorId = "string",
    ProxySessionId = "string",
    Name = "string",
    Status = "Open"|"InProgress"|"Closed",
    ExpiryMinutes = 123,
    Capabilities = list(
      "Voice"|"SMS"
    ),
    CreatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    UpdatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    EndedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    Participants = list(
      list(
        PhoneNumber = "string",
        ProxyPhoneNumber = "string"
      )
    ),
    NumberSelectionBehavior = "PreferSticky"|"AvoidSticky",
    GeoMatchLevel = "Country"|"AreaCode",
    GeoMatchParams = list(
      Country = "string",
      AreaCode = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets the specified proxy session details for the specified Amazon Chime
Voice Connector.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_proxy_session(
  VoiceConnectorId = "string",
  ProxySessionId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
