% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_retention_settings}
\alias{chime_get_retention_settings}
\title{Gets the retention settings for the specified Amazon Chime Enterprise
account}
\usage{
chime_get_retention_settings(AccountId)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  RetentionSettings = list(
    RoomRetentionSettings = list(
      RetentionDays = 123
    ),
    ConversationRetentionSettings = list(
      RetentionDays = 123
    )
  ),
  InitiateDeletionTimestamp = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets the retention settings for the specified Amazon Chime Enterprise
account. For more information about retention settings, see \href{https://docs.aws.amazon.com/chime/latest/ag/archive-retention.html}{Managing Chat Retention Policies}
in the \emph{Amazon Chime Administration Guide}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_retention_settings(
  AccountId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
