% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_sip_media_application}
\alias{chime_get_sip_media_application}
\title{Retrieves the information for a SIP media application, including name,
AWS Region, and endpoints}
\usage{
chime_get_sip_media_application(SipMediaApplicationId)
}
\arguments{
\item{SipMediaApplicationId}{[required] The SIP media application ID.}
}
\description{
Retrieves the information for a SIP media application, including name,
AWS Region, and endpoints.
}
\section{Request syntax}{
\preformatted{svc$get_sip_media_application(
  SipMediaApplicationId = "string"
)
}
}

\keyword{internal}
