% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_sip_rule}
\alias{chime_get_sip_rule}
\title{Retrieves the details of a SIP rule, such as the rule ID, name,
triggers, and target endpoints}
\usage{
chime_get_sip_rule(SipRuleId)
}
\arguments{
\item{SipRuleId}{[required] The SIP rule ID.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  SipRule = list(
    SipRuleId = "string",
    Name = "string",
    Disabled = TRUE|FALSE,
    TriggerType = "ToPhoneNumber"|"RequestUriHostname",
    TriggerValue = "string",
    TargetApplications = list(
      list(
        SipMediaApplicationId = "string",
        Priority = 123,
        AwsRegion = "string"
      )
    ),
    CreatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    UpdatedTimestamp = as.POSIXct(
      "2015-01-01"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the details of a SIP rule, such as the rule ID, name,
triggers, and target endpoints.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_sip_rule(
  SipRuleId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
