% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_user}
\alias{chime_get_user}
\title{Retrieves details for the specified user ID, such as primary email
address, license type, and personal meeting PIN}
\usage{
chime_get_user(AccountId, UserId)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{UserId}{[required] The user ID.}
}
\value{
A list with the following syntax:\preformatted{list(
  User = list(
    UserId = "string",
    AccountId = "string",
    PrimaryEmail = "string",
    PrimaryProvisionedNumber = "string",
    DisplayName = "string",
    LicenseType = "Basic"|"Plus"|"Pro"|"ProTrial",
    UserType = "PrivateUser"|"SharedDevice",
    UserRegistrationStatus = "Unregistered"|"Registered"|"Suspended",
    UserInvitationStatus = "Pending"|"Accepted"|"Failed",
    RegisteredOn = as.POSIXct(
      "2015-01-01"
    ),
    InvitedOn = as.POSIXct(
      "2015-01-01"
    ),
    AlexaForBusinessMetadata = list(
      IsAlexaForBusinessEnabled = TRUE|FALSE,
      AlexaForBusinessRoomArn = "string"
    ),
    PersonalPIN = "string"
  )
)
}
}
\description{
Retrieves details for the specified user ID, such as primary email
address, license type, and personal meeting PIN.

To retrieve user details with an email address instead of a user ID, use
the \code{\link[=chime_list_users]{list_users}} action, and then filter by email
address.
}
\section{Request syntax}{
\preformatted{svc$get_user(
  AccountId = "string",
  UserId = "string"
)
}
}

\keyword{internal}
