% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_voice_connector_emergency_calling_configuration}
\alias{chime_get_voice_connector_emergency_calling_configuration}
\title{Gets the emergency calling configuration details for the specified
Amazon Chime Voice Connector}
\usage{
chime_get_voice_connector_emergency_calling_configuration(
  VoiceConnectorId)
}
\arguments{
\item{VoiceConnectorId}{[required] The Amazon Chime Voice Connector ID.}
}
\value{
A list with the following syntax:\preformatted{list(
  EmergencyCallingConfiguration = list(
    DNIS = list(
      list(
        EmergencyPhoneNumber = "string",
        TestPhoneNumber = "string",
        CallingCountry = "string"
      )
    )
  )
)
}
}
\description{
Gets the emergency calling configuration details for the specified
Amazon Chime Voice Connector.
}
\section{Request syntax}{
\preformatted{svc$get_voice_connector_emergency_calling_configuration(
  VoiceConnectorId = "string"
)
}
}

\keyword{internal}
