% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_voice_connector_logging_configuration}
\alias{chime_get_voice_connector_logging_configuration}
\title{Retrieves the logging configuration details for the specified Amazon
Chime Voice Connector}
\usage{
chime_get_voice_connector_logging_configuration(VoiceConnectorId)
}
\arguments{
\item{VoiceConnectorId}{[required] The Amazon Chime Voice Connector ID.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  LoggingConfiguration = list(
    EnableSIPLogs = TRUE|FALSE
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the logging configuration details for the specified Amazon
Chime Voice Connector. Shows whether SIP message logs are enabled for
sending to Amazon CloudWatch.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_voice_connector_logging_configuration(
  VoiceConnectorId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
