% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_voice_connector_origination}
\alias{chime_get_voice_connector_origination}
\title{Retrieves origination setting details for the specified Amazon Chime
Voice Connector}
\usage{
chime_get_voice_connector_origination(VoiceConnectorId)
}
\arguments{
\item{VoiceConnectorId}{[required] The Amazon Chime Voice Connector ID.}
}
\value{
A list with the following syntax:\preformatted{list(
  Origination = list(
    Routes = list(
      list(
        Host = "string",
        Port = 123,
        Protocol = "TCP"|"UDP",
        Priority = 123,
        Weight = 123
      )
    ),
    Disabled = TRUE|FALSE
  )
)
}
}
\description{
Retrieves origination setting details for the specified Amazon Chime
Voice Connector.
}
\section{Request syntax}{
\preformatted{svc$get_voice_connector_origination(
  VoiceConnectorId = "string"
)
}
}

\keyword{internal}
