% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_voice_connector_proxy}
\alias{chime_get_voice_connector_proxy}
\title{Gets the proxy configuration details for the specified Amazon Chime
Voice Connector}
\usage{
chime_get_voice_connector_proxy(VoiceConnectorId)
}
\arguments{
\item{VoiceConnectorId}{[required] The Amazon Chime voice connector ID.}
}
\value{
A list with the following syntax:\preformatted{list(
  Proxy = list(
    DefaultSessionExpiryMinutes = 123,
    Disabled = TRUE|FALSE,
    FallBackPhoneNumber = "string",
    PhoneNumberCountries = list(
      "string"
    )
  )
)
}
}
\description{
Gets the proxy configuration details for the specified Amazon Chime
Voice Connector.
}
\section{Request syntax}{
\preformatted{svc$get_voice_connector_proxy(
  VoiceConnectorId = "string"
)
}
}

\keyword{internal}
