% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_voice_connector_streaming_configuration}
\alias{chime_get_voice_connector_streaming_configuration}
\title{Retrieves the streaming configuration details for the specified Amazon
Chime Voice Connector}
\usage{
chime_get_voice_connector_streaming_configuration(VoiceConnectorId)
}
\arguments{
\item{VoiceConnectorId}{[required] The Amazon Chime Voice Connector ID.}
}
\description{
Retrieves the streaming configuration details for the specified Amazon
Chime Voice Connector. Shows whether media streaming is enabled for
sending to Amazon Kinesis. It also shows the retention period, in hours,
for the Amazon Kinesis data.
}
\section{Request syntax}{
\preformatted{svc$get_voice_connector_streaming_configuration(
  VoiceConnectorId = "string"
)
}
}

\keyword{internal}
