% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_voice_connector_termination}
\alias{chime_get_voice_connector_termination}
\title{Retrieves termination setting details for the specified Amazon Chime
Voice Connector}
\usage{
chime_get_voice_connector_termination(VoiceConnectorId)
}
\arguments{
\item{VoiceConnectorId}{[required] The Amazon Chime Voice Connector ID.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Termination = list(
    CpsLimit = 123,
    DefaultPhoneNumber = "string",
    CallingRegions = list(
      "string"
    ),
    CidrAllowedList = list(
      "string"
    ),
    Disabled = TRUE|FALSE
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves termination setting details for the specified Amazon Chime
Voice Connector.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_voice_connector_termination(
  VoiceConnectorId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
