% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_voice_connector_termination_health}
\alias{chime_get_voice_connector_termination_health}
\title{Retrieves information about the last time a SIP OPTIONS ping was
received from your SIP infrastructure for the specified Amazon Chime
Voice Connector}
\usage{
chime_get_voice_connector_termination_health(VoiceConnectorId)
}
\arguments{
\item{VoiceConnectorId}{[required] The Amazon Chime Voice Connector ID.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  TerminationHealth = list(
    Timestamp = as.POSIXct(
      "2015-01-01"
    ),
    Source = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves information about the last time a SIP \code{OPTIONS} ping was
received from your SIP infrastructure for the specified Amazon Chime
Voice Connector.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_voice_connector_termination_health(
  VoiceConnectorId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
