% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_list_app_instance_admins}
\alias{chime_list_app_instance_admins}
\title{Returns a list of the administrators in the app instance}
\usage{
chime_list_app_instance_admins(AppInstanceArn, MaxResults, NextToken)
}
\arguments{
\item{AppInstanceArn}{[required] The ARN of the app instance.}

\item{MaxResults}{The maximum number of administrators that you want to return.}

\item{NextToken}{The token returned from previous API requests until the number of
administrators is reached.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AppInstanceArn = "string",
  AppInstanceAdmins = list(
    list(
      Admin = list(
        Arn = "string",
        Name = "string"
      )
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns a list of the administrators in the app instance.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_app_instance_admins(
  AppInstanceArn = "string",
  MaxResults = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
