% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_list_app_instance_users}
\alias{chime_list_app_instance_users}
\title{List all AppInstanceUsers created under a single app instance}
\usage{
chime_list_app_instance_users(AppInstanceArn, MaxResults, NextToken)
}
\arguments{
\item{AppInstanceArn}{[required] The ARN of the app instance.}

\item{MaxResults}{The maximum number of requests that you want returned.}

\item{NextToken}{The token passed by previous API calls until all requested users are
returned.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AppInstanceArn = "string",
  AppInstanceUsers = list(
    list(
      AppInstanceUserArn = "string",
      Name = "string",
      Metadata = "string"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
List all \code{AppInstanceUsers} created under a single app instance.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_app_instance_users(
  AppInstanceArn = "string",
  MaxResults = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
