% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_list_bots}
\alias{chime_list_bots}
\title{Lists the bots associated with the administrator's Amazon Chime
Enterprise account ID}
\usage{
chime_list_bots(AccountId, MaxResults, NextToken)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{MaxResults}{The maximum number of results to return in a single call. The default is
10.}

\item{NextToken}{The token to use to retrieve the next page of results.}
}
\value{
A list with the following syntax:\preformatted{list(
  Bots = list(
    list(
      BotId = "string",
      UserId = "string",
      DisplayName = "string",
      BotType = "ChatBot",
      Disabled = TRUE|FALSE,
      CreatedTimestamp = as.POSIXct(
        "2015-01-01"
      ),
      UpdatedTimestamp = as.POSIXct(
        "2015-01-01"
      ),
      BotEmail = "string",
      SecurityToken = "string"
    )
  ),
  NextToken = "string"
)
}
}
\description{
Lists the bots associated with the administrator's Amazon Chime
Enterprise account ID.
}
\section{Request syntax}{
\preformatted{svc$list_bots(
  AccountId = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
