% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_list_channel_bans}
\alias{chime_list_channel_bans}
\title{Lists all the users banned from a particular channel}
\usage{
chime_list_channel_bans(ChannelArn, MaxResults, NextToken)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel.}

\item{MaxResults}{The maximum number of bans that you want returned.}

\item{NextToken}{The token passed by previous API calls until all requested bans are
returned.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ChannelArn = "string",
  NextToken = "string",
  ChannelBans = list(
    list(
      Member = list(
        Arn = "string",
        Name = "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Lists all the users banned from a particular channel.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_channel_bans(
  ChannelArn = "string",
  MaxResults = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
