% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_list_channel_memberships_for_app_instance_user}
\alias{chime_list_channel_memberships_for_app_instance_user}
\title{Lists all channels that a particular AppInstanceUser is a part of}
\usage{
chime_list_channel_memberships_for_app_instance_user(AppInstanceUserArn,
  MaxResults, NextToken)
}
\arguments{
\item{AppInstanceUserArn}{The ARN of the app instance users}

\item{MaxResults}{The maximum number of users that you want returned.}

\item{NextToken}{The token returned from previous API requests until the number of
channel memberships is reached.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ChannelMemberships = list(
    list(
      ChannelSummary = list(
        Name = "string",
        ChannelArn = "string",
        Mode = "UNRESTRICTED"|"RESTRICTED",
        Privacy = "PUBLIC"|"PRIVATE",
        Metadata = "string",
        LastMessageTimestamp = as.POSIXct(
          "2015-01-01"
        )
      ),
      AppInstanceUserMembershipSummary = list(
        Type = "DEFAULT"|"HIDDEN",
        ReadMarkerTimestamp = as.POSIXct(
          "2015-01-01"
        )
      )
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists all channels that a particular \code{AppInstanceUser} is a part of.
Only an \code{AppInstanceAdmin} can call the API with a user ARN that is not
their own.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_channel_memberships_for_app_instance_user(
  AppInstanceUserArn = "string",
  MaxResults = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
