% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_list_channel_messages}
\alias{chime_list_channel_messages}
\title{List all the messages in a channel}
\usage{
chime_list_channel_messages(ChannelArn, SortOrder, NotBefore, NotAfter,
  MaxResults, NextToken)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel.}

\item{SortOrder}{The order in which you want messages sorted. Default is Descending,
based on time created.}

\item{NotBefore}{The initial or starting time stamp for your requested messages.}

\item{NotAfter}{The final or ending time stamp for your requested messages.}

\item{MaxResults}{The maximum number of messages that you want returned.}

\item{NextToken}{The token passed by previous API calls until all requested messages are
returned.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ChannelArn = "string",
  NextToken = "string",
  ChannelMessages = list(
    list(
      MessageId = "string",
      Content = "string",
      Metadata = "string",
      Type = "STANDARD"|"CONTROL",
      CreatedTimestamp = as.POSIXct(
        "2015-01-01"
      ),
      LastUpdatedTimestamp = as.POSIXct(
        "2015-01-01"
      ),
      LastEditedTimestamp = as.POSIXct(
        "2015-01-01"
      ),
      Sender = list(
        Arn = "string",
        Name = "string"
      ),
      Redacted = TRUE|FALSE
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
List all the messages in a channel. Returns a paginated list of
\code{ChannelMessages}. Sorted in descending order by default, based on the
creation timestamp.

Redacted messages appear in the results as empty, since they are only
redacted, not deleted. Deleted messages do not appear in the results.
This action always returns the latest version of an edited message.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_channel_messages(
  ChannelArn = "string",
  SortOrder = "ASCENDING"|"DESCENDING",
  NotBefore = as.POSIXct(
    "2015-01-01"
  ),
  NotAfter = as.POSIXct(
    "2015-01-01"
  ),
  MaxResults = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
