% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_list_channel_moderators}
\alias{chime_list_channel_moderators}
\title{Lists all the moderators for a channel}
\usage{
chime_list_channel_moderators(ChannelArn, MaxResults, NextToken)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel.}

\item{MaxResults}{The maximum number of moderators that you want returned.}

\item{NextToken}{The token passed by previous API calls until all requested moderators
are returned.}
}
\description{
Lists all the moderators for a channel.
}
\section{Request syntax}{
\preformatted{svc$list_channel_moderators(
  ChannelArn = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
