% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_list_channels}
\alias{chime_list_channels}
\title{Lists all Channels created under a single Chime App as a paginated list}
\usage{
chime_list_channels(AppInstanceArn, Privacy, MaxResults, NextToken,
  ChimeBearer)
}
\arguments{
\item{AppInstanceArn}{[required] The ARN of the \code{AppInstance}.}

\item{Privacy}{The privacy setting. \code{PUBLIC} retrieves all the public channels.
\code{PRIVATE} retrieves private channels. Only an \code{AppInstanceAdmin} can
retrieve private channels.}

\item{MaxResults}{The maximum number of channels that you want to return.}

\item{NextToken}{The token passed by previous API calls until all requested channels are
returned.}

\item{ChimeBearer}{The \code{AppInstanceUserArn} of the user that makes the API call.}
}
\value{
A list with the following syntax:\preformatted{list(
  Channels = list(
    list(
      Name = "string",
      ChannelArn = "string",
      Mode = "UNRESTRICTED"|"RESTRICTED",
      Privacy = "PUBLIC"|"PRIVATE",
      Metadata = "string",
      LastMessageTimestamp = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  NextToken = "string"
)
}
}
\description{
Lists all Channels created under a single Chime App as a paginated list.
You can specify filters to narrow results.

\strong{Functionality & restrictions}
\itemize{
\item Use privacy = \code{PUBLIC} to retrieve all public channels in the
account.
\item Only an \code{AppInstanceAdmin} can set privacy = \code{PRIVATE} to list the
private channels in an account.
}

The \code{x-amz-chime-bearer} request header is mandatory. Use the
\code{AppInstanceUserArn} of the user that makes the API call as the value in
the header.
}
\section{Request syntax}{
\preformatted{svc$list_channels(
  AppInstanceArn = "string",
  Privacy = "PUBLIC"|"PRIVATE",
  MaxResults = 123,
  NextToken = "string",
  ChimeBearer = "string"
)
}
}

\keyword{internal}
