% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_list_meeting_tags}
\alias{chime_list_meeting_tags}
\title{Lists the tags applied to an Amazon Chime SDK meeting resource}
\usage{
chime_list_meeting_tags(MeetingId)
}
\arguments{
\item{MeetingId}{[required] The Amazon Chime SDK meeting ID.}
}
\value{
A list with the following syntax:\preformatted{list(
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}
\description{
Lists the tags applied to an Amazon Chime SDK meeting resource.
}
\section{Request syntax}{
\preformatted{svc$list_meeting_tags(
  MeetingId = "string"
)
}
}

\keyword{internal}
