% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_list_phone_numbers}
\alias{chime_list_phone_numbers}
\title{Lists the phone numbers for the specified Amazon Chime account, Amazon
Chime user, Amazon Chime Voice Connector, or Amazon Chime Voice
Connector group}
\usage{
chime_list_phone_numbers(Status, ProductType, FilterName, FilterValue,
  MaxResults, NextToken)
}
\arguments{
\item{Status}{The phone number status.}

\item{ProductType}{The phone number product type.}

\item{FilterName}{The filter to use to limit the number of results.}

\item{FilterValue}{The value to use for the filter.}

\item{MaxResults}{The maximum number of results to return in a single call.}

\item{NextToken}{The token to use to retrieve the next page of results.}
}
\value{
A list with the following syntax:\preformatted{list(
  PhoneNumbers = list(
    list(
      PhoneNumberId = "string",
      E164PhoneNumber = "string",
      Type = "Local"|"TollFree",
      ProductType = "BusinessCalling"|"VoiceConnector",
      Status = "AcquireInProgress"|"AcquireFailed"|"Unassigned"|"Assigned"|"ReleaseInProgress"|"DeleteInProgress"|"ReleaseFailed"|"DeleteFailed",
      Capabilities = list(
        InboundCall = TRUE|FALSE,
        OutboundCall = TRUE|FALSE,
        InboundSMS = TRUE|FALSE,
        OutboundSMS = TRUE|FALSE,
        InboundMMS = TRUE|FALSE,
        OutboundMMS = TRUE|FALSE
      ),
      Associations = list(
        list(
          Value = "string",
          Name = "AccountId"|"UserId"|"VoiceConnectorId"|"VoiceConnectorGroupId"|"SipRuleId",
          AssociatedTimestamp = as.POSIXct(
            "2015-01-01"
          )
        )
      ),
      CallingName = "string",
      CallingNameStatus = "Unassigned"|"UpdateInProgress"|"UpdateSucceeded"|"UpdateFailed",
      CreatedTimestamp = as.POSIXct(
        "2015-01-01"
      ),
      UpdatedTimestamp = as.POSIXct(
        "2015-01-01"
      ),
      DeletionTimestamp = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  NextToken = "string"
)
}
}
\description{
Lists the phone numbers for the specified Amazon Chime account, Amazon
Chime user, Amazon Chime Voice Connector, or Amazon Chime Voice
Connector group.
}
\section{Request syntax}{
\preformatted{svc$list_phone_numbers(
  Status = "AcquireInProgress"|"AcquireFailed"|"Unassigned"|"Assigned"|"ReleaseInProgress"|"DeleteInProgress"|"ReleaseFailed"|"DeleteFailed",
  ProductType = "BusinessCalling"|"VoiceConnector",
  FilterName = "AccountId"|"UserId"|"VoiceConnectorId"|"VoiceConnectorGroupId"|"SipRuleId",
  FilterValue = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
