% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_list_room_memberships}
\alias{chime_list_room_memberships}
\title{Lists the membership details for the specified room in an Amazon Chime
Enterprise account, such as the members' IDs, email addresses, and names}
\usage{
chime_list_room_memberships(AccountId, RoomId, MaxResults, NextToken)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{RoomId}{[required] The room ID.}

\item{MaxResults}{The maximum number of results to return in a single call.}

\item{NextToken}{The token to use to retrieve the next page of results.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  RoomMemberships = list(
    list(
      RoomId = "string",
      Member = list(
        MemberId = "string",
        MemberType = "User"|"Bot"|"Webhook",
        Email = "string",
        FullName = "string",
        AccountId = "string"
      ),
      Role = "Administrator"|"Member",
      InvitedBy = "string",
      UpdatedTimestamp = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists the membership details for the specified room in an Amazon Chime
Enterprise account, such as the members' IDs, email addresses, and
names.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_room_memberships(
  AccountId = "string",
  RoomId = "string",
  MaxResults = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
