% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_list_rooms}
\alias{chime_list_rooms}
\title{Lists the room details for the specified Amazon Chime Enterprise account}
\usage{
chime_list_rooms(AccountId, MemberId, MaxResults, NextToken)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{MemberId}{The member ID (user ID or bot ID).}

\item{MaxResults}{The maximum number of results to return in a single call.}

\item{NextToken}{The token to use to retrieve the next page of results.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Rooms = list(
    list(
      RoomId = "string",
      Name = "string",
      AccountId = "string",
      CreatedBy = "string",
      CreatedTimestamp = as.POSIXct(
        "2015-01-01"
      ),
      UpdatedTimestamp = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists the room details for the specified Amazon Chime Enterprise
account. Optionally, filter the results by a member ID (user ID or bot
ID) to see a list of rooms that the member belongs to.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_rooms(
  AccountId = "string",
  MemberId = "string",
  MaxResults = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
