% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_list_sip_media_applications}
\alias{chime_list_sip_media_applications}
\title{Lists the SIP media applications under the administrator's AWS account}
\usage{
chime_list_sip_media_applications(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The maximum number of results to return in a single call. Defaults to
100.}

\item{NextToken}{The token to use to retrieve the next page of results.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  SipMediaApplications = list(
    list(
      SipMediaApplicationId = "string",
      AwsRegion = "string",
      Name = "string",
      Endpoints = list(
        list(
          LambdaArn = "string"
        )
      ),
      CreatedTimestamp = as.POSIXct(
        "2015-01-01"
      ),
      UpdatedTimestamp = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists the SIP media applications under the administrator's AWS account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_sip_media_applications(
  MaxResults = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
