% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_list_sip_rules}
\alias{chime_list_sip_rules}
\title{Lists the SIP rules under the administrator's AWS account}
\usage{
chime_list_sip_rules(SipMediaApplicationId, MaxResults, NextToken)
}
\arguments{
\item{SipMediaApplicationId}{The SIP media application ID.}

\item{MaxResults}{The maximum number of results to return in a single call. Defaults to
100.}

\item{NextToken}{The token to use to retrieve the next page of results.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  SipRules = list(
    list(
      SipRuleId = "string",
      Name = "string",
      Disabled = TRUE|FALSE,
      TriggerType = "ToPhoneNumber"|"RequestUriHostname",
      TriggerValue = "string",
      TargetApplications = list(
        list(
          SipMediaApplicationId = "string",
          Priority = 123,
          AwsRegion = "string"
        )
      ),
      CreatedTimestamp = as.POSIXct(
        "2015-01-01"
      ),
      UpdatedTimestamp = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists the SIP rules under the administrator's AWS account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_sip_rules(
  SipMediaApplicationId = "string",
  MaxResults = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
