% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_list_supported_phone_number_countries}
\alias{chime_list_supported_phone_number_countries}
\title{Lists supported phone number countries}
\usage{
chime_list_supported_phone_number_countries(ProductType)
}
\arguments{
\item{ProductType}{[required] The phone number product type.}
}
\value{
A list with the following syntax:\preformatted{list(
  PhoneNumberCountries = list(
    list(
      CountryCode = "string",
      SupportedPhoneNumberTypes = list(
        "Local"|"TollFree"
      )
    )
  )
)
}
}
\description{
Lists supported phone number countries.
}
\section{Request syntax}{
\preformatted{svc$list_supported_phone_number_countries(
  ProductType = "BusinessCalling"|"VoiceConnector"|"SipMediaApplicationDialIn"
)
}
}

\keyword{internal}
