% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_list_tags_for_resource}
\alias{chime_list_tags_for_resource}
\title{Lists the tags applied to an Amazon Chime SDK meeting resource}
\usage{
chime_list_tags_for_resource(ResourceARN)
}
\arguments{
\item{ResourceARN}{[required] The resource ARN.}
}
\value{
A list with the following syntax:\preformatted{list(
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}
\description{
Lists the tags applied to an Amazon Chime SDK meeting resource.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  ResourceARN = "string"
)
}
}

\keyword{internal}
