% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_list_voice_connector_termination_credentials}
\alias{chime_list_voice_connector_termination_credentials}
\title{Lists the SIP credentials for the specified Amazon Chime Voice Connector}
\usage{
chime_list_voice_connector_termination_credentials(VoiceConnectorId)
}
\arguments{
\item{VoiceConnectorId}{[required] The Amazon Chime Voice Connector ID.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Usernames = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Lists the SIP credentials for the specified Amazon Chime Voice
Connector.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_voice_connector_termination_credentials(
  VoiceConnectorId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
