% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_put_app_instance_retention_settings}
\alias{chime_put_app_instance_retention_settings}
\title{Sets the amount of time in days that a given app instance retains data}
\usage{
chime_put_app_instance_retention_settings(AppInstanceArn,
  AppInstanceRetentionSettings)
}
\arguments{
\item{AppInstanceArn}{[required] The ARN of the app instance.}

\item{AppInstanceRetentionSettings}{[required] The time in days to retain data. Data type: number.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AppInstanceRetentionSettings = list(
    ChannelRetentionSettings = list(
      RetentionDays = 123
    )
  ),
  InitiateDeletionTimestamp = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Sets the amount of time in days that a given app instance retains data.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$put_app_instance_retention_settings(
  AppInstanceArn = "string",
  AppInstanceRetentionSettings = list(
    ChannelRetentionSettings = list(
      RetentionDays = 123
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
