% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_put_app_instance_streaming_configurations}
\alias{chime_put_app_instance_streaming_configurations}
\title{The data streaming configurations of an AppInstance}
\usage{
chime_put_app_instance_streaming_configurations(AppInstanceArn,
  AppInstanceStreamingConfigurations)
}
\arguments{
\item{AppInstanceArn}{[required] The ARN of the \code{AppInstance}.}

\item{AppInstanceStreamingConfigurations}{[required] The streaming configurations set for an \code{AppInstance}.}
}
\value{
A list with the following syntax:\preformatted{list(
  AppInstanceStreamingConfigurations = list(
    list(
      AppInstanceDataType = "Channel"|"ChannelMessage",
      ResourceArn = "string"
    )
  )
)
}
}
\description{
The data streaming configurations of an \code{AppInstance}.
}
\section{Request syntax}{
\preformatted{svc$put_app_instance_streaming_configurations(
  AppInstanceArn = "string",
  AppInstanceStreamingConfigurations = list(
    list(
      AppInstanceDataType = "Channel"|"ChannelMessage",
      ResourceArn = "string"
    )
  )
)
}
}

\keyword{internal}
