% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_put_retention_settings}
\alias{chime_put_retention_settings}
\title{Puts retention settings for the specified Amazon Chime Enterprise
account}
\usage{
chime_put_retention_settings(AccountId, RetentionSettings)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{RetentionSettings}{[required] The retention settings.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  RetentionSettings = list(
    RoomRetentionSettings = list(
      RetentionDays = 123
    ),
    ConversationRetentionSettings = list(
      RetentionDays = 123
    )
  ),
  InitiateDeletionTimestamp = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Puts retention settings for the specified Amazon Chime Enterprise
account. We recommend using AWS CloudTrail to monitor usage of this API
for your account. For more information, see \href{https://docs.aws.amazon.com/chime/latest/ag/cloudtrail.html}{Logging Amazon Chime API Calls with AWS CloudTrail}
in the \emph{Amazon Chime Administration Guide}.

To turn off existing retention settings, remove the number of days from
the corresponding \strong{RetentionDays} field in the \strong{RetentionSettings}
object. For more information about retention settings, see \href{https://docs.aws.amazon.com/chime/latest/ag/archive-retention.html}{Managing Chat Retention Policies}
in the \emph{Amazon Chime Administration Guide}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$put_retention_settings(
  AccountId = "string",
  RetentionSettings = list(
    RoomRetentionSettings = list(
      RetentionDays = 123
    ),
    ConversationRetentionSettings = list(
      RetentionDays = 123
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
