% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_put_voice_connector_emergency_calling_configuration}
\alias{chime_put_voice_connector_emergency_calling_configuration}
\title{Puts emergency calling configuration details to the specified Amazon
Chime Voice Connector, such as emergency phone numbers and calling
countries}
\usage{
chime_put_voice_connector_emergency_calling_configuration(
  VoiceConnectorId, EmergencyCallingConfiguration)
}
\arguments{
\item{VoiceConnectorId}{[required] The Amazon Chime Voice Connector ID.}

\item{EmergencyCallingConfiguration}{[required] The emergency calling configuration details.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  EmergencyCallingConfiguration = list(
    DNIS = list(
      list(
        EmergencyPhoneNumber = "string",
        TestPhoneNumber = "string",
        CallingCountry = "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Puts emergency calling configuration details to the specified Amazon
Chime Voice Connector, such as emergency phone numbers and calling
countries. Origination and termination settings must be enabled for the
Amazon Chime Voice Connector before emergency calling can be configured.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$put_voice_connector_emergency_calling_configuration(
  VoiceConnectorId = "string",
  EmergencyCallingConfiguration = list(
    DNIS = list(
      list(
        EmergencyPhoneNumber = "string",
        TestPhoneNumber = "string",
        CallingCountry = "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
