% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_put_voice_connector_termination_credentials}
\alias{chime_put_voice_connector_termination_credentials}
\title{Adds termination SIP credentials for the specified Amazon Chime Voice
Connector}
\usage{
chime_put_voice_connector_termination_credentials(VoiceConnectorId,
  Credentials)
}
\arguments{
\item{VoiceConnectorId}{[required] The Amazon Chime Voice Connector ID.}

\item{Credentials}{The termination SIP credentials.}
}
\value{
An empty list.
}
\description{
Adds termination SIP credentials for the specified Amazon Chime Voice
Connector.
}
\section{Request syntax}{
\preformatted{svc$put_voice_connector_termination_credentials(
  VoiceConnectorId = "string",
  Credentials = list(
    list(
      Username = "string",
      Password = "string"
    )
  )
)
}
}

\keyword{internal}
