% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_redact_conversation_message}
\alias{chime_redact_conversation_message}
\title{Redacts the specified message from the specified Amazon Chime
conversation}
\usage{
chime_redact_conversation_message(AccountId, ConversationId, MessageId)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{ConversationId}{[required] The conversation ID.}

\item{MessageId}{[required] The message ID.}
}
\value{
An empty list.
}
\description{
Redacts the specified message from the specified Amazon Chime
conversation.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$redact_conversation_message(
  AccountId = "string",
  ConversationId = "string",
  MessageId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
