% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_regenerate_security_token}
\alias{chime_regenerate_security_token}
\title{Regenerates the security token for a bot}
\usage{
chime_regenerate_security_token(AccountId, BotId)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{BotId}{[required] The bot ID.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Bot = list(
    BotId = "string",
    UserId = "string",
    DisplayName = "string",
    BotType = "ChatBot",
    Disabled = TRUE|FALSE,
    CreatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    UpdatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    BotEmail = "string",
    SecurityToken = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Regenerates the security token for a bot.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$regenerate_security_token(
  AccountId = "string",
  BotId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
