% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_send_channel_message}
\alias{chime_send_channel_message}
\title{Sends a message to a particular channel that the member is a part of}
\usage{
chime_send_channel_message(ChannelArn, Content, Type, Persistence,
  Metadata, ClientRequestToken, ChimeBearer)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel.}

\item{Content}{[required] The content of the message.}

\item{Type}{[required] The type of message, \code{STANDARD} or \code{CONTROL}.}

\item{Persistence}{[required] Boolean that controls whether the message is persisted on the back end.
Required.}

\item{Metadata}{The optional metadata for each message.}

\item{ClientRequestToken}{[required] The \code{Idempotency} token for each client request.}

\item{ChimeBearer}{The \code{AppInstanceUserArn} of the user that makes the API call.}
}
\value{
A list with the following syntax:\preformatted{list(
  ChannelArn = "string",
  MessageId = "string"
)
}
}
\description{
Sends a message to a particular channel that the member is a part of.

The \code{x-amz-chime-bearer} request header is mandatory. Use the
\code{AppInstanceUserArn} of the user that makes the API call as the value in
the header.

Also, \code{STANDARD} messages can contain 4KB of data and the 1KB of
metadata. \code{CONTROL} messages can contain 30 bytes of data and no
metadata.
}
\section{Request syntax}{
\preformatted{svc$send_channel_message(
  ChannelArn = "string",
  Content = "string",
  Type = "STANDARD"|"CONTROL",
  Persistence = "PERSISTENT"|"NON_PERSISTENT",
  Metadata = "string",
  ClientRequestToken = "string",
  ChimeBearer = "string"
)
}
}

\keyword{internal}
