% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_tag_meeting}
\alias{chime_tag_meeting}
\title{Applies the specified tags to the specified Amazon Chime SDK meeting}
\usage{
chime_tag_meeting(MeetingId, Tags)
}
\arguments{
\item{MeetingId}{[required] The Amazon Chime SDK meeting ID.}

\item{Tags}{[required] The tag key-value pairs.}
}
\value{
An empty list.
}
\description{
Applies the specified tags to the specified Amazon Chime SDK meeting.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$tag_meeting(
  MeetingId = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
