% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_untag_attendee}
\alias{chime_untag_attendee}
\title{Untags the specified tags from the specified Amazon Chime SDK attendee}
\usage{
chime_untag_attendee(MeetingId, AttendeeId, TagKeys)
}
\arguments{
\item{MeetingId}{[required] The Amazon Chime SDK meeting ID.}

\item{AttendeeId}{[required] The Amazon Chime SDK attendee ID.}

\item{TagKeys}{[required] The tag keys.}
}
\description{
Untags the specified tags from the specified Amazon Chime SDK attendee.
}
\section{Request syntax}{
\preformatted{svc$untag_attendee(
  MeetingId = "string",
  AttendeeId = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
