% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_untag_resource}
\alias{chime_untag_resource}
\title{Untags the specified tags from the specified Amazon Chime SDK meeting
resource}
\usage{
chime_untag_resource(ResourceARN, TagKeys)
}
\arguments{
\item{ResourceARN}{[required] The resource ARN.}

\item{TagKeys}{[required] The tag keys.}
}
\value{
An empty list.
}
\description{
Untags the specified tags from the specified Amazon Chime SDK meeting
resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceARN = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
