% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_update_account}
\alias{chime_update_account}
\title{Updates account details for the specified Amazon Chime account}
\usage{
chime_update_account(AccountId, Name)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{Name}{The new name for the specified Amazon Chime account.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Account = list(
    AwsAccountId = "string",
    AccountId = "string",
    Name = "string",
    AccountType = "Team"|"EnterpriseDirectory"|"EnterpriseLWA"|"EnterpriseOIDC",
    CreatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    DefaultLicense = "Basic"|"Plus"|"Pro"|"ProTrial",
    SupportedLicenses = list(
      "Basic"|"Plus"|"Pro"|"ProTrial"
    ),
    SigninDelegateGroups = list(
      list(
        GroupName = "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Updates account details for the specified Amazon Chime account.
Currently, only account name updates are supported for this action.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_account(
  AccountId = "string",
  Name = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
