% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_update_app_instance}
\alias{chime_update_app_instance}
\title{Updates AppInstance metadata}
\usage{
chime_update_app_instance(AppInstanceArn, Name, Metadata)
}
\arguments{
\item{AppInstanceArn}{[required] The ARN of the \code{AppInstance}.}

\item{Name}{[required] The name that you want to change.}

\item{Metadata}{The metadata that you want to change.}
}
\value{
A list with the following syntax:\preformatted{list(
  AppInstanceArn = "string"
)
}
}
\description{
Updates \code{AppInstance} metadata.
}
\section{Request syntax}{
\preformatted{svc$update_app_instance(
  AppInstanceArn = "string",
  Name = "string",
  Metadata = "string"
)
}
}

\keyword{internal}
