% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_update_bot}
\alias{chime_update_bot}
\title{Updates the status of the specified bot, such as starting or stopping
the bot from running in your Amazon Chime Enterprise account}
\usage{
chime_update_bot(AccountId, BotId, Disabled)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{BotId}{[required] The bot ID.}

\item{Disabled}{When true, stops the specified bot from running in your account.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Bot = list(
    BotId = "string",
    UserId = "string",
    DisplayName = "string",
    BotType = "ChatBot",
    Disabled = TRUE|FALSE,
    CreatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    UpdatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    BotEmail = "string",
    SecurityToken = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Updates the status of the specified bot, such as starting or stopping
the bot from running in your Amazon Chime Enterprise account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_bot(
  AccountId = "string",
  BotId = "string",
  Disabled = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
