% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_update_channel}
\alias{chime_update_channel}
\title{Update a channel's attributes}
\usage{
chime_update_channel(ChannelArn, Name, Mode, Metadata, ChimeBearer)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel.}

\item{Name}{[required] The name of the channel.}

\item{Mode}{[required] The mode of the update request.}

\item{Metadata}{The metadata for the update request.}

\item{ChimeBearer}{The \code{AppInstanceUserArn} of the user that makes the API call.}
}
\value{
A list with the following syntax:\preformatted{list(
  ChannelArn = "string"
)
}
}
\description{
Update a channel's attributes.

\strong{Restriction}: You can't change a channel's privacy.

The \code{x-amz-chime-bearer} request header is mandatory. Use the
\code{AppInstanceUserArn} of the user that makes the API call as the value in
the header.
}
\section{Request syntax}{
\preformatted{svc$update_channel(
  ChannelArn = "string",
  Name = "string",
  Mode = "UNRESTRICTED"|"RESTRICTED",
  Metadata = "string",
  ChimeBearer = "string"
)
}
}

\keyword{internal}
