% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_update_channel_message}
\alias{chime_update_channel_message}
\title{Updates the content of a message}
\usage{
chime_update_channel_message(ChannelArn, MessageId, Content, Metadata)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel.}

\item{MessageId}{[required] The ID string of the message being updated.}

\item{Content}{The content of the message being updated.}

\item{Metadata}{The metadata of the message being updated.}
}
\value{
A list with the following syntax:\preformatted{list(
  ChannelArn = "string",
  MessageId = "string"
)
}
}
\description{
Updates the content of a message.
}
\section{Request syntax}{
\preformatted{svc$update_channel_message(
  ChannelArn = "string",
  MessageId = "string",
  Content = "string",
  Metadata = "string"
)
}
}

\keyword{internal}
