% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_update_channel_read_marker}
\alias{chime_update_channel_read_marker}
\title{The details of the time when a user last read messages in a channel}
\usage{
chime_update_channel_read_marker(ChannelArn, ChimeBearer)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel.}

\item{ChimeBearer}{The \code{AppInstanceUserArn} of the user that makes the API call.}
}
\value{
A list with the following syntax:\preformatted{list(
  ChannelArn = "string"
)
}
}
\description{
The details of the time when a user last read messages in a channel.

The \code{x-amz-chime-bearer} request header is mandatory. Use the
\code{AppInstanceUserArn} of the user that makes the API call as the value in
the header.
}
\section{Request syntax}{
\preformatted{svc$update_channel_read_marker(
  ChannelArn = "string",
  ChimeBearer = "string"
)
}
}

\keyword{internal}
