% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_update_phone_number}
\alias{chime_update_phone_number}
\title{Updates phone number details, such as product type or calling name, for
the specified phone number ID}
\usage{
chime_update_phone_number(PhoneNumberId, ProductType, CallingName)
}
\arguments{
\item{PhoneNumberId}{[required] The phone number ID.}

\item{ProductType}{The product type.}

\item{CallingName}{The outbound calling name associated with the phone number.}
}
\value{
A list with the following syntax:\preformatted{list(
  PhoneNumber = list(
    PhoneNumberId = "string",
    E164PhoneNumber = "string",
    Country = "string",
    Type = "Local"|"TollFree",
    ProductType = "BusinessCalling"|"VoiceConnector"|"SipMediaApplicationDialIn",
    Status = "AcquireInProgress"|"AcquireFailed"|"Unassigned"|"Assigned"|"ReleaseInProgress"|"DeleteInProgress"|"ReleaseFailed"|"DeleteFailed",
    Capabilities = list(
      InboundCall = TRUE|FALSE,
      OutboundCall = TRUE|FALSE,
      InboundSMS = TRUE|FALSE,
      OutboundSMS = TRUE|FALSE,
      InboundMMS = TRUE|FALSE,
      OutboundMMS = TRUE|FALSE
    ),
    Associations = list(
      list(
        Value = "string",
        Name = "AccountId"|"UserId"|"VoiceConnectorId"|"VoiceConnectorGroupId"|"SipRuleId",
        AssociatedTimestamp = as.POSIXct(
          "2015-01-01"
        )
      )
    ),
    CallingName = "string",
    CallingNameStatus = "Unassigned"|"UpdateInProgress"|"UpdateSucceeded"|"UpdateFailed",
    CreatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    UpdatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    DeletionTimestamp = as.POSIXct(
      "2015-01-01"
    )
  )
)
}
}
\description{
Updates phone number details, such as product type or calling name, for
the specified phone number ID. You can update one phone number detail at
a time. For example, you can update either the product type or the
calling name in one action.

For toll-free numbers, you cannot use the Amazon Chime Business Calling
product type. For numbers outside the U.S., you must use the Amazon
Chime SIP Media Application Dial-In product type.

Updates to outbound calling names can take 72 hours to complete. Pending
updates to outbound calling names must be complete before you can
request another update.
}
\section{Request syntax}{
\preformatted{svc$update_phone_number(
  PhoneNumberId = "string",
  ProductType = "BusinessCalling"|"VoiceConnector"|"SipMediaApplicationDialIn",
  CallingName = "string"
)
}
}

\keyword{internal}
