% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_update_phone_number_settings}
\alias{chime_update_phone_number_settings}
\title{Updates the phone number settings for the administrator's AWS account,
such as the default outbound calling name}
\usage{
chime_update_phone_number_settings(CallingName)
}
\arguments{
\item{CallingName}{[required] The default outbound calling name for the account.}
}
\description{
Updates the phone number settings for the administrator\'s AWS account,
such as the default outbound calling name. You can update the default
outbound calling name once every seven days. Outbound calling names can
take up to 72 hours to update.
}
\section{Request syntax}{
\preformatted{svc$update_phone_number_settings(
  CallingName = "string"
)
}
}

\keyword{internal}
