% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_update_proxy_session}
\alias{chime_update_proxy_session}
\title{Updates the specified proxy session details, such as voice or SMS
capabilities}
\usage{
chime_update_proxy_session(VoiceConnectorId, ProxySessionId,
  Capabilities, ExpiryMinutes)
}
\arguments{
\item{VoiceConnectorId}{[required] The Amazon Chime voice connector ID.}

\item{ProxySessionId}{[required] The proxy session ID.}

\item{Capabilities}{[required] The proxy session capabilities.}

\item{ExpiryMinutes}{The number of minutes allowed for the proxy session.}
}
\description{
Updates the specified proxy session details, such as voice or SMS
capabilities.
}
\section{Request syntax}{
\preformatted{svc$update_proxy_session(
  VoiceConnectorId = "string",
  ProxySessionId = "string",
  Capabilities = list(
    "Voice"|"SMS"
  ),
  ExpiryMinutes = 123
)
}
}

\keyword{internal}
