% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_update_room_membership}
\alias{chime_update_room_membership}
\title{Updates room membership details, such as member role}
\usage{
chime_update_room_membership(AccountId, RoomId, MemberId, Role)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{RoomId}{[required] The room ID.}

\item{MemberId}{[required] The member ID.}

\item{Role}{The role of the member.}
}
\description{
Updates room membership details, such as member role. The member role
designates whether the member is a chat room administrator or a general
chat room member. Member role can only be updated for user IDs.
}
\section{Request syntax}{
\preformatted{svc$update_room_membership(
  AccountId = "string",
  RoomId = "string",
  MemberId = "string",
  Role = "Administrator"|"Member"
)
}
}

\keyword{internal}
