% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_update_sip_media_application}
\alias{chime_update_sip_media_application}
\title{Updates the details of the specified SIP media application}
\usage{
chime_update_sip_media_application(SipMediaApplicationId, Name,
  Endpoints)
}
\arguments{
\item{SipMediaApplicationId}{[required] The SIP media application ID.}

\item{Name}{The new name for the specified SIP media application.}

\item{Endpoints}{The new set of endpoints for the specified SIP media application.}
}
\value{
A list with the following syntax:\preformatted{list(
  SipMediaApplication = list(
    SipMediaApplicationId = "string",
    AwsRegion = "string",
    Name = "string",
    Endpoints = list(
      list(
        LambdaArn = "string"
      )
    ),
    CreatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    UpdatedTimestamp = as.POSIXct(
      "2015-01-01"
    )
  )
)
}
}
\description{
Updates the details of the specified SIP media application.
}
\section{Request syntax}{
\preformatted{svc$update_sip_media_application(
  SipMediaApplicationId = "string",
  Name = "string",
  Endpoints = list(
    list(
      LambdaArn = "string"
    )
  )
)
}
}

\keyword{internal}
