% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_update_sip_rule}
\alias{chime_update_sip_rule}
\title{Updates the details for the specified SIP rule}
\usage{
chime_update_sip_rule(SipRuleId, Name, Disabled, TargetApplications)
}
\arguments{
\item{SipRuleId}{[required] The SIP rule ID.}

\item{Name}{[required] The new name for the specified SIP rule.}

\item{Disabled}{The new value specified to indicate whether the rule is disabled.}

\item{TargetApplications}{The new value of the list of target applications.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  SipRule = list(
    SipRuleId = "string",
    Name = "string",
    Disabled = TRUE|FALSE,
    TriggerType = "ToPhoneNumber"|"RequestUriHostname",
    TriggerValue = "string",
    TargetApplications = list(
      list(
        SipMediaApplicationId = "string",
        Priority = 123,
        AwsRegion = "string"
      )
    ),
    CreatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    UpdatedTimestamp = as.POSIXct(
      "2015-01-01"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Updates the details for the specified SIP rule.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_sip_rule(
  SipRuleId = "string",
  Name = "string",
  Disabled = TRUE|FALSE,
  TargetApplications = list(
    list(
      SipMediaApplicationId = "string",
      Priority = 123,
      AwsRegion = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
