% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_update_user}
\alias{chime_update_user}
\title{Updates user details for a specified user ID}
\usage{
chime_update_user(AccountId, UserId, LicenseType, UserType,
  AlexaForBusinessMetadata)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{UserId}{[required] The user ID.}

\item{LicenseType}{The user license type to update. This must be a supported license type
for the Amazon Chime account that the user belongs to.}

\item{UserType}{The user type.}

\item{AlexaForBusinessMetadata}{The Alexa for Business metadata.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  User = list(
    UserId = "string",
    AccountId = "string",
    PrimaryEmail = "string",
    PrimaryProvisionedNumber = "string",
    DisplayName = "string",
    LicenseType = "Basic"|"Plus"|"Pro"|"ProTrial",
    UserType = "PrivateUser"|"SharedDevice",
    UserRegistrationStatus = "Unregistered"|"Registered"|"Suspended",
    UserInvitationStatus = "Pending"|"Accepted"|"Failed",
    RegisteredOn = as.POSIXct(
      "2015-01-01"
    ),
    InvitedOn = as.POSIXct(
      "2015-01-01"
    ),
    AlexaForBusinessMetadata = list(
      IsAlexaForBusinessEnabled = TRUE|FALSE,
      AlexaForBusinessRoomArn = "string"
    ),
    PersonalPIN = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Updates user details for a specified user ID. Currently, only
\code{LicenseType} updates are supported for this action.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_user(
  AccountId = "string",
  UserId = "string",
  LicenseType = "Basic"|"Plus"|"Pro"|"ProTrial",
  UserType = "PrivateUser"|"SharedDevice",
  AlexaForBusinessMetadata = list(
    IsAlexaForBusinessEnabled = TRUE|FALSE,
    AlexaForBusinessRoomArn = "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
