% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_update_user_settings}
\alias{chime_update_user_settings}
\title{Updates the settings for the specified user, such as phone number
settings}
\usage{
chime_update_user_settings(AccountId, UserId, UserSettings)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{UserId}{[required] The user ID.}

\item{UserSettings}{[required] The user settings to update.}
}
\value{
An empty list.
}
\description{
Updates the settings for the specified user, such as phone number
settings.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_user_settings(
  AccountId = "string",
  UserId = "string",
  UserSettings = list(
    Telephony = list(
      InboundCalling = TRUE|FALSE,
      OutboundCalling = TRUE|FALSE,
      SMS = TRUE|FALSE
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
