% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_update_voice_connector}
\alias{chime_update_voice_connector}
\title{Updates details for the specified Amazon Chime Voice Connector}
\usage{
chime_update_voice_connector(VoiceConnectorId, Name, RequireEncryption)
}
\arguments{
\item{VoiceConnectorId}{[required] The Amazon Chime Voice Connector ID.}

\item{Name}{[required] The name of the Amazon Chime Voice Connector.}

\item{RequireEncryption}{[required] When enabled, requires encryption for the Amazon Chime Voice Connector.}
}
\description{
Updates details for the specified Amazon Chime Voice Connector.
}
\section{Request syntax}{
\preformatted{svc$update_voice_connector(
  VoiceConnectorId = "string",
  Name = "string",
  RequireEncryption = TRUE|FALSE
)
}
}

\keyword{internal}
