% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_associate_member_to_group}
\alias{workmail_associate_member_to_group}
\title{Adds a member (user or group) to the group's set}
\usage{
workmail_associate_member_to_group(OrganizationId, GroupId, MemberId)
}
\arguments{
\item{OrganizationId}{[required] The organization under which the group exists.}

\item{GroupId}{[required] The group to which the member (user or group) is associated.}

\item{MemberId}{[required] The member (user or group) to associate to the group.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Adds a member (user or group) to the group's set.
}
\section{Request syntax}{
\preformatted{svc$associate_member_to_group(
  OrganizationId = "string",
  GroupId = "string",
  MemberId = "string"
)
}
}

\keyword{internal}
