% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_cancel_mailbox_export_job}
\alias{workmail_cancel_mailbox_export_job}
\title{Cancels a mailbox export job}
\usage{
workmail_cancel_mailbox_export_job(ClientToken, JobId, OrganizationId)
}
\arguments{
\item{ClientToken}{[required] The idempotency token for the client request.}

\item{JobId}{[required] The job ID.}

\item{OrganizationId}{[required] The organization ID.}
}
\value{
An empty list.
}
\description{
Cancels a mailbox export job.

If the mailbox export job is near completion, it might not be possible
to cancel it.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$cancel_mailbox_export_job(
  ClientToken = "string",
  JobId = "string",
  OrganizationId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
