% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_create_resource}
\alias{workmail_create_resource}
\title{Creates a new Amazon WorkMail resource}
\usage{
workmail_create_resource(OrganizationId, Name, Type)
}
\arguments{
\item{OrganizationId}{[required] The identifier associated with the organization for which the resource
is created.}

\item{Name}{[required] The name of the new resource.}

\item{Type}{[required] The type of the new resource. The available types are \code{equipment} and
\code{room}.}
}
\value{
A list with the following syntax:\preformatted{list(
  ResourceId = "string"
)
}
}
\description{
Creates a new Amazon WorkMail resource.
}
\section{Request syntax}{
\preformatted{svc$create_resource(
  OrganizationId = "string",
  Name = "string",
  Type = "ROOM"|"EQUIPMENT"
)
}
}

\keyword{internal}
