% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_delete_group}
\alias{workmail_delete_group}
\title{Deletes a group from Amazon WorkMail}
\usage{
workmail_delete_group(OrganizationId, GroupId)
}
\arguments{
\item{OrganizationId}{[required] The organization that contains the group.}

\item{GroupId}{[required] The identifier of the group to be deleted.}
}
\value{
An empty list.
}
\description{
Deletes a group from Amazon WorkMail.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_group(
  OrganizationId = "string",
  GroupId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
