% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_delete_resource}
\alias{workmail_delete_resource}
\title{Deletes the specified resource}
\usage{
workmail_delete_resource(OrganizationId, ResourceId)
}
\arguments{
\item{OrganizationId}{[required] The identifier associated with the organization from which the resource
is deleted.}

\item{ResourceId}{[required] The identifier of the resource to be deleted.}
}
\value{
An empty list.
}
\description{
Deletes the specified resource.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_resource(
  OrganizationId = "string",
  ResourceId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
