% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_describe_mailbox_export_job}
\alias{workmail_describe_mailbox_export_job}
\title{Describes the current status of a mailbox export job}
\usage{
workmail_describe_mailbox_export_job(JobId, OrganizationId)
}
\arguments{
\item{JobId}{[required] The mailbox export job ID.}

\item{OrganizationId}{[required] The organization ID.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  EntityId = "string",
  Description = "string",
  RoleArn = "string",
  KmsKeyArn = "string",
  S3BucketName = "string",
  S3Prefix = "string",
  S3Path = "string",
  EstimatedProgress = 123,
  State = "RUNNING"|"COMPLETED"|"FAILED"|"CANCELLED",
  ErrorInfo = "string",
  StartTime = as.POSIXct(
    "2015-01-01"
  ),
  EndTime = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes the current status of a mailbox export job.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_mailbox_export_job(
  JobId = "string",
  OrganizationId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
