% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_describe_organization}
\alias{workmail_describe_organization}
\title{Provides more information regarding a given organization based on its
identifier}
\usage{
workmail_describe_organization(OrganizationId)
}
\arguments{
\item{OrganizationId}{[required] The identifier for the organization to be described.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  OrganizationId = "string",
  Alias = "string",
  State = "string",
  DirectoryId = "string",
  DirectoryType = "string",
  DefaultMailDomain = "string",
  CompletedDate = as.POSIXct(
    "2015-01-01"
  ),
  ErrorMessage = "string",
  ARN = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Provides more information regarding a given organization based on its
identifier.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_organization(
  OrganizationId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
