% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_describe_user}
\alias{workmail_describe_user}
\title{Provides information regarding the user}
\usage{
workmail_describe_user(OrganizationId, UserId)
}
\arguments{
\item{OrganizationId}{[required] The identifier for the organization under which the user exists.}

\item{UserId}{[required] The identifier for the user to be described.}
}
\value{
A list with the following syntax:\preformatted{list(
  UserId = "string",
  Name = "string",
  Email = "string",
  DisplayName = "string",
  State = "ENABLED"|"DISABLED"|"DELETED",
  UserRole = "USER"|"RESOURCE"|"SYSTEM_USER",
  EnabledDate = as.POSIXct(
    "2015-01-01"
  ),
  DisabledDate = as.POSIXct(
    "2015-01-01"
  )
)
}
}
\description{
Provides information regarding the user.
}
\section{Request syntax}{
\preformatted{svc$describe_user(
  OrganizationId = "string",
  UserId = "string"
)
}
}

\keyword{internal}
