% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_get_access_control_effect}
\alias{workmail_get_access_control_effect}
\title{Gets the effects of an organization's access control rules as they apply
to a specified IPv4 address, access protocol action, or user ID}
\usage{
workmail_get_access_control_effect(OrganizationId, IpAddress, Action,
  UserId)
}
\arguments{
\item{OrganizationId}{[required] The identifier for the organization.}

\item{IpAddress}{[required] The IPv4 address.}

\item{Action}{[required] The access protocol action. Valid values include \code{ActiveSync},
\code{AutoDiscover}, \code{EWS}, \code{IMAP}, \code{SMTP}, \code{WindowsOutlook}, and \code{WebMail}.}

\item{UserId}{[required] The user ID.}
}
\description{
Gets the effects of an organization\'s access control rules as they
apply to a specified IPv4 address, access protocol action, or user ID.
}
\section{Request syntax}{
\preformatted{svc$get_access_control_effect(
  OrganizationId = "string",
  IpAddress = "string",
  Action = "string",
  UserId = "string"
)
}
}

\keyword{internal}
