% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_get_default_retention_policy}
\alias{workmail_get_default_retention_policy}
\title{Gets the default retention policy details for the specified organization}
\usage{
workmail_get_default_retention_policy(OrganizationId)
}
\arguments{
\item{OrganizationId}{[required] The organization ID.}
}
\description{
Gets the default retention policy details for the specified
organization.
}
\section{Request syntax}{
\preformatted{svc$get_default_retention_policy(
  OrganizationId = "string"
)
}
}

\keyword{internal}
