% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_list_mobile_device_access_rules}
\alias{workmail_list_mobile_device_access_rules}
\title{Lists the mobile device access rules for the specified Amazon WorkMail
organization}
\usage{
workmail_list_mobile_device_access_rules(OrganizationId)
}
\arguments{
\item{OrganizationId}{[required] The Amazon WorkMail organization for which to list the rules.}
}
\value{
A list with the following syntax:\preformatted{list(
  Rules = list(
    list(
      MobileDeviceAccessRuleId = "string",
      Name = "string",
      Description = "string",
      Effect = "ALLOW"|"DENY",
      DeviceTypes = list(
        "string"
      ),
      NotDeviceTypes = list(
        "string"
      ),
      DeviceModels = list(
        "string"
      ),
      NotDeviceModels = list(
        "string"
      ),
      DeviceOperatingSystems = list(
        "string"
      ),
      NotDeviceOperatingSystems = list(
        "string"
      ),
      DeviceUserAgents = list(
        "string"
      ),
      NotDeviceUserAgents = list(
        "string"
      ),
      DateCreated = as.POSIXct(
        "2015-01-01"
      ),
      DateModified = as.POSIXct(
        "2015-01-01"
      )
    )
  )
)
}
}
\description{
Lists the mobile device access rules for the specified Amazon WorkMail
organization.
}
\section{Request syntax}{
\preformatted{svc$list_mobile_device_access_rules(
  OrganizationId = "string"
)
}
}

\keyword{internal}
