% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_list_organizations}
\alias{workmail_list_organizations}
\title{Returns summaries of the customer's organizations}
\usage{
workmail_list_organizations(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{The token to use to retrieve the next page of results. The first call
does not contain any tokens.}

\item{MaxResults}{The maximum number of results to return in a single call.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  OrganizationSummaries = list(
    list(
      OrganizationId = "string",
      Alias = "string",
      DefaultMailDomain = "string",
      ErrorMessage = "string",
      State = "string"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns summaries of the customer's organizations.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_organizations(
  NextToken = "string",
  MaxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
