% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_list_users}
\alias{workmail_list_users}
\title{Returns summaries of the organization's users}
\usage{
workmail_list_users(OrganizationId, NextToken, MaxResults)
}
\arguments{
\item{OrganizationId}{[required] The identifier for the organization under which the users exist.}

\item{NextToken}{The token to use to retrieve the next page of results. The first call
does not contain any tokens.}

\item{MaxResults}{The maximum number of results to return in a single call.}
}
\value{
A list with the following syntax:\preformatted{list(
  Users = list(
    list(
      Id = "string",
      Email = "string",
      Name = "string",
      DisplayName = "string",
      State = "ENABLED"|"DISABLED"|"DELETED",
      UserRole = "USER"|"RESOURCE"|"SYSTEM_USER",
      EnabledDate = as.POSIXct(
        "2015-01-01"
      ),
      DisabledDate = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  NextToken = "string"
)
}
}
\description{
Returns summaries of the organization's users.
}
\section{Request syntax}{
\preformatted{svc$list_users(
  OrganizationId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
