% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_put_retention_policy}
\alias{workmail_put_retention_policy}
\title{Puts a retention policy to the specified organization}
\usage{
workmail_put_retention_policy(OrganizationId, Id, Name, Description,
  FolderConfigurations)
}
\arguments{
\item{OrganizationId}{[required] The organization ID.}

\item{Id}{The retention policy ID.}

\item{Name}{[required] The retention policy name.}

\item{Description}{The retention policy description.}

\item{FolderConfigurations}{[required] The retention policy folder configurations.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Puts a retention policy to the specified organization.
}
\section{Request syntax}{
\preformatted{svc$put_retention_policy(
  OrganizationId = "string",
  Id = "string",
  Name = "string",
  Description = "string",
  FolderConfigurations = list(
    list(
      Name = "INBOX"|"DELETED_ITEMS"|"SENT_ITEMS"|"DRAFTS"|"JUNK_EMAIL",
      Action = "NONE"|"DELETE"|"PERMANENTLY_DELETE",
      Period = 123
    )
  )
)
}
}

\keyword{internal}
