% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_reset_password}
\alias{workmail_reset_password}
\title{Allows the administrator to reset the password for a user}
\usage{
workmail_reset_password(OrganizationId, UserId, Password)
}
\arguments{
\item{OrganizationId}{[required] The identifier of the organization that contains the user for which the
password is reset.}

\item{UserId}{[required] The identifier of the user for whom the password is reset.}

\item{Password}{[required] The new password for the user.}
}
\value{
An empty list.
}
\description{
Allows the administrator to reset the password for a user.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$reset_password(
  OrganizationId = "string",
  UserId = "string",
  Password = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
