% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_untag_resource}
\alias{workmail_untag_resource}
\title{Untags the specified tags from the specified Amazon WorkMail
organization resource}
\usage{
workmail_untag_resource(ResourceARN, TagKeys)
}
\arguments{
\item{ResourceARN}{[required] The resource ARN.}

\item{TagKeys}{[required] The tag keys.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Untags the specified tags from the specified Amazon WorkMail
organization resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceARN = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
