% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_create_room}
\alias{alexaforbusiness_create_room}
\title{Creates a room with the specified details}
\usage{
alexaforbusiness_create_room(RoomName, Description, ProfileArn,
  ProviderCalendarId, ClientRequestToken, Tags)
}
\arguments{
\item{RoomName}{[required] The name for the room.}

\item{Description}{The description for the room.}

\item{ProfileArn}{The profile ARN for the room. This is required.}

\item{ProviderCalendarId}{The calendar ARN for the room.}

\item{ClientRequestToken}{A unique, user-specified identifier for this request that ensures
idempotency.}

\item{Tags}{The tags for the room.}
}
\description{
Creates a room with the specified details.
}
\section{Request syntax}{
\preformatted{svc$create_room(
  RoomName = "string",
  Description = "string",
  ProfileArn = "string",
  ProviderCalendarId = "string",
  ClientRequestToken = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
