% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_delete_business_report_schedule}
\alias{alexaforbusiness_delete_business_report_schedule}
\title{Deletes the recurring report delivery schedule with the specified
schedule ARN}
\usage{
alexaforbusiness_delete_business_report_schedule(ScheduleArn)
}
\arguments{
\item{ScheduleArn}{[required] The ARN of the business report schedule.}
}
\value{
An empty list.
}
\description{
Deletes the recurring report delivery schedule with the specified
schedule ARN.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_business_report_schedule(
  ScheduleArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
