% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_get_conference_preference}
\alias{alexaforbusiness_get_conference_preference}
\title{Retrieves the existing conference preferences}
\usage{
alexaforbusiness_get_conference_preference()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Preference = list(
    DefaultConferenceProviderArn = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the existing conference preferences.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_conference_preference()
}\if{html}{\out{</div>}}
}

\keyword{internal}
