% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_get_conference_preference}
\alias{alexaforbusiness_get_conference_preference}
\title{Retrieves the existing conference preferences}
\usage{
alexaforbusiness_get_conference_preference()
}
\value{
A list with the following syntax:\preformatted{list(
  Preference = list(
    DefaultConferenceProviderArn = "string"
  )
)
}
}
\description{
Retrieves the existing conference preferences.
}
\section{Request syntax}{
\preformatted{svc$get_conference_preference()
}
}

\keyword{internal}
