% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_get_conference_provider}
\alias{alexaforbusiness_get_conference_provider}
\title{Gets details about a specific conference provider}
\usage{
alexaforbusiness_get_conference_provider(ConferenceProviderArn)
}
\arguments{
\item{ConferenceProviderArn}{[required] The ARN of the newly created conference provider.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ConferenceProvider = list(
    Arn = "string",
    Name = "string",
    Type = "CHIME"|"BLUEJEANS"|"FUZE"|"GOOGLE_HANGOUTS"|"POLYCOM"|"RINGCENTRAL"|"SKYPE_FOR_BUSINESS"|"WEBEX"|"ZOOM"|"CUSTOM",
    IPDialIn = list(
      Endpoint = "string",
      CommsProtocol = "SIP"|"SIPS"|"H323"
    ),
    PSTNDialIn = list(
      CountryCode = "string",
      PhoneNumber = "string",
      OneClickIdDelay = "string",
      OneClickPinDelay = "string"
    ),
    MeetingSetting = list(
      RequirePin = "YES"|"NO"|"OPTIONAL"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets details about a specific conference provider.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_conference_provider(
  ConferenceProviderArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
