% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_get_room}
\alias{alexaforbusiness_get_room}
\title{Gets room details by room ARN}
\usage{
alexaforbusiness_get_room(RoomArn)
}
\arguments{
\item{RoomArn}{The ARN of the room for which to request details. Required.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Room = list(
    RoomArn = "string",
    RoomName = "string",
    Description = "string",
    ProviderCalendarId = "string",
    ProfileArn = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets room details by room ARN.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_room(
  RoomArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
