% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_get_skill_group}
\alias{alexaforbusiness_get_skill_group}
\title{Gets skill group details by skill group ARN}
\usage{
alexaforbusiness_get_skill_group(SkillGroupArn)
}
\arguments{
\item{SkillGroupArn}{The ARN of the skill group for which to get details. Required.}
}
\value{
A list with the following syntax:\preformatted{list(
  SkillGroup = list(
    SkillGroupArn = "string",
    SkillGroupName = "string",
    Description = "string"
  )
)
}
}
\description{
Gets skill group details by skill group ARN.
}
\section{Request syntax}{
\preformatted{svc$get_skill_group(
  SkillGroupArn = "string"
)
}
}

\keyword{internal}
