% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_list_conference_providers}
\alias{alexaforbusiness_list_conference_providers}
\title{Lists conference providers under a specific AWS account}
\usage{
alexaforbusiness_list_conference_providers(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{The tokens used for pagination.}

\item{MaxResults}{The maximum number of conference providers to be returned, per paginated
calls.}
}
\description{
Lists conference providers under a specific AWS account.
}
\section{Request syntax}{
\preformatted{svc$list_conference_providers(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
