% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_list_gateway_groups}
\alias{alexaforbusiness_list_gateway_groups}
\title{Retrieves a list of gateway group summaries}
\usage{
alexaforbusiness_list_gateway_groups(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{The token used to paginate though multiple pages of gateway group
summaries.}

\item{MaxResults}{The maximum number of gateway group summaries to return. The default is
50.}
}
\description{
Retrieves a list of gateway group summaries. Use GetGatewayGroup to
retrieve details of a specific gateway group.
}
\section{Request syntax}{
\preformatted{svc$list_gateway_groups(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
