% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_list_skills_store_skills_by_category}
\alias{alexaforbusiness_list_skills_store_skills_by_category}
\title{Lists all skills in the Alexa skill store by category}
\usage{
alexaforbusiness_list_skills_store_skills_by_category(CategoryId,
  NextToken, MaxResults)
}
\arguments{
\item{CategoryId}{[required] The category ID for which the skills are being retrieved from the skill
store.}

\item{NextToken}{The tokens used for pagination.}

\item{MaxResults}{The maximum number of skills returned per paginated calls.}
}
\value{
A list with the following syntax:\preformatted{list(
  SkillsStoreSkills = list(
    list(
      SkillId = "string",
      SkillName = "string",
      ShortDescription = "string",
      IconUrl = "string",
      SampleUtterances = list(
        "string"
      ),
      SkillDetails = list(
        ProductDescription = "string",
        InvocationPhrase = "string",
        ReleaseDate = "string",
        EndUserLicenseAgreement = "string",
        GenericKeywords = list(
          "string"
        ),
        BulletPoints = list(
          "string"
        ),
        NewInThisVersionBulletPoints = list(
          "string"
        ),
        SkillTypes = list(
          "string"
        ),
        Reviews = list(
          "string"
        ),
        DeveloperInfo = list(
          DeveloperName = "string",
          PrivacyPolicy = "string",
          Email = "string",
          Url = "string"
        )
      ),
      SupportsLinking = TRUE|FALSE
    )
  ),
  NextToken = "string"
)
}
}
\description{
Lists all skills in the Alexa skill store by category.
}
\section{Request syntax}{
\preformatted{svc$list_skills_store_skills_by_category(
  CategoryId = 123,
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
