% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_resolve_room}
\alias{alexaforbusiness_resolve_room}
\title{Determines the details for the room from which a skill request was
invoked}
\usage{
alexaforbusiness_resolve_room(UserId, SkillId)
}
\arguments{
\item{UserId}{[required] The ARN of the user. Required.}

\item{SkillId}{[required] The ARN of the skill that was requested. Required.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  RoomArn = "string",
  RoomName = "string",
  RoomSkillParameters = list(
    list(
      ParameterKey = "string",
      ParameterValue = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Determines the details for the room from which a skill request was
invoked. This operation is used by skill developers.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$resolve_room(
  UserId = "string",
  SkillId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
