% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_update_skill_group}
\alias{alexaforbusiness_update_skill_group}
\title{Updates skill group details by skill group ARN}
\usage{
alexaforbusiness_update_skill_group(SkillGroupArn, SkillGroupName,
  Description)
}
\arguments{
\item{SkillGroupArn}{The ARN of the skill group to update.}

\item{SkillGroupName}{The updated name for the skill group.}

\item{Description}{The updated description for the skill group.}
}
\value{
An empty list.
}
\description{
Updates skill group details by skill group ARN.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_skill_group(
  SkillGroupArn = "string",
  SkillGroupName = "string",
  Description = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
