% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_create_app_instance}
\alias{chime_create_app_instance}
\title{Creates an Amazon Chime Messaging SDK AppInstance under an AWS Account}
\usage{
chime_create_app_instance(Name, Metadata, ClientRequestToken)
}
\arguments{
\item{Name}{[required] The name of the app instance.}

\item{Metadata}{The metadata of the app instance. Limited to a 1KB string in UTF-8.}

\item{ClientRequestToken}{[required] The \code{ClientRequestToken} of the app instance.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AppInstanceArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Creates an Amazon Chime Messaging SDK \code{AppInstance} under an AWS
Account. Only Messaging SDK customers use this API.
\code{\link[=chime_create_app_instance]{create_app_instance}} supports
\code{idempotency} behavior as described in the AWS API Standard.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_app_instance(
  Name = "string",
  Metadata = "string",
  ClientRequestToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
