% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_create_app_instance_admin}
\alias{chime_create_app_instance_admin}
\title{Promotes an AppInstanceUser to an AppInstanceAdmin}
\usage{
chime_create_app_instance_admin(AppInstanceAdminArn, AppInstanceArn)
}
\arguments{
\item{AppInstanceAdminArn}{[required] The ARN of the administrator of the current app instance.}

\item{AppInstanceArn}{[required] The ARN of the app instance.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AppInstanceAdmin = list(
    Arn = "string",
    Name = "string"
  ),
  AppInstanceArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Promotes an \code{AppInstanceUser} to an \code{AppInstanceAdmin}. The promoted
user can perform the following actions.
\itemize{
\item \code{ChannelModerator} actions across all channels in the app instance.
\item \code{\link[=chime_delete_channel_message]{delete_channel_message}} actions.
}

Only an \code{AppInstanceUser} can be promoted to an \code{AppInstanceAdmin} role.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_app_instance_admin(
  AppInstanceAdminArn = "string",
  AppInstanceArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
