% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_create_attendee}
\alias{chime_create_attendee}
\title{Creates a new attendee for an active Amazon Chime SDK meeting}
\usage{
chime_create_attendee(MeetingId, ExternalUserId, Tags)
}
\arguments{
\item{MeetingId}{[required] The Amazon Chime SDK meeting ID.}

\item{ExternalUserId}{[required] The Amazon Chime SDK external user ID. An idempotency token. Links the
attendee to an identity managed by a builder application.}

\item{Tags}{The tag key-value pairs.}
}
\value{
A list with the following syntax:\preformatted{list(
  Attendee = list(
    ExternalUserId = "string",
    AttendeeId = "string",
    JoinToken = "string"
  )
)
}
}
\description{
Creates a new attendee for an active Amazon Chime SDK meeting. For more
information about the Amazon Chime SDK, see \href{https://docs.aws.amazon.com/chime/latest/dg/meetings-sdk.html}{Using the Amazon Chime SDK} in
the \emph{Amazon Chime Developer Guide}.
}
\section{Request syntax}{
\preformatted{svc$create_attendee(
  MeetingId = "string",
  ExternalUserId = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
