% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_create_channel_ban}
\alias{chime_create_channel_ban}
\title{Permanently bans a member from a channel}
\usage{
chime_create_channel_ban(ChannelArn, MemberArn, ChimeBearer)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the ban request.}

\item{MemberArn}{[required] The ARN of the member being banned.}

\item{ChimeBearer}{The \code{AppInstanceUserArn} of the user that makes the API call.}
}
\value{
A list with the following syntax:\preformatted{list(
  ChannelArn = "string",
  Member = list(
    Arn = "string",
    Name = "string"
  )
)
}
}
\description{
Permanently bans a member from a channel. Moderators can't add banned
members to a channel. To undo a ban, you first have to
\code{\link[=chime_delete_channel_ban]{delete_channel_ban}}, and then
\code{\link[=chime_create_channel_membership]{create_channel_membership}}. Bans are
cleaned up when you delete users or channels.

If you ban a user who is already part of a channel, that user is
automatically kicked from the channel.

The \code{x-amz-chime-bearer} request header is mandatory. Use the
\code{AppInstanceUserArn} of the user that makes the API call as the value in
the header.
}
\section{Request syntax}{
\preformatted{svc$create_channel_ban(
  ChannelArn = "string",
  MemberArn = "string",
  ChimeBearer = "string"
)
}
}

\keyword{internal}
