% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_create_phone_number_order}
\alias{chime_create_phone_number_order}
\title{Creates an order for phone numbers to be provisioned}
\usage{
chime_create_phone_number_order(ProductType, E164PhoneNumbers)
}
\arguments{
\item{ProductType}{[required] The phone number product type.}

\item{E164PhoneNumbers}{[required] List of phone numbers, in E.164 format.}
}
\description{
Creates an order for phone numbers to be provisioned. Choose from Amazon
Chime Business Calling and Amazon Chime Voice Connector product types.
}
\section{Request syntax}{
\preformatted{svc$create_phone_number_order(
  ProductType = "BusinessCalling"|"VoiceConnector",
  E164PhoneNumbers = list(
    "string"
  )
)
}
}

\keyword{internal}
