% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_create_proxy_session}
\alias{chime_create_proxy_session}
\title{Creates a proxy session on the specified Amazon Chime Voice Connector
for the specified participant phone numbers}
\usage{
chime_create_proxy_session(VoiceConnectorId, ParticipantPhoneNumbers,
  Name, ExpiryMinutes, Capabilities, NumberSelectionBehavior,
  GeoMatchLevel, GeoMatchParams)
}
\arguments{
\item{VoiceConnectorId}{[required] The Amazon Chime voice connector ID.}

\item{ParticipantPhoneNumbers}{[required] The participant phone numbers.}

\item{Name}{The name of the proxy session.}

\item{ExpiryMinutes}{The number of minutes allowed for the proxy session.}

\item{Capabilities}{[required] The proxy session capabilities.}

\item{NumberSelectionBehavior}{The preference for proxy phone number reuse, or stickiness, between the
same participants across sessions.}

\item{GeoMatchLevel}{The preference for matching the country or area code of the proxy phone
number with that of the first participant.}

\item{GeoMatchParams}{The country and area code for the proxy phone number.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ProxySession = list(
    VoiceConnectorId = "string",
    ProxySessionId = "string",
    Name = "string",
    Status = "Open"|"InProgress"|"Closed",
    ExpiryMinutes = 123,
    Capabilities = list(
      "Voice"|"SMS"
    ),
    CreatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    UpdatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    EndedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    Participants = list(
      list(
        PhoneNumber = "string",
        ProxyPhoneNumber = "string"
      )
    ),
    NumberSelectionBehavior = "PreferSticky"|"AvoidSticky",
    GeoMatchLevel = "Country"|"AreaCode",
    GeoMatchParams = list(
      Country = "string",
      AreaCode = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Creates a proxy session on the specified Amazon Chime Voice Connector
for the specified participant phone numbers.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_proxy_session(
  VoiceConnectorId = "string",
  ParticipantPhoneNumbers = list(
    "string"
  ),
  Name = "string",
  ExpiryMinutes = 123,
  Capabilities = list(
    "Voice"|"SMS"
  ),
  NumberSelectionBehavior = "PreferSticky"|"AvoidSticky",
  GeoMatchLevel = "Country"|"AreaCode",
  GeoMatchParams = list(
    Country = "string",
    AreaCode = "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
