% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_create_sip_media_application_call}
\alias{chime_create_sip_media_application_call}
\title{Creates an outbound call to a phone number from the phone number
specified in the request, and it invokes the endpoint of the specified
sipMediaApplicationId}
\usage{
chime_create_sip_media_application_call(FromPhoneNumber, ToPhoneNumber,
  SipMediaApplicationId)
}
\arguments{
\item{FromPhoneNumber}{[required] The phone number that a user calls from. This is a phone number in your
Amazon Chime phone number inventory.}

\item{ToPhoneNumber}{[required] The phone number that the service should call.}

\item{SipMediaApplicationId}{[required] The ID of the SIP media application.}
}
\value{
A list with the following syntax:\preformatted{list(
  SipMediaApplicationCall = list(
    TransactionId = "string"
  )
)
}
}
\description{
Creates an outbound call to a phone number from the phone number
specified in the request, and it invokes the endpoint of the specified
\code{sipMediaApplicationId}.
}
\section{Request syntax}{
\preformatted{svc$create_sip_media_application_call(
  FromPhoneNumber = "string",
  ToPhoneNumber = "string",
  SipMediaApplicationId = "string"
)
}
}

\keyword{internal}
