% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_create_voice_connector_group}
\alias{chime_create_voice_connector_group}
\title{Creates an Amazon Chime Voice Connector group under the administrator's
AWS account}
\usage{
chime_create_voice_connector_group(Name, VoiceConnectorItems)
}
\arguments{
\item{Name}{[required] The name of the Amazon Chime Voice Connector group.}

\item{VoiceConnectorItems}{The Amazon Chime Voice Connectors to route inbound calls to.}
}
\description{
Creates an Amazon Chime Voice Connector group under the administrator's
AWS account. You can associate Amazon Chime Voice Connectors with the
Amazon Chime Voice Connector group by including \code{VoiceConnectorItems} in
the request.
}
\details{
You can include Amazon Chime Voice Connectors from different AWS Regions
in your group. This creates a fault tolerant mechanism for fallback in
case of availability events.
}
\section{Request syntax}{
\preformatted{svc$create_voice_connector_group(
  Name = "string",
  VoiceConnectorItems = list(
    list(
      VoiceConnectorId = "string",
      Priority = 123
    )
  )
)
}
}

\keyword{internal}
