% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_delete_channel_ban}
\alias{chime_delete_channel_ban}
\title{Removes a user from a channel's ban list}
\usage{
chime_delete_channel_ban(ChannelArn, MemberArn)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel from which the app instance user was banned.}

\item{MemberArn}{[required] The ARN of the app instance user that you want to reinstate.}
}
\value{
An empty list.
}
\description{
Removes a user from a channel's ban list.
}
\section{Request syntax}{
\preformatted{svc$delete_channel_ban(
  ChannelArn = "string",
  MemberArn = "string"
)
}
}

\keyword{internal}
