% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_delete_channel_membership}
\alias{chime_delete_channel_membership}
\title{Removes a member from a channel}
\usage{
chime_delete_channel_membership(ChannelArn, MemberArn)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel from which you want to remove the user.}

\item{MemberArn}{[required] The ARN of the member that you're removing from the channel.}
}
\description{
Removes a member from a channel.
}
\section{Request syntax}{
\preformatted{svc$delete_channel_membership(
  ChannelArn = "string",
  MemberArn = "string"
)
}
}

\keyword{internal}
