% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_delete_channel_message}
\alias{chime_delete_channel_message}
\title{Deletes a channel message}
\usage{
chime_delete_channel_message(ChannelArn, MessageId)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel.}

\item{MessageId}{[required] The ID of the message being deleted.}
}
\value{
An empty list.
}
\description{
Deletes a channel message. Only admins can perform this action. Deletion
makes messages inaccessible immediately. A background process deletes
any revisions created by
\code{\link[=chime_update_channel_message]{update_channel_message}}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_channel_message(
  ChannelArn = "string",
  MessageId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
