% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_delete_channel_moderator}
\alias{chime_delete_channel_moderator}
\title{Deletes a channel moderator}
\usage{
chime_delete_channel_moderator(ChannelArn, ChannelModeratorArn)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel.}

\item{ChannelModeratorArn}{[required] The ARN of the moderator being deleted.}
}
\description{
Deletes a channel moderator.
}
\section{Request syntax}{
\preformatted{svc$delete_channel_moderator(
  ChannelArn = "string",
  ChannelModeratorArn = "string"
)
}
}

\keyword{internal}
