% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_describe_app_instance}
\alias{chime_describe_app_instance}
\title{Returns the full details of an AppInstance}
\usage{
chime_describe_app_instance(AppInstanceArn)
}
\arguments{
\item{AppInstanceArn}{[required] The ARN of the \code{AppInstance}.}
}
\value{
A list with the following syntax:\preformatted{list(
  AppInstance = list(
    AppInstanceArn = "string",
    Name = "string",
    Metadata = "string",
    CreatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    LastUpdatedTimestamp = as.POSIXct(
      "2015-01-01"
    )
  )
)
}
}
\description{
Returns the full details of an \code{AppInstance}.
}
\section{Request syntax}{
\preformatted{svc$describe_app_instance(
  AppInstanceArn = "string"
)
}
}

\keyword{internal}
