% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_describe_app_instance_user}
\alias{chime_describe_app_instance_user}
\title{Returns the full details of an AppInstanceUser}
\usage{
chime_describe_app_instance_user(AppInstanceUserArn)
}
\arguments{
\item{AppInstanceUserArn}{[required] The ARN of the app instance user.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AppInstanceUser = list(
    AppInstanceUserArn = "string",
    Name = "string",
    CreatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    Metadata = "string",
    LastUpdatedTimestamp = as.POSIXct(
      "2015-01-01"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns the full details of an \code{AppInstanceUser}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_app_instance_user(
  AppInstanceUserArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
