% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_describe_channel}
\alias{chime_describe_channel}
\title{Returns the full details of a channel in an Amazon Chime app instance}
\usage{
chime_describe_channel(ChannelArn)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Channel = list(
    Name = "string",
    ChannelArn = "string",
    Mode = "UNRESTRICTED"|"RESTRICTED",
    Privacy = "PUBLIC"|"PRIVATE",
    Metadata = "string",
    CreatedBy = list(
      Arn = "string",
      Name = "string"
    ),
    CreatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    LastMessageTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    LastUpdatedTimestamp = as.POSIXct(
      "2015-01-01"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns the full details of a channel in an Amazon Chime app instance.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_channel(
  ChannelArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
