% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_describe_channel_membership}
\alias{chime_describe_channel_membership}
\title{Returns the full details of a user's channel membership}
\usage{
chime_describe_channel_membership(ChannelArn, MemberArn)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel.}

\item{MemberArn}{[required] The ARN of the member.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ChannelMembership = list(
    InvitedBy = list(
      Arn = "string",
      Name = "string"
    ),
    Type = "DEFAULT"|"HIDDEN",
    Member = list(
      Arn = "string",
      Name = "string"
    ),
    ChannelArn = "string",
    CreatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    LastUpdatedTimestamp = as.POSIXct(
      "2015-01-01"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns the full details of a user's channel membership.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_channel_membership(
  ChannelArn = "string",
  MemberArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
