% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_describe_channel_membership}
\alias{chime_describe_channel_membership}
\title{Returns the full details of a user's channel membership}
\usage{
chime_describe_channel_membership(ChannelArn, MemberArn, ChimeBearer)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel.}

\item{MemberArn}{[required] The ARN of the member.}

\item{ChimeBearer}{The \code{AppInstanceUserArn} of the user that makes the API call.}
}
\value{
A list with the following syntax:\preformatted{list(
  ChannelMembership = list(
    InvitedBy = list(
      Arn = "string",
      Name = "string"
    ),
    Type = "DEFAULT"|"HIDDEN",
    Member = list(
      Arn = "string",
      Name = "string"
    ),
    ChannelArn = "string",
    CreatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    LastUpdatedTimestamp = as.POSIXct(
      "2015-01-01"
    )
  )
)
}
}
\description{
Returns the full details of a user's channel membership.

The \code{x-amz-chime-bearer} request header is mandatory. Use the
\code{AppInstanceUserArn} of the user that makes the API call as the value in
the header.
}
\section{Request syntax}{
\preformatted{svc$describe_channel_membership(
  ChannelArn = "string",
  MemberArn = "string",
  ChimeBearer = "string"
)
}
}

\keyword{internal}
