% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_describe_channel_membership_for_app_instance_user}
\alias{chime_describe_channel_membership_for_app_instance_user}
\title{Returns the details of a channel based on the membership of the
AppInstanceUser specified}
\usage{
chime_describe_channel_membership_for_app_instance_user(ChannelArn,
  AppInstanceUserArn)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel to which the user belongs.}

\item{AppInstanceUserArn}{[required] The ARN of the user in a channel.}
}
\value{
A list with the following syntax:\preformatted{list(
  ChannelMembership = list(
    ChannelSummary = list(
      Name = "string",
      ChannelArn = "string",
      Mode = "UNRESTRICTED"|"RESTRICTED",
      Privacy = "PUBLIC"|"PRIVATE",
      Metadata = "string",
      LastMessageTimestamp = as.POSIXct(
        "2015-01-01"
      )
    ),
    AppInstanceUserMembershipSummary = list(
      Type = "DEFAULT"|"HIDDEN",
      ReadMarkerTimestamp = as.POSIXct(
        "2015-01-01"
      )
    )
  )
)
}
}
\description{
Returns the details of a channel based on the membership of the
\code{AppInstanceUser} specified.
}
\section{Request syntax}{
\preformatted{svc$describe_channel_membership_for_app_instance_user(
  ChannelArn = "string",
  AppInstanceUserArn = "string"
)
}
}

\keyword{internal}
