% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_describe_channel_moderated_by_app_instance_user}
\alias{chime_describe_channel_moderated_by_app_instance_user}
\title{Returns the full details of a channel moderated by the specified
AppInstanceUser}
\usage{
chime_describe_channel_moderated_by_app_instance_user(ChannelArn,
  AppInstanceUserArn)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the moderated channel.}

\item{AppInstanceUserArn}{[required] The ARN of the app instance user in the moderated channel.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Channel = list(
    ChannelSummary = list(
      Name = "string",
      ChannelArn = "string",
      Mode = "UNRESTRICTED"|"RESTRICTED",
      Privacy = "PUBLIC"|"PRIVATE",
      Metadata = "string",
      LastMessageTimestamp = as.POSIXct(
        "2015-01-01"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns the full details of a channel moderated by the specified
\code{AppInstanceUser}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_channel_moderated_by_app_instance_user(
  ChannelArn = "string",
  AppInstanceUserArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
