% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_describe_channel_moderator}
\alias{chime_describe_channel_moderator}
\title{Returns the full details of a single ChannelModerator}
\usage{
chime_describe_channel_moderator(ChannelArn, ChannelModeratorArn)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel.}

\item{ChannelModeratorArn}{[required] The ARN of the channel moderator.}
}
\value{
A list with the following syntax:\preformatted{list(
  ChannelModerator = list(
    Moderator = list(
      Arn = "string",
      Name = "string"
    ),
    ChannelArn = "string",
    CreatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    CreatedBy = list(
      Arn = "string",
      Name = "string"
    )
  )
)
}
}
\description{
Returns the full details of a single ChannelModerator.
}
\section{Request syntax}{
\preformatted{svc$describe_channel_moderator(
  ChannelArn = "string",
  ChannelModeratorArn = "string"
)
}
}

\keyword{internal}
