% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_app_instance_retention_settings}
\alias{chime_get_app_instance_retention_settings}
\title{Gets the retention settings for an app instance}
\usage{
chime_get_app_instance_retention_settings(AppInstanceArn)
}
\arguments{
\item{AppInstanceArn}{[required] The ARN of the app instance.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AppInstanceRetentionSettings = list(
    ChannelRetentionSettings = list(
      RetentionDays = 123
    )
  ),
  InitiateDeletionTimestamp = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets the retention settings for an app instance.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_app_instance_retention_settings(
  AppInstanceArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
