% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_app_instance_retention_settings}
\alias{chime_get_app_instance_retention_settings}
\title{Gets the retention settings for an AppInstance}
\usage{
chime_get_app_instance_retention_settings(AppInstanceArn)
}
\arguments{
\item{AppInstanceArn}{[required] The ARN of the \code{AppInstance}.}
}
\value{
A list with the following syntax:\preformatted{list(
  AppInstanceRetentionSettings = list(
    ChannelRetentionSettings = list(
      RetentionDays = 123
    )
  ),
  InitiateDeletionTimestamp = as.POSIXct(
    "2015-01-01"
  )
)
}
}
\description{
Gets the retention settings for an \code{AppInstance}.
}
\section{Request syntax}{
\preformatted{svc$get_app_instance_retention_settings(
  AppInstanceArn = "string"
)
}
}

\keyword{internal}
