% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_app_instance_streaming_configurations}
\alias{chime_get_app_instance_streaming_configurations}
\title{Gets the streaming settings for an app instance}
\usage{
chime_get_app_instance_streaming_configurations(AppInstanceArn)
}
\arguments{
\item{AppInstanceArn}{[required] The ARN of the app instance.}
}
\value{
A list with the following syntax:\preformatted{list(
  AppInstanceStreamingConfigurations = list(
    list(
      AppInstanceDataType = "Channel"|"ChannelMessage",
      ResourceArn = "string"
    )
  )
)
}
}
\description{
Gets the streaming settings for an app instance.
}
\section{Request syntax}{
\preformatted{svc$get_app_instance_streaming_configurations(
  AppInstanceArn = "string"
)
}
}

\keyword{internal}
