% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_bot}
\alias{chime_get_bot}
\title{Retrieves details for the specified bot, such as bot email address, bot
type, status, and display name}
\usage{
chime_get_bot(AccountId, BotId)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{BotId}{[required] The bot ID.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Bot = list(
    BotId = "string",
    UserId = "string",
    DisplayName = "string",
    BotType = "ChatBot",
    Disabled = TRUE|FALSE,
    CreatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    UpdatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    BotEmail = "string",
    SecurityToken = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves details for the specified bot, such as bot email address, bot
type, status, and display name.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_bot(
  AccountId = "string",
  BotId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
