% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_channel_message}
\alias{chime_get_channel_message}
\title{Gets the full details of a channel message}
\usage{
chime_get_channel_message(ChannelArn, MessageId)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel.}

\item{MessageId}{[required] The ID of the message.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ChannelMessage = list(
    ChannelArn = "string",
    MessageId = "string",
    Content = "string",
    Metadata = "string",
    Type = "STANDARD"|"CONTROL",
    CreatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    LastEditedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    LastUpdatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    Sender = list(
      Arn = "string",
      Name = "string"
    ),
    Redacted = TRUE|FALSE,
    Persistence = "PERSISTENT"|"NON_PERSISTENT"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets the full details of a channel message.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_channel_message(
  ChannelArn = "string",
  MessageId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
