% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_global_settings}
\alias{chime_get_global_settings}
\title{Retrieves global settings for the administrator's AWS account, such as
Amazon Chime Business Calling and Amazon Chime Voice Connector settings}
\usage{
chime_get_global_settings()
}
\description{
Retrieves global settings for the administrator's AWS account, such as
Amazon Chime Business Calling and Amazon Chime Voice Connector settings.
}
\keyword{internal}
