% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_meeting}
\alias{chime_get_meeting}
\title{Gets the Amazon Chime SDK meeting details for the specified meeting ID}
\usage{
chime_get_meeting(MeetingId)
}
\arguments{
\item{MeetingId}{[required] The Amazon Chime SDK meeting ID.}
}
\description{
Gets the Amazon Chime SDK meeting details for the specified meeting ID.
For more information about the Amazon Chime SDK, see \href{https://docs.aws.amazon.com/chime/latest/dg/meetings-sdk.html}{Using the Amazon Chime SDK} in
the \emph{Amazon Chime Developer Guide}.
}
\section{Request syntax}{
\preformatted{svc$get_meeting(
  MeetingId = "string"
)
}
}

\keyword{internal}
