% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_room}
\alias{chime_get_room}
\title{Retrieves room details, such as the room name, for a room in an Amazon
Chime Enterprise account}
\usage{
chime_get_room(AccountId, RoomId)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{RoomId}{[required] The room ID.}
}
\description{
Retrieves room details, such as the room name, for a room in an Amazon
Chime Enterprise account.
}
\section{Request syntax}{
\preformatted{svc$get_room(
  AccountId = "string",
  RoomId = "string"
)
}
}

\keyword{internal}
