% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_sip_media_application}
\alias{chime_get_sip_media_application}
\title{Retrieves the information for a SIP media application, including name,
AWS Region, and endpoints}
\usage{
chime_get_sip_media_application(SipMediaApplicationId)
}
\arguments{
\item{SipMediaApplicationId}{[required] The SIP media application ID.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  SipMediaApplication = list(
    SipMediaApplicationId = "string",
    AwsRegion = "string",
    Name = "string",
    Endpoints = list(
      list(
        LambdaArn = "string"
      )
    ),
    CreatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    UpdatedTimestamp = as.POSIXct(
      "2015-01-01"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the information for a SIP media application, including name,
AWS Region, and endpoints.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_sip_media_application(
  SipMediaApplicationId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
