% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_sip_media_application_logging_configuration}
\alias{chime_get_sip_media_application_logging_configuration}
\title{Returns the logging configuration for the specified SIP media
application}
\usage{
chime_get_sip_media_application_logging_configuration(
  SipMediaApplicationId)
}
\arguments{
\item{SipMediaApplicationId}{[required] The ID of the SIP media application.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  SipMediaApplicationLoggingConfiguration = list(
    EnableSipMediaApplicationMessageLogs = TRUE|FALSE
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns the logging configuration for the specified SIP media
application.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_sip_media_application_logging_configuration(
  SipMediaApplicationId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
