% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_voice_connector}
\alias{chime_get_voice_connector}
\title{Retrieves details for the specified Amazon Chime Voice Connector, such
as timestamps, name, outbound host, and encryption requirements}
\usage{
chime_get_voice_connector(VoiceConnectorId)
}
\arguments{
\item{VoiceConnectorId}{[required] The Amazon Chime Voice Connector ID.}
}
\description{
Retrieves details for the specified Amazon Chime Voice Connector, such
as timestamps, name, outbound host, and encryption requirements.
}
\section{Request syntax}{
\preformatted{svc$get_voice_connector(
  VoiceConnectorId = "string"
)
}
}

\keyword{internal}
