% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_voice_connector_logging_configuration}
\alias{chime_get_voice_connector_logging_configuration}
\title{Retrieves the logging configuration details for the specified Amazon
Chime Voice Connector}
\usage{
chime_get_voice_connector_logging_configuration(VoiceConnectorId)
}
\arguments{
\item{VoiceConnectorId}{[required] The Amazon Chime Voice Connector ID.}
}
\description{
Retrieves the logging configuration details for the specified Amazon
Chime Voice Connector. Shows whether SIP message logs are enabled for
sending to Amazon CloudWatch Logs.
}
\section{Request syntax}{
\preformatted{svc$get_voice_connector_logging_configuration(
  VoiceConnectorId = "string"
)
}
}

\keyword{internal}
