% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_list_attendee_tags}
\alias{chime_list_attendee_tags}
\title{Lists the tags applied to an Amazon Chime SDK attendee resource}
\usage{
chime_list_attendee_tags(MeetingId, AttendeeId)
}
\arguments{
\item{MeetingId}{[required] The Amazon Chime SDK meeting ID.}

\item{AttendeeId}{[required] The Amazon Chime SDK attendee ID.}
}
\value{
A list with the following syntax:\preformatted{list(
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}
\description{
Lists the tags applied to an Amazon Chime SDK attendee resource.
}
\section{Request syntax}{
\preformatted{svc$list_attendee_tags(
  MeetingId = "string",
  AttendeeId = "string"
)
}
}

\keyword{internal}
