% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_list_channel_bans}
\alias{chime_list_channel_bans}
\title{Lists all the users banned from a particular channel}
\usage{
chime_list_channel_bans(ChannelArn, MaxResults, NextToken, ChimeBearer)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel.}

\item{MaxResults}{The maximum number of bans that you want returned.}

\item{NextToken}{The token passed by previous API calls until all requested bans are
returned.}

\item{ChimeBearer}{The \code{AppInstanceUserArn} of the user that makes the API call.}
}
\value{
A list with the following syntax:\preformatted{list(
  ChannelArn = "string",
  NextToken = "string",
  ChannelBans = list(
    list(
      Member = list(
        Arn = "string",
        Name = "string"
      )
    )
  )
)
}
}
\description{
Lists all the users banned from a particular channel.

The \code{x-amz-chime-bearer} request header is mandatory. Use the
\code{AppInstanceUserArn} of the user that makes the API call as the value in
the header.
}
\section{Request syntax}{
\preformatted{svc$list_channel_bans(
  ChannelArn = "string",
  MaxResults = 123,
  NextToken = "string",
  ChimeBearer = "string"
)
}
}

\keyword{internal}
