% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_list_channel_moderators}
\alias{chime_list_channel_moderators}
\title{Lists all the moderators for a channel}
\usage{
chime_list_channel_moderators(ChannelArn, MaxResults, NextToken,
  ChimeBearer)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel.}

\item{MaxResults}{The maximum number of moderators that you want returned.}

\item{NextToken}{The token passed by previous API calls until all requested moderators
are returned.}

\item{ChimeBearer}{The \code{AppInstanceUserArn} of the user that makes the API call.}
}
\value{
A list with the following syntax:\preformatted{list(
  ChannelArn = "string",
  NextToken = "string",
  ChannelModerators = list(
    list(
      Moderator = list(
        Arn = "string",
        Name = "string"
      )
    )
  )
)
}
}
\description{
Lists all the moderators for a channel.

The \code{x-amz-chime-bearer} request header is mandatory. Use the
\code{AppInstanceUserArn} of the user that makes the API call as the value in
the header.
}
\section{Request syntax}{
\preformatted{svc$list_channel_moderators(
  ChannelArn = "string",
  MaxResults = 123,
  NextToken = "string",
  ChimeBearer = "string"
)
}
}

\keyword{internal}
