% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_list_meetings}
\alias{chime_list_meetings}
\title{Lists up to 100 active Amazon Chime SDK meetings}
\usage{
chime_list_meetings(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{The token to use to retrieve the next page of results.}

\item{MaxResults}{The maximum number of results to return in a single call.}
}
\description{
Lists up to 100 active Amazon Chime SDK meetings. For more information
about the Amazon Chime SDK, see \href{https://docs.aws.amazon.com/chime/latest/dg/meetings-sdk.html}{Using the Amazon Chime SDK} in
the \emph{Amazon Chime Developer Guide}.
}
\section{Request syntax}{
\preformatted{svc$list_meetings(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
