% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_list_phone_number_orders}
\alias{chime_list_phone_number_orders}
\title{Lists the phone number orders for the administrator's Amazon Chime
account}
\usage{
chime_list_phone_number_orders(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{The token to use to retrieve the next page of results.}

\item{MaxResults}{The maximum number of results to return in a single call.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  PhoneNumberOrders = list(
    list(
      PhoneNumberOrderId = "string",
      ProductType = "BusinessCalling"|"VoiceConnector",
      Status = "Processing"|"Successful"|"Failed"|"Partial",
      OrderedPhoneNumbers = list(
        list(
          E164PhoneNumber = "string",
          Status = "Processing"|"Acquired"|"Failed"
        )
      ),
      CreatedTimestamp = as.POSIXct(
        "2015-01-01"
      ),
      UpdatedTimestamp = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists the phone number orders for the administrator's Amazon Chime
account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_phone_number_orders(
  NextToken = "string",
  MaxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
