% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_list_room_memberships}
\alias{chime_list_room_memberships}
\title{Lists the membership details for the specified room, such as member IDs,
member email addresses, and member names}
\usage{
chime_list_room_memberships(AccountId, RoomId, MaxResults, NextToken)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{RoomId}{[required] The room ID.}

\item{MaxResults}{The maximum number of results to return in a single call.}

\item{NextToken}{The token to use to retrieve the next page of results.}
}
\description{
Lists the membership details for the specified room, such as member IDs,
member email addresses, and member names.
}
\section{Request syntax}{
\preformatted{svc$list_room_memberships(
  AccountId = "string",
  RoomId = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
