% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_list_voice_connectors}
\alias{chime_list_voice_connectors}
\title{Lists the Amazon Chime Voice Connectors for the administrator's AWS
account}
\usage{
chime_list_voice_connectors(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{The token to use to retrieve the next page of results.}

\item{MaxResults}{The maximum number of results to return in a single call.}
}
\value{
A list with the following syntax:\preformatted{list(
  VoiceConnectors = list(
    list(
      VoiceConnectorId = "string",
      AwsRegion = "us-east-1"|"us-west-2",
      Name = "string",
      OutboundHostName = "string",
      RequireEncryption = TRUE|FALSE,
      CreatedTimestamp = as.POSIXct(
        "2015-01-01"
      ),
      UpdatedTimestamp = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  NextToken = "string"
)
}
}
\description{
Lists the Amazon Chime Voice Connectors for the administrator's AWS
account.
}
\section{Request syntax}{
\preformatted{svc$list_voice_connectors(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
