% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_put_app_instance_retention_settings}
\alias{chime_put_app_instance_retention_settings}
\title{Sets the amount of time in days that a given AppInstance retains data}
\usage{
chime_put_app_instance_retention_settings(AppInstanceArn,
  AppInstanceRetentionSettings)
}
\arguments{
\item{AppInstanceArn}{[required] The ARN of the \code{AppInstance}.}

\item{AppInstanceRetentionSettings}{[required] The time in days to retain data. Data type: number.}
}
\value{
A list with the following syntax:\preformatted{list(
  AppInstanceRetentionSettings = list(
    ChannelRetentionSettings = list(
      RetentionDays = 123
    )
  ),
  InitiateDeletionTimestamp = as.POSIXct(
    "2015-01-01"
  )
)
}
}
\description{
Sets the amount of time in days that a given \code{AppInstance} retains data.
}
\section{Request syntax}{
\preformatted{svc$put_app_instance_retention_settings(
  AppInstanceArn = "string",
  AppInstanceRetentionSettings = list(
    ChannelRetentionSettings = list(
      RetentionDays = 123
    )
  )
)
}
}

\keyword{internal}
