% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_put_app_instance_streaming_configurations}
\alias{chime_put_app_instance_streaming_configurations}
\title{The data streaming configurations of an app instance}
\usage{
chime_put_app_instance_streaming_configurations(AppInstanceArn,
  AppInstanceStreamingConfigurations)
}
\arguments{
\item{AppInstanceArn}{[required] The ARN of the app instance.}

\item{AppInstanceStreamingConfigurations}{[required] The streaming configurations set for an app instance.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AppInstanceStreamingConfigurations = list(
    list(
      AppInstanceDataType = "Channel"|"ChannelMessage",
      ResourceArn = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
The data streaming configurations of an app instance.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$put_app_instance_streaming_configurations(
  AppInstanceArn = "string",
  AppInstanceStreamingConfigurations = list(
    list(
      AppInstanceDataType = "Channel"|"ChannelMessage",
      ResourceArn = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
