% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_put_voice_connector_streaming_configuration}
\alias{chime_put_voice_connector_streaming_configuration}
\title{Adds a streaming configuration for the specified Amazon Chime Voice
Connector}
\usage{
chime_put_voice_connector_streaming_configuration(VoiceConnectorId,
  StreamingConfiguration)
}
\arguments{
\item{VoiceConnectorId}{[required] The Amazon Chime Voice Connector ID.}

\item{StreamingConfiguration}{[required] The streaming configuration details to add.}
}
\value{
A list with the following syntax:\preformatted{list(
  StreamingConfiguration = list(
    DataRetentionInHours = 123,
    Disabled = TRUE|FALSE,
    StreamingNotificationTargets = list(
      list(
        NotificationTarget = "EventBridge"|"SNS"|"SQS"
      )
    )
  )
)
}
}
\description{
Adds a streaming configuration for the specified Amazon Chime Voice
Connector. The streaming configuration specifies whether media streaming
is enabled for sending to Indonesians. It also sets the retention
period, in hours, for the Amazon Kinesis data.
}
\section{Request syntax}{
\preformatted{svc$put_voice_connector_streaming_configuration(
  VoiceConnectorId = "string",
  StreamingConfiguration = list(
    DataRetentionInHours = 123,
    Disabled = TRUE|FALSE,
    StreamingNotificationTargets = list(
      list(
        NotificationTarget = "EventBridge"|"SNS"|"SQS"
      )
    )
  )
)
}
}

\keyword{internal}
