% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_reset_personal_pin}
\alias{chime_reset_personal_pin}
\title{Resets the personal meeting PIN for the specified user on an Amazon
Chime account}
\usage{
chime_reset_personal_pin(AccountId, UserId)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{UserId}{[required] The user ID.}
}
\value{
A list with the following syntax:\preformatted{list(
  User = list(
    UserId = "string",
    AccountId = "string",
    PrimaryEmail = "string",
    PrimaryProvisionedNumber = "string",
    DisplayName = "string",
    LicenseType = "Basic"|"Plus"|"Pro"|"ProTrial",
    UserType = "PrivateUser"|"SharedDevice",
    UserRegistrationStatus = "Unregistered"|"Registered"|"Suspended",
    UserInvitationStatus = "Pending"|"Accepted"|"Failed",
    RegisteredOn = as.POSIXct(
      "2015-01-01"
    ),
    InvitedOn = as.POSIXct(
      "2015-01-01"
    ),
    AlexaForBusinessMetadata = list(
      IsAlexaForBusinessEnabled = TRUE|FALSE,
      AlexaForBusinessRoomArn = "string"
    ),
    PersonalPIN = "string"
  )
)
}
}
\description{
Resets the personal meeting PIN for the specified user on an Amazon
Chime account. Returns the User object with the updated personal meeting
PIN.
}
\section{Request syntax}{
\preformatted{svc$reset_personal_pin(
  AccountId = "string",
  UserId = "string"
)
}
}

\keyword{internal}
