% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_update_channel}
\alias{chime_update_channel}
\title{Update a channel's attributes}
\usage{
chime_update_channel(ChannelArn, Name, Mode, Metadata)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel.}

\item{Name}{[required] The name of the channel.}

\item{Mode}{[required] The mode of the update request.}

\item{Metadata}{The metadata of the channel.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ChannelArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Update a channel's attributes.

\strong{Restriction}: You can't change a channel's privacy.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_channel(
  ChannelArn = "string",
  Name = "string",
  Mode = "UNRESTRICTED"|"RESTRICTED",
  Metadata = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
