% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_update_channel_read_marker}
\alias{chime_update_channel_read_marker}
\title{Sets the timestamp to the point when a user last read messages in a
channel}
\usage{
chime_update_channel_read_marker(ChannelArn)
}
\arguments{
\item{ChannelArn}{[required] The ARN of the channel.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ChannelArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Sets the timestamp to the point when a user last read messages in a
channel.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_channel_read_marker(
  ChannelArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
