% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_update_phone_number}
\alias{chime_update_phone_number}
\title{Updates phone number details, such as product type, for the specified
phone number ID}
\usage{
chime_update_phone_number(PhoneNumberId, ProductType)
}
\arguments{
\item{PhoneNumberId}{[required] The phone number ID.}

\item{ProductType}{The product type.}
}
\description{
Updates phone number details, such as product type, for the specified
phone number ID. For toll-free numbers, you can use only the Amazon
Chime Voice Connector product type.
}
\section{Request syntax}{
\preformatted{svc$update_phone_number(
  PhoneNumberId = "string",
  ProductType = "BusinessCalling"|"VoiceConnector"
)
}
}

\keyword{internal}
