% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_update_proxy_session}
\alias{chime_update_proxy_session}
\title{Updates the specified proxy session details, such as voice or SMS
capabilities}
\usage{
chime_update_proxy_session(VoiceConnectorId, ProxySessionId,
  Capabilities, ExpiryMinutes)
}
\arguments{
\item{VoiceConnectorId}{[required] The Amazon Chime voice connector ID.}

\item{ProxySessionId}{[required] The proxy session ID.}

\item{Capabilities}{[required] The proxy session capabilities.}

\item{ExpiryMinutes}{The number of minutes allowed for the proxy session.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ProxySession = list(
    VoiceConnectorId = "string",
    ProxySessionId = "string",
    Name = "string",
    Status = "Open"|"InProgress"|"Closed",
    ExpiryMinutes = 123,
    Capabilities = list(
      "Voice"|"SMS"
    ),
    CreatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    UpdatedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    EndedTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    Participants = list(
      list(
        PhoneNumber = "string",
        ProxyPhoneNumber = "string"
      )
    ),
    NumberSelectionBehavior = "PreferSticky"|"AvoidSticky",
    GeoMatchLevel = "Country"|"AreaCode",
    GeoMatchParams = list(
      Country = "string",
      AreaCode = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Updates the specified proxy session details, such as voice or SMS
capabilities.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_proxy_session(
  VoiceConnectorId = "string",
  ProxySessionId = "string",
  Capabilities = list(
    "Voice"|"SMS"
  ),
  ExpiryMinutes = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
