% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_update_room}
\alias{chime_update_room}
\title{Updates room details, such as the room name, for a room in an Amazon
Chime Enterprise account}
\usage{
chime_update_room(AccountId, RoomId, Name)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{RoomId}{[required] The room ID.}

\item{Name}{The room name.}
}
\description{
Updates room details, such as the room name, for a room in an Amazon
Chime Enterprise account.
}
\section{Request syntax}{
\preformatted{svc$update_room(
  AccountId = "string",
  RoomId = "string",
  Name = "string"
)
}
}

\keyword{internal}
