% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_delete_user}
\alias{workmail_delete_user}
\title{Deletes a user from Amazon WorkMail and all subsequent systems}
\usage{
workmail_delete_user(OrganizationId, UserId)
}
\arguments{
\item{OrganizationId}{[required] The organization that contains the user to be deleted.}

\item{UserId}{[required] The identifier of the user to be deleted.}
}
\value{
An empty list.
}
\description{
Deletes a user from Amazon WorkMail and all subsequent systems. Before
you can delete a user, the user state must be \code{DISABLED}. Use the
\code{\link[=workmail_describe_user]{describe_user}} action to confirm the user
state.

Deleting a user is permanent and cannot be undone. WorkMail archives
user mailboxes for 30 days before they are permanently removed.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_user(
  OrganizationId = "string",
  UserId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
