% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_describe_group}
\alias{workmail_describe_group}
\title{Returns the data available for the group}
\usage{
workmail_describe_group(OrganizationId, GroupId)
}
\arguments{
\item{OrganizationId}{[required] The identifier for the organization under which the group exists.}

\item{GroupId}{[required] The identifier for the group to be described.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  GroupId = "string",
  Name = "string",
  Email = "string",
  State = "ENABLED"|"DISABLED"|"DELETED",
  EnabledDate = as.POSIXct(
    "2015-01-01"
  ),
  DisabledDate = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns the data available for the group.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_group(
  OrganizationId = "string",
  GroupId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
