% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_get_default_retention_policy}
\alias{workmail_get_default_retention_policy}
\title{Gets the default retention policy details for the specified organization}
\usage{
workmail_get_default_retention_policy(OrganizationId)
}
\arguments{
\item{OrganizationId}{[required] The organization ID.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Id = "string",
  Name = "string",
  Description = "string",
  FolderConfigurations = list(
    list(
      Name = "INBOX"|"DELETED_ITEMS"|"SENT_ITEMS"|"DRAFTS"|"JUNK_EMAIL",
      Action = "NONE"|"DELETE"|"PERMANENTLY_DELETE",
      Period = 123
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets the default retention policy details for the specified
organization.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_default_retention_policy(
  OrganizationId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
