% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_list_access_control_rules}
\alias{workmail_list_access_control_rules}
\title{Lists the access control rules for the specified organization}
\usage{
workmail_list_access_control_rules(OrganizationId)
}
\arguments{
\item{OrganizationId}{[required] The identifier for the organization.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Rules = list(
    list(
      Name = "string",
      Effect = "ALLOW"|"DENY",
      Description = "string",
      IpRanges = list(
        "string"
      ),
      NotIpRanges = list(
        "string"
      ),
      Actions = list(
        "string"
      ),
      NotActions = list(
        "string"
      ),
      UserIds = list(
        "string"
      ),
      NotUserIds = list(
        "string"
      ),
      DateCreated = as.POSIXct(
        "2015-01-01"
      ),
      DateModified = as.POSIXct(
        "2015-01-01"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Lists the access control rules for the specified organization.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_access_control_rules(
  OrganizationId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
