% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_list_groups}
\alias{workmail_list_groups}
\title{Returns summaries of the organization's groups}
\usage{
workmail_list_groups(OrganizationId, NextToken, MaxResults)
}
\arguments{
\item{OrganizationId}{[required] The identifier for the organization under which the groups exist.}

\item{NextToken}{The token to use to retrieve the next page of results. The first call
does not contain any tokens.}

\item{MaxResults}{The maximum number of results to return in a single call.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Groups = list(
    list(
      Id = "string",
      Email = "string",
      Name = "string",
      State = "ENABLED"|"DISABLED"|"DELETED",
      EnabledDate = as.POSIXct(
        "2015-01-01"
      ),
      DisabledDate = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns summaries of the organization's groups.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_groups(
  OrganizationId = "string",
  NextToken = "string",
  MaxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
