% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_list_tags_for_resource}
\alias{workmail_list_tags_for_resource}
\title{Lists the tags applied to an Amazon WorkMail organization resource}
\usage{
workmail_list_tags_for_resource(ResourceARN)
}
\arguments{
\item{ResourceARN}{[required] The resource ARN.}
}
\description{
Lists the tags applied to an Amazon WorkMail organization resource.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  ResourceARN = "string"
)
}
}

\keyword{internal}
